% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{addAdministration}
\alias{addAdministration}
\title{Add an administration to an arm.}
\usage{
addAdministration(object, value)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{value}{An object from the class \linkS4class{Administration}.}
}
\value{
The object \code{Arm} with its administration.
}
\description{
Add an administration to an arm.
}
