% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\name{addAdministrationConstraint}
\alias{addAdministrationConstraint}
\title{Add constraints on the administration for a \code{DesignConstraint} object.}
\usage{
addAdministrationConstraint(object, value)
}
\arguments{
\item{object}{A \code{DesignConstraint} object.}

\item{value}{An \code{AdministrationConstraint} object.}
}
\value{
The \code{DesignConstraint} object with its administration constraints.
}
\description{
Add constraints on the administration for a \code{DesignConstraint} object.
}
