% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{addArm}
\alias{addArm}
\title{Add an arm to a design.}
\usage{
addArm(object, arm)
}
\arguments{
\item{object}{A \code{Design} object.}

\item{arm}{An \code{Arm} object.}
}
\value{
The \code{Design} object with the new arm.
}
\description{
Add an arm to a design.
}
