% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{addArms}
\alias{addArms}
\title{Add arms to a design.}
\usage{
addArms(object, listOfArms)
}
\arguments{
\item{object}{A \code{Design} object.}

\item{listOfArms}{A list of \code{Arm} object.}
}
\value{
The \code{Design} object with the new arms.
}
\description{
Add arms to a design.
}
