% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{addDesign}
\alias{addDesign}
\title{Add a design to the \code{PFIMProject} object.}
\usage{
addDesign(object, design)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}

\item{design}{A \code{Design} object.}
}
\value{
The \code{PFIMProject} object with the \code{Design} object added.
}
\description{
Add a design to the \code{PFIMProject} object.
}
