% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\name{addDesignConstraints}
\alias{addDesignConstraints}
\title{Add design constraints on the sampling for a design.}
\usage{
addDesignConstraints(object, listOfConstraints)
}
\arguments{
\item{object}{A \code{DesignConstraint} object.}

\item{listOfConstraints}{A list of \code{Constraint} object.}
}
\value{
The \code{DesignConstraint} object constraints with the constraints from \code{listOfConstraints}.
}
\description{
Add design constraints on the sampling for a design.
}
