% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{addDesigns}
\alias{addDesigns}
\title{Add a list of designs to the \code{PFIMProject} object.}
\usage{
addDesigns(object, listOfDesigns)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}

\item{listOfDesigns}{A list of \code{Design} objects.}
}
\value{
The \code{PFIMProject} object with the \code{Design} objects added.
}
\description{
Add a list of designs to the \code{PFIMProject} object.
}
