% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Additive.R
\name{Additive}
\alias{Additive}
\title{Additive}
\usage{
Additive(beta = 0, combinedEffect = 0, value = 0)
}
\arguments{
\item{beta}{beta coefficient}

\item{combinedEffect}{combined effect}

\item{value}{value}
}
\description{
The class \code{Additive} is used to define Additive model for covariates.
}
\details{
Additive
}
