% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{addModel}
\alias{addModel}
\title{Add a \code{Model} object in the \code{LibraryOfModels}.}
\usage{
addModel(object, model)
}
\arguments{
\item{object}{A \code{LibraryOfModels} object.}

\item{model}{The model to add in the library (PK, PD or PKPD model).}
}
\value{
The \code{LibraryOfModels} object with the loaded library of models.
}
\description{
Add a \code{Model} object in the \code{LibraryOfModels}.
}
