% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{addModels}
\alias{addModels}
\alias{addModels,LibraryOfModels-method}
\title{Add a models to a library of models.}
\usage{
addModels(object, models)

\S4method{addModels}{LibraryOfModels}(object, models)
}
\arguments{
\item{object}{An object from the class \linkS4class{LibraryOfModels}.}

\item{models}{A list of object from the class \linkS4class{Model}.}
}
\value{
The library of models with the added models.
}
\description{
Add a models to a library of models.
}
