% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{addResponse}
\alias{addResponse}
\title{Add a response to a statistical model.}
\usage{
addResponse(object, value)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{value}{A character string giving the name of the  response to add.}
}
\value{
The \code{StatisticalModel} object with the added response.
}
\description{
Add a response to a statistical model.
}
