% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{addResponses}
\alias{addResponses}
\title{Add responses to a statistical model.}
\usage{
addResponses(object, listOfResponses)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{listOfResponses}{A list of character string giving the names of the  responses to add.}
}
\value{
The \code{StatisticalModel} object with the added responses.
}
\description{
Add responses to a statistical model.
}
