% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{addSampling}
\alias{addSampling}
\title{Add sampling time for an arm and for a response.}
\usage{
addSampling(object, value)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{value}{An object from the class \code{SamplingTimes}.}
}
\value{
The object \code{Arm} with its the new sampling times.
}
\description{
Add sampling time for an arm and for a response.
}
