% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\name{addSamplingConstraint}
\alias{addSamplingConstraint}
\title{Add a constraint on the sampling for a design.}
\usage{
addSamplingConstraint(object, value)
}
\arguments{
\item{object}{A \code{DesignConstraint} object.}

\item{value}{A \code{SamplingConstraint} object.}
}
\value{
The \code{DesignConstraint} object constraints with the constraints from \code{SamplingConstraint} object added.
}
\description{
Add a constraint on the sampling for a design.
}
