% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{addSamplingConstraints}
\alias{addSamplingConstraints}
\title{Add sampling constraints to an arm.}
\usage{
addSamplingConstraints(object, sampling_constraints)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{sampling_constraints}{A \code{SamplingConstraint} object giving the sampling constraints added to the \code{Arm} object.}
}
\value{
The object \code{Arm} with the new sampling constraints.
}
\description{
Add sampling constraints to an arm.
}
