% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{addSamplings}
\alias{addSamplings}
\title{Add sampling times for an arm and for a response.}
\usage{
addSamplings(object, listOfSamplings)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{listOfSamplings}{The objects from the class \code{SamplingTimes}.}
}
\value{
The object \code{Arm} with its new sampling times.
}
\description{
Add sampling times for an arm and for a response.
}
