% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{adjustGradient}
\alias{adjustGradient}
\title{Adjust the gradient for the log normal distribution.}
\arguments{
\item{distribution}{An object \code{Distribution} giving the distribution.}

\item{gradient}{The gradient of the model responses.}
}
\value{
The adjusted gradient of the model responses.
}
\description{
Adjust the gradient for the log normal distribution.
}
