% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormalDistribution.R
\docType{methods}
\name{AdjustLogNormalDistribution}
\alias{AdjustLogNormalDistribution}
\alias{AdjustLogNormalDistribution,LogNormalDistribution-method}
\title{Adjust the mean of a \code{\link{LogNormalDistribution}} object.}
\usage{
AdjustLogNormalDistribution(object, mu, df_total)

\S4method{AdjustLogNormalDistribution}{LogNormalDistribution}(object, mu, df_total)
}
\arguments{
\item{object}{A \code{\link{AdjustLogNormalDistribution}}.}

\item{mu}{A numeric giving the mean mu.}

\item{df_total}{numeric giving df_total}
}
\value{
A \code{StandardDistribution} object giving the adjusted Log Normal distribution.
}
\description{
Adjust the mean of a \code{\link{LogNormalDistribution}} object.
}
