% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalDistribution.R
\docType{methods}
\name{AdjustNormalDistribution}
\alias{AdjustNormalDistribution}
\alias{AdjustNormalDistribution,NormalDistribution-method}
\title{Adjust the Normal Distribution.}
\usage{
AdjustNormalDistribution(object, mu, df_total)

\S4method{AdjustNormalDistribution}{NormalDistribution}(object, mu, df_total)
}
\arguments{
\item{object}{A \code{\link{NormalDistribution}}.}

\item{mu}{A numeric giving the mean mu.}

\item{df_total}{numeric giving df_total}
}
\value{
A \code{StandardDistribution} object giving the adjusted Normal distribution.
}
\description{
Adjust the Normal Distribution.
}
