% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\docType{class}
\name{Administration-class}
\alias{Administration-class}
\alias{Administration}
\title{Class "Administration"}
\description{
The class \code{Administration} defines information concerning the parametrization and the type of administration:
single dose, multiple doses. Constraints can also be added on the allowed times, doses and infusion duration.
}
\section{Objects from the class}{

Objects form the class \code{Administration} can be created by calls of the form \code{Administration(...)} where
(...) are the parameters for the \code{Administration} objects.
}

\section{Slots for \code{Administration} objects}{

 \describe{
   \item{\code{outcome}:}{A character string giving the name for the reponse of the model.}
   \item{\code{time_dose}:}{A numeric vector giving the times when doses are given. By default set to 0.}
   \item{\code{amount_dose}:}{A numeric vector giving the amount of doses.}
   \item{\code{tau}:}{A numeric giving the frequency.}
   \item{\code{Tinf}:}{A numeric vector giving the infusion duration Tinf (Tinf can be null).}
   \item{\code{allowed_time}:}{Constraint object containing the constraints on allowed times.}
   \item{\code{allowed_dose}:}{Constraint object containing the constraints on allowed dose.}
   \item{\code{allowed_tinf}:}{Constraint object containing the constraints on Tinf.}
 }
}

