% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraint.R
\docType{class}
\name{AdministrationConstraint-class}
\alias{AdministrationConstraint-class}
\alias{AdministrationConstraint}
\title{Class "AdministrationConstraint"}
\description{
The class \code{AdministrationConstraint} represents the constraint of an input to the system.
The class stores information concerning the constraints for the dosage regimen :
response of the model, type of administration, amount of dose.
}
\section{Objects from the class}{

Objects form the class \code{AdministrationConstraint} can be created by calls of the form #'\code{AdministrationConstraint(...)} where (...) are the parameters for the \code{AdministrationConstraint} objects.
}

\section{Slots for the \code{AdministrationConstraint} objects}{

\describe{
\item{\code{response}:}{A character giving the response of the model.}
\item{\code{Optimisability}:}{A boolean giving if a dose is optimisable or not. If not the dose is fixed.}
\item{\code{fixedDoses}:}{A vector giving the fixed doses.}
\item{\code{AllowedDoses}:}{A vector giving the allowed amount of doses.}
}
}

