% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraints.R
\docType{class}
\name{AdministrationConstraints-class}
\alias{AdministrationConstraints-class}
\alias{AdministrationConstraints}
\title{Class "AdministrationConstraints"}
\description{
The class \code{AdministrationConstraints} represents the constraint of an input to the system.
The class stores information concerning the constraints for the dosage regimen:
response of the model, amount of dose.
}
\section{Objects from the class}{

Objects form the class \code{AdministrationConstraints} can be created by calls of the form \code{AdministrationConstraints(...)} where
(...) are the parameters for the \code{AdministrationConstraints} objects.
}

\section{Slots for \code{AdministrationConstraints} objects}{

 \describe{
   \item{\code{outcome}:}{A character string giving the name for the response of the model.}
   \item{\code{doses}:}{A numeric vector giving the amount of doses.}
 }
}

