% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\name{allowedContinuousSamplingTimes}
\alias{allowedContinuousSamplingTimes}
\title{Set the allowed continuous sampling times.}
\usage{
allowedContinuousSamplingTimes(object, allowedTimes)
}
\arguments{
\item{object}{A \code{SamplingConstraint} object.}

\item{allowedTimes}{A list giving the vectors for the allowed continuous Sampling Times.}
}
\value{
The object \code{SamplingConstraint} object with the allowed continuous Sampling Times.
}
\description{
Set the allowed continuous sampling times.
}
