% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\docType{methods}
\name{allowedDiscretSamplingTimes}
\alias{allowedDiscretSamplingTimes}
\title{Set the allowed discret sampling times.}
\usage{
allowedDiscretSamplingTimes(object, allowedTimes)
}
\arguments{
\item{object}{A \code{SamplingConstraint} object.}

\item{allowedTimes}{A list giving the vectors for the allowed Continous Sampling Times.}
}
\value{
The object \code{SamplingConstraint} object with the allowed Continous Sampling Times.
}
\description{
Set the allowed discret sampling times.
}
