% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraint.R
\name{AllowedDoses}
\alias{AllowedDoses}
\title{Define the vector of allowed amount of dose.}
\usage{
AllowedDoses(object, value)
}
\arguments{
\item{object}{An object \code{AdministrationConstraint} from the class \linkS4class{AdministrationConstraint}.}

\item{value}{A numeric vector giving the allowed amount of doses.}
}
\value{
The \code{AdministrationConstraint} object with its new allowed amount of doses.
}
\description{
Define the vector of allowed amount of dose.
}
