% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\docType{class}
\name{Arm-class}
\alias{Arm-class}
\alias{Arm}
\alias{[,Arm-method}
\alias{[<-,Arm-method}
\title{Class "Arm"}
\description{
The class "\code{Arm} combines the treatment (the class \linkS4class{Administration}) and the sampling schedule
(the class \linkS4class{SamplingTimes}).
}
\section{Objects from the class}{

Objects form the class \code{Arm} can be created by calls of the form \code{Arm(...)} where (...) are the parameters for the \code{Arm} objects.
}

\section{Slots for the \code{Arm} objects}{

\describe{
\item{\code{arm_size}:}{An integer the number of subjects in the arm. By default set to 1.}
\item{\code{administrations}:}{A list of \linkS4class{Administration} objects.}
\item{\code{cond_init}:}{A list of the initial conditions.}
\item{\code{samplings}:}{A list of \linkS4class{SamplingTimes} objects.}
\item{\code{constraints}:}{A list of \linkS4class{SamplingConstraint} objects.}
\item{\code{sampling_constraints}:}{A list of objects from \linkS4class{SamplingConstraint} class.}
}
}

