% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{armAdministration}
\alias{armAdministration}
\title{Get administration parameters of an arm}
\usage{
armAdministration(arm, ...)
}
\arguments{
\item{arm}{An object of class \code{Arm}.}

\item{...}{Additional arguments.}
}
\value{
A list of named lists, one per administration.
}
\description{
Extracts dosing information (outcome, dose, time dose, tau, Tinf) for
each administration in the arm.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
