% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R
\docType{class}
\name{BayesianFim-class}
\alias{BayesianFim-class}
\alias{BayesianFim}
\title{Class "BayesianFim" representing the population Fisher information matrix.}
\description{
A class storing information regarding the population Fisher computation matrix.
}
\section{Objects from the class}{

\code{BayesianFim} objects are typically created by calls to \code{{BayesianFim}}:
\describe{
\item{Bayesianfim:}{Create a new \code{BayesianFim}}
}
}

