% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R
\name{BayesianFim}
\alias{BayesianFim}
\title{BayesianFim}
\usage{
BayesianFim(
  fisherMatrix = numeric(0),
  fixedEffects = numeric(0),
  varianceEffects = numeric(0),
  SEAndRSE = list(),
  condNumberFixedEffects = 0,
  condNumberVarianceEffects = 0,
  shrinkage = numeric(0)
)
}
\arguments{
\item{fisherMatrix}{A matrix giving the numerical values of the FIM.}

\item{fixedEffects}{A matrix giving the fixed-effects block of the FIM.}

\item{varianceEffects}{A matrix giving the variance-effects block of the FIM.}

\item{SEAndRSE}{A data frame of SE and RSE values.}

\item{condNumberFixedEffects}{Condition number of the fixed-effects block.}

\item{condNumberVarianceEffects}{Condition number of the variance-effects block.}

\item{shrinkage}{A vector giving the shrinkage values.}
}
\description{
The class \code{BayesianFim} represents and stores information for the Bayesian FIM.
}
