% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{CalculatedResidualVariance}
\alias{CalculatedResidualVariance}
\title{Compute the residual variance thanks to the function g of the model error.}
\usage{
CalculatedResidualVariance(objectStatisticalModel, objectModelError, x_i)
}
\arguments{
\item{objectStatisticalModel}{A \code{StatisticalModel} object.}

\item{objectModelError}{A \code{ModelError} object.}

\item{x_i}{variable x_i of the model error.
#' @return CalculatedResidualVariance}
}
\description{
Compute the residual variance thanks to the function g of the model error.
}
