% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalCovariate.R
\name{CategoricalCovariate}
\alias{CategoricalCovariate}
\title{CategoricalCovariate}
\usage{
CategoricalCovariate(name, categories, categoriesProportions, effects = list())
}
\arguments{
\item{name}{name}

\item{categories}{categories}

\item{effects}{effects}

\item{proportions}{proportions}
}
\description{
The class \code{CategoricalCovariate} implements the Categorical distribution.
}
