% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalCovariateWithIOV.R
\name{CategoricalCovariateWithIOV}
\alias{CategoricalCovariateWithIOV}
\title{CategoricalCovariate}
\usage{
CategoricalCovariateWithIOV(
  name,
  categories,
  sequences,
  sequencesProportions,
  effects = list()
)
}
\arguments{
\item{name}{name}

\item{categories}{categories}

\item{sequences}{sequences}

\item{effects}{effects}

\item{seq_proportions}{sequence proportions}
}
\description{
The class \code{CategoricalCovariateWithIOV} implements the CategoricalCovariateWithIOV

The class \code{CategoricalCovariateWithIOV} implements the CategoricalCovariateWithIOV
}
