% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKModel.R
\name{changeVariablePKModel}
\alias{changeVariablePKModel}
\title{Change variable in a PK Model.}
\usage{
changeVariablePKModel(object)
}
\arguments{
\item{object}{\code{PKModel} object.}
}
\value{
A expression giving the equations of the PK model with variable changed.
}
\description{
Change variable in a PK Model.
}
