% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{checkParameterInEquations}
\alias{checkParameterInEquations}
\title{Check the parameters in the model equations.}
\usage{
checkParameterInEquations(object)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}
}
\value{
The \code{ModelParameter} objects of the model parameters in the model equation.
}
\description{
Check the parameters in the model equations.
}
