% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{checkSamplingTimeConstraintsForContinuousOptimization}
\alias{checkSamplingTimeConstraintsForContinuousOptimization}
\title{Check for the samplingTime constraints for continuous optimization}
\usage{
checkSamplingTimeConstraintsForContinuousOptimization(
  object,
  arm,
  newSamplings,
  outcome
)
}
\arguments{
\item{object}{An object from the class \linkS4class{SamplingTimeConstraints}.}

\item{arm}{An object from the class \linkS4class{Arm}.}

\item{newSamplings}{A vector giving the new sampling.}

\item{outcome}{The outcomes for the model.}
}
\value{
A list of Boolean giving true if
the minimal sampling times is in the vector of sampling times & the number of sampling for each windows is respected
false otherwise.
}
\description{
Check for the samplingTime constraints for continuous optimization
}
