% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{checkValiditySamplingConstraint}
\alias{checkValiditySamplingConstraint}
\title{Validate optimization constraints}
\arguments{
\item{design}{An object \code{Design}.}
}
\value{
Returns nothing if valid, or stops with an error message if
the sampling window/delta constraints are mathematically impossible.
}
\description{
Validate optimization constraints
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
