% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{Combinaison}
\alias{Combinaison}
\title{Create all the possible combinaison for each Design and each Arms.}
\usage{
Combinaison(object, times, nTimesForEachVector, fixedTimes, n, combin)
}
\arguments{
\item{object}{A \code{Optimization} object.}

\item{times}{A \code{SAmplingTimes} object.}

\item{nTimesForEachVector}{the number of sampling times for each vector of sampling times.}

\item{fixedTimes}{the fixed sampling times.}

\item{n}{parameter n}

\item{combin}{parameter combin}
}
\value{
All the possible combination for each Design and each Arms.
}
\description{
Create all the possible combinaison for each Design and each Arms.
}
