% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined1.R
\docType{class}
\name{Combined1-class}
\alias{Combined1-class}
\alias{Combined1}
\title{Class "Combined1"}
\description{
The class  \code{Combined1} defines the the residual error variance according
to the formula g(sigma_inter, sigma_slope, c_error, f(x, theta)) = sigma_inter + sigma_slope*f(x,theta)).
}
\section{Objects from the class}{

Combined1 objects are typically created by calls to \code{Combined1} and contain the following slots that are herited from
the class \linkS4class{Combined1c}:

\describe{
\item{\code{.Object}:}{An object of the Class \linkS4class{Combined1}}
\item{\code{sigma_inter}:}{A numeric value giving the sigma inter of the error model.}
\item{\code{sigma_slope}:}{A numeric value giving the sigma slope of the error model.}
}
}

