% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined1c.R
\docType{class}
\name{Combined1c-class}
\alias{Combined1c-class}
\alias{Combined1c}
\title{Class "Combined1c"}
\description{
The class \code{Combined1c} defines the the residual error variance according
to the formula g(sigma_inter, sigma_slope, c_error, f(x, theta)) = sigma_inter + sigma_slope*f(x,theta)^c_error).
}
\section{Objects from the class}{
 Combined1c objects
are typically created by calls to \code{{Combined1c}} and contain the following slots that are heritated from
the class \linkS4class{ModelError}:

\describe{
\item{\code{.Object}:}{An object of the class \code{ModelError}}
\item{\code{sigma_inter}:}{A numeric value giving the sigma inter of the error model.}
\item{\code{sigma_slope}:}{A numeric value giving the sigma slope of the error model.}
\item{\code{c_error}:}{A numeric value giving the exponant c of the error model.}
}
}

