% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined2.R
\docType{class}
\name{Combined2-class}
\alias{Combined2-class}
\alias{Combined2}
\title{Class "Combined2"}
\description{
The class \code{{Combined2}} defines the the residual error variance according
to the formula g(sigma_inter, sigma_slope, c_error, f(x, theta)) = sigma_inter^2 + sigma_slope^2 * f_x_i_theta.
}
\section{Objects from the class}{
\code{{Combined2}} objects are typically created by calls to \code{{Combined2}}
and contain the following slots that are heritated from the class \linkS4class{Combined2c}:
\describe{
\item{\code{.Object}:}{An object of the class \linkS4class{ModelError}}
\item{\code{sigma_inter}:}{A numeric value giving the sigma inter of the error model.}
\item{\code{sigma_slope}:}{A numeric value giving the sigma slope of the error model.}
}
}

