% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined2c.R
\docType{class}
\name{Combined2c-class}
\alias{Combined2c-class}
\alias{Combined2c}
\title{Class "Combined2c"}
\description{
The class \code{Combined2c} defines the the residual error variance according
to the formula g(sigma_inter, sigma_slope, c_error, f(x, theta)) = sigma_inter^2 + sigma_slope^2 * f(x,theta)^(2 * c_error)
}
\section{Objects from the class}{
 \code{Combined2c}  objects
are typically created by calls to \code{Combined2c} and contain the following slots that are heritated
from the class \linkS4class{ModelError}:
\describe{
\item{\code{.Object}:}{An object of the class \linkS4class{ModelError}}
\item{\code{sigma_inter}:}{A numeric value giving the sigma inter of the error model.}
\item{\code{sigma_slope}:}{A numeric value giving the sigma slope of the error model.}
\item{\code{c_error}:}{A numeric value giving the exponant c of the error model.}
}
}

