% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{Compute NNI for Equivalence Test}
\alias{Compute NNI for Equivalence Test}
\title{computeNNIEquivalence}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{parameter_name}{String. The name of the parameter to test.}

\item{target_power}{Numeric. Target power (default: 0.9).}

\item{interval_eq}{Numeric vector. Equivalence bounds.}

\item{alpha}{Numeric. Type I error rate.}
}
\description{
Estimates the number of individuals needed to achieve a target power for equivalence.
}
