% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{Compute Power for Comparison Test}
\alias{Compute Power for Comparison Test}
\title{computePowerComparison}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{parameter_name}{String. The name of the parameter to test (e.g., "beta_Cl_treat_B").}

\item{alpha}{Numeric. Type I error rate (default: 0.05).}
}
\description{
Calculates the expected power for a bilateral comparison test (H0: beta = 0) based on the FIM.
}
