% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{Compute Power for Equivalence Test}
\alias{Compute Power for Equivalence Test}
\title{computePowerEquivalence}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{parameter_name}{String. The name of the parameter to test.}

\item{interval_eq}{Numeric vector of length 2. Equivalence bounds (default: log(0.8), log(1.25)).}

\item{alpha}{Numeric. Type I error rate (default: 0.05).}
}
\description{
Calculates the power for an equivalence test (TOST) given a specific interval.
}
