% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationFim.R
\name{computeVMat}
\alias{computeVMat}
\title{computeVMat: compute (1/2) * Tr(V^-1 A V^-1 B)}
\usage{
computeVMat(varParam1, varParam2, invCholV)
}
\arguments{
\item{varParam1}{Matrix A (derivative dV/dlambda_i)}

\item{varParam2}{Matrix B (derivative dV/dlambda_j)}

\item{invCholV}{\eqn{V^{-1}} computed via Cholesky}
}
\value{
Scalar (1/2) * Tr(V^-1 A V^-1 B)
}
\description{
computeVMat: compute (1/2) * Tr(V^-1 A V^-1 B)
}
