% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constant.R
\docType{class}
\name{Constant-class}
\alias{Constant-class}
\alias{Constant}
\title{Class "Constant"}
\description{
The class \code{Constant} defines the the residual error variance according
to the formula g(sigma_inter, sigma_slope, c_error, f(x, theta)) = sigma_inter.
}
\section{Objects from the class}{
 \code{Constant} objects
are typically created by calls to \code{Constant} and contain the following slots
that are heritated from the class Combined1:
\describe{
\item{\code{.Object}:}{An object of the class \linkS4class{ModelError}}
\item{\code{sigma_inter}:}{A numeric value giving the sigma inter of the error model.}
}
}

