% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraint.R
\docType{class}
\name{Constraint-class}
\alias{Constraint-class}
\alias{Constraint}
\title{Class "Constraint"}
\description{
The class \code{Constraint} stores the constraints for a variable. Constraints are given either as:
a continuous range, a discrete set of values, or a Design constraint.
}
\section{Objects from the class \code{Constraint}}{

Objects form the class \code{Constraint} can be created by calls of the objects from the following classes:
\describe{
\item{-}{\linkS4class{AdministrationConstraint}}
\item{-}{\linkS4class{ContinuousConstraint}}
\item{-}{\linkS4class{DesignConstraint}}
\item{-}{\linkS4class{DiscreteConstraint}}
}
}

