% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuousConstraint.R
\docType{class}
\name{ContinuousConstraint-class}
\alias{ContinuousConstraint-class}
\alias{ContinuousConstraint}
\alias{[,ContinuousConstraint-method}
\alias{[<-,ContinuousConstraint-method}
\title{Class "ContinuousConstraint" representing the constraints for a variable}
\description{
The class  \code{ContinuousConstraint} stores constraints for a variable.
Constraints are given either as a continuous range (any value between a min and a max boundary is admissible)
or a discrete set of values (any value belonging to the set is admissible).
}
\section{Objects from the class}{
 Constraint objects are typically created by calls to \code{constraint}
and contain the following slots:

\describe{
\item{\code{type}:}{A character string, one of 'continuous' or 'discrete'.}
\item{\code{range}:}{A numeric vector with two values giving the min/max of the continuous range.}
\item{\code{minimalDelay}:}{A numeric value giving the minimal timestep between two sampling times.}
}
}

