% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEquations.R
\name{convertAnalyticToODE}
\alias{convertAnalyticToODE}
\title{Convert an equation of a PD model of a \code{ModelEquations} object from analytic to ODE.}
\usage{
convertAnalyticToODE(object)
}
\arguments{
\item{object}{\code{ModelEquations} object.}
}
\value{
A list of expression \code{output} giving the equations of the analytic PD model in ODE form.
}
\description{
Convert an equation of a PD model of a \code{ModelEquations} object from analytic to ODE.
}
