% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate.R
\name{Covariate}
\alias{Covariate}
\title{Covariate}
\usage{
Covariate(
  name,
  effects = list(),
  categories = NULL,
  categoriesProportions = NULL,
  sequences = NULL,
  sequencesProportions = NULL
)
}
\arguments{
\item{name}{Character vector specifying the covariate name.}

\item{effects}{List specifying the effects.}
}
\description{
The class \code{Covariate} represents the Covariate.
}
