% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateModelEquation.R
\name{CovariateModelEquation}
\alias{CovariateModelEquation}
\title{CovariateModelEquation}
\usage{
CovariateModelEquation(beta = 0, combinedEffect = 0, value = 0)
}
\arguments{
\item{beta}{beta coefficient}

\item{combinedEffect}{combined effect}

\item{value}{value}
}
\description{
The class \code{CovariateModelEquation} is used to defined a CovariateModelEquation.
}
\details{
CovariateModelEquation
}
