% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate.R
\name{createEffectVector}
\alias{createEffectVector}
\title{createEffectVector: create the vector for the effect}
\arguments{
\item{covariate}{A object \code{Covariate} giving the covariate}

\item{category}{A string giving the category of the covariate}

\item{effectVector}{A vector of double named with the parameters names}
}
\value{
The vector effectVector
}
\description{
createEffectVector: create the vector for the effect
}
