% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{dataForArmEvaluation}
\alias{dataForArmEvaluation}
\alias{dataForArmEvaluation,Design-method}
\title{dataForArmEvaluation}
\usage{
dataForArmEvaluation(object, arm, model)

\S4method{dataForArmEvaluation}{Design}(object, arm, model)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}

\item{arm}{...}

\item{model}{An object \code{Model} from the class \linkS4class{Model}.}
}
\value{
A list containing data for arm evaluation in the design.
}
\description{
dataForArmEvaluation
}
