% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{Dcriterion}
\alias{Dcriterion}
\title{Dcriterion}
\arguments{
\item{fim}{An object of class \code{Fim}.}
}
\value{
A double giving the D-criterion of the Fim.
}
\description{
Computes the D-criterion of the Fisher Information Matrix. The D-criterion is
calculated as the determinant of the FIM raised to the power of 1 over the
number of parameters.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
