% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{defineCorrelation}
\alias{defineCorrelation}
\title{Set the correlation.}
\usage{
defineCorrelation(object, correlationlist)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{correlationlist}{...}
}
\value{
Return correlationlist
}
\description{
Set the correlation.
}
