% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{defineModel}
\alias{defineModel}
\title{Define a model.}
\usage{
defineModel(object, designs)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{designs}{A list of objects from the class \linkS4class{Design}.}
}
\value{
A model defined either from the library of models or user defined.
}
\description{
Define a model.
}
