% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{defineModelEquations}
\alias{defineModelEquations}
\title{Define model equations}
\usage{
defineModelEquations(object, equations)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{equations}{An expression giving the equations of the model.}
}
\value{
The \code{StatisticalModel} object with the equations.
}
\description{
Define model equations
}
