% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R, R/ModelAnalytic.R,
%   R/ModelAnalyticSteadyState.R, R/ModelAnalyticInfusion.R,
%   R/ModelAnalyticInfusionSteadyState.R, R/ModelODEBolus.R,
%   R/ModelODEDoseInEquations.R, R/ModelODEDoseNotInEquations.R,
%   R/ModelODEInfusionDoseInEquations.R
\name{defineModelEquationsFromStringToFunction}
\alias{defineModelEquationsFromStringToFunction}
\alias{defineModelEquationsFromStringToFunction,ModelAnalytic-method}
\alias{defineModelEquationsFromStringToFunction,ModelAnalyticSteadyState-method}
\alias{defineModelEquationsFromStringToFunction,ModelAnalyticInfusion-method}
\alias{defineModelEquationsFromStringToFunction,ModelAnalyticInfusionSteadyState-method}
\alias{defineModelEquationsFromStringToFunction,ModelODEBolus-method}
\alias{defineModelEquationsFromStringToFunction,ModelODEDoseInEquations-method}
\alias{defineModelEquationsFromStringToFunction,ModelODEDoseNotInEquations-method}
\alias{defineModelEquationsFromStringToFunction,ModelODEInfusionDoseInEquations-method}
\title{defineModelEquationsFromStringToFunction}
\usage{
defineModelEquationsFromStringToFunction(
  object,
  parametersNames,
  outcomesWithAdministration,
  outcomesWithNoAdministration
)

\S4method{defineModelEquationsFromStringToFunction}{ModelAnalytic}(
  object,
  parametersNames,
  outcomesWithAdministration,
  outcomesWithNoAdministration
)

\S4method{defineModelEquationsFromStringToFunction}{ModelAnalyticSteadyState}(
  object,
  parametersNames,
  outcomesWithAdministration,
  outcomesWithNoAdministration
)

\S4method{defineModelEquationsFromStringToFunction}{ModelAnalyticInfusion}(
  object,
  parametersNames,
  outcomesWithAdministration,
  outcomesWithNoAdministration
)

\S4method{defineModelEquationsFromStringToFunction}{ModelAnalyticInfusionSteadyState}(
  object,
  parametersNames,
  outcomesWithAdministration,
  outcomesWithNoAdministration
)

\S4method{defineModelEquationsFromStringToFunction}{ModelODEBolus}(
  object,
  parametersNames,
  outcomesWithAdministration,
  outcomesWithNoAdministration
)

\S4method{defineModelEquationsFromStringToFunction}{ModelODEDoseInEquations}(
  object,
  parametersNames,
  outcomesWithAdministration,
  outcomesWithNoAdministration
)

\S4method{defineModelEquationsFromStringToFunction}{ModelODEDoseNotInEquations}(
  object,
  parametersNames,
  outcomesWithAdministration,
  outcomesWithNoAdministration
)

\S4method{defineModelEquationsFromStringToFunction}{ModelODEInfusionDoseInEquations}(
  object,
  parametersNames,
  outcomesWithAdministration,
  outcomesWithNoAdministration
)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{parametersNames}{Vector of parameter names.}

\item{outcomesWithAdministration}{Vector of the name of the outcome with administration.}

\item{outcomesWithNoAdministration}{Vector of the name of the outcome with no administration.}
}
\value{
....
}
\description{
defineModelEquationsFromStringToFunction
}
