% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{defineModelUserDefined}
\alias{defineModelUserDefined}
\alias{defineModelUserDefined,Model-method}
\title{Define a user defined model.}
\usage{
defineModelUserDefined(object, designs)

\S4method{defineModelUserDefined}{Model}(object, designs)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{designs}{A list of objects from the class \linkS4class{Design}.}
}
\value{
A model giving a user defined model.
}
\description{
Define a user defined model.
}
