% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R, R/ModelAnalyticInfusion.R,
%   R/ModelAnalyticInfusionSteadyState.R, R/ModelAnalyticSteadyState.R,
%   R/ModelODEBolus.R, R/ModelODEDoseInEquations.R,
%   R/ModelODEDoseNotInEquations.R, R/ModelODEInfusionDoseInEquation.R
\name{defineModelWrapper}
\alias{defineModelWrapper}
\title{define the model wrapper for the ode solver}
\arguments{
\item{model}{An object of class \code{ModelAnalytic} that defines the model.}

\item{evaluation}{An object of class \code{Evaluation} that defines the evaluation}
}
\value{
The model with wrapperModelAnalytic, functionArgumentsModelAnalytic, functionArgumentsSymbolModelAnalytic, outputNames, outcomesWithAdministration
}
\description{
define the model wrapper for the ode solver
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
