% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{defineOptimizationAlgorithm}
\alias{defineOptimizationAlgorithm}
\title{defineOptimizationAlgorithm: instantiate the algorithm from the optimizer slot}
\arguments{
\item{optimization}{An \code{Optimization} object.}
}
\value{
An optimization algorithm object.
}
\description{
defineOptimizationAlgorithm: instantiate the algorithm from the optimizer slot
}
