% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{defineParameter}
\alias{defineParameter}
\title{Define a parameter of a statistical model.}
\usage{
defineParameter(object, parameter)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{parameter}{An expression giving a parameter of the \code{StatisticalModel} object.}
}
\value{
Return \code{StatisticalModel} object with new parameters.
}
\description{
Define a parameter of a statistical model.
}
