% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{defineParameters}
\alias{defineParameters}
\title{Define the parameters of a statistical model.}
\usage{
defineParameters(object, listOfParameters)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{listOfParameters}{A list of string giving the parameters of the \code{StatisticalModel} object.}
}
\value{
Return \code{StatisticalModel} object with new parameters.
}
\description{
Define the parameters of a statistical model.
}
