% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R, R/ModelAnalyticInfusion.R,
%   R/ModelAnalyticInfusionSteadyState.R, R/ModelAnalyticSteadyState.R,
%   R/ModelODEBolus.R, R/ModelODEDoseInEquations.R,
%   R/ModelODEDoseNotInEquations.R, R/ModelODEInfusionDoseInEquation.R
\name{definePKModel}
\alias{definePKModel}
\title{Define a PK model from library of model}
\arguments{
\item{pkModel}{An object of class \code{ModelAnalytic} that defines the PK model.}

\item{pfimproject}{An object of class \code{PFIMProject} that defines the pfimproject.}
}
\description{
Define a PK model from library of model
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
