% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{defineStatisticalModel}
\alias{defineStatisticalModel}
\title{Define the \code{StatisticalModel} object of the PFIMProject object.}
\usage{
defineStatisticalModel(object, value)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}

\item{value}{A \code{StatisticalModel} or \code{ODEStatisticalModel} object.}
}
\value{
The \code{StatisticalModel} object of the PFIMProject object.
}
\description{
Define the \code{StatisticalModel} object of the PFIMProject object.
}
