% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{defineVariable}
\alias{defineVariable}
\title{Define a variable in a statistical model.}
\usage{
defineVariable(.Object, variable)
}
\arguments{
\item{.Object}{A \code{StatisticalModel} object.}

\item{variable}{A character string giving the variable to be defined.}
}
\value{
The \code{StatisticalModel} object with the new variable.
}
\description{
Define a variable in a statistical model.
}
