% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{defineVariables}
\alias{defineVariables}
\title{Define variables in a statistical model.}
\usage{
defineVariables(.Object, listOfVariables)
}
\arguments{
\item{.Object}{A \code{StatisticalModel} object.}

\item{listOfVariables}{A list of character string giving the variables to be defined.}
}
\value{
The \code{StatisticalModel} object with the new variables.
}
\description{
Define variables in a statistical model.
}
