% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\docType{class}
\name{Design-class}
\alias{Design-class}
\alias{Design}
\title{Class "Design"}
\description{
The class \code{Design} defines information concerning the parametrization of the designs.
}
\section{Objects from the class}{

Objects form the class \code{Design} can be created by calls of the form \code{Design(...)} where (...) are the parameters for the \code{Design} objects.
}

\section{Slots for the \code{Design} objects}{

\describe{
\item{\code{name}:}{A string giving the name of the design.}
\item{\code{size}:}{An integer giving the number of subjects in the design.}
\item{\code{arms}:}{A list of the arms.}
\item{\code{outcomesEvaluation}:}{A list of the results of the design evaluation for the outcomes.}
\item{\code{outcomesGradient}:}{A list of the results of the design evaluation for the sensitivity indices.}
\item{\code{numberOfArms}:}{A numeric giving the number of arms in the design.}
\item{\code{fim}:}{An object of the class \code{Fim} containing the Fisher Information Matrix of the design.}
}
}

