% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\docType{class}
\name{DesignConstraint-class}
\alias{DesignConstraint-class}
\alias{DesignConstraint}
\title{Class "DesignConstraint"}
\description{
The class \code{DesignConstraint}
defines information concerning the parametrization of the constraints on a design.
}
\section{Objects from the class \code{DesignConstraint}}{

Objects form the class \linkS4class{DesignConstraint} can be created by calls of the form \code{DesignConstraint(...)} where
(...) are the parameters for the \code{DesignConstraint} objects.
}

\section{Slots for the \code{DesignConstraint} objects}{

\describe{
\item{\code{name}:}{A character string giving the name of the design - optional.}
\item{\code{PossibleArms}:}{A list of arms for optimization.}
\item{\code{totalNumberOfIndividuals}:}{A numeric giving the total number of individuals in the design.}
\item{\code{amountOfArm}:}{A numeric giving the number of arms in the design.}
\item{\code{samplingConstraints}:}{A list giving the sampling constraints for the design.}
\item{\code{administrationConstraints}:}{A list giving the administration constraints for the design.}
}
}

