% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteConstraint.R
\docType{class}
\name{DiscreteConstraint-class}
\alias{DiscreteConstraint-class}
\alias{DiscreteConstraint}
\title{Class "DiscreteConstraint" representing the constraints for a variable}
\description{
The class \code{DiscreteConstraint} stores constraints for a variable. Constraints are given either as a continuous range
(any value between a min and a max boundary is admissible) or a discrete set of values
(any value belonging to the set is admissible)
}
\section{Objects from the class \code{DiscreteConstraint}}{

Objects form the class \code{DiscreteConstraint} can be created by calls of the form \code{DiscreteConstraint(...)} where
(...) are the parameters for the \code{DiscreteConstraint} Model.
}

\section{Slots for the \code{DiscreteConstraint} objects}{

\describe{
\item{\code{type}:}{A character string, one of 'continuous' or 'discrete'.}
\item{\code{range}:}{A numeric vector with two values giving the min/max of the continuous range.}
\item{\code{discrete}:}{A numeric vector giving the set of possible values.}
}
}

