% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{class}
\name{Distribution-class}
\alias{Distribution-class}
\alias{Distribution}
\title{Class "Distribution"}
\description{
The class defines all the required methods for a distribution object.
However, the actual functionality needs to implemented by the inheriting class.
}
