% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{.buildEvalFromOptimization}
\alias{.buildEvalFromOptimization}
\title{Helper: Build an Evaluation from an Optimization and Design}
\usage{
.buildEvalFromOptimization(optimizationObject, design)
}
\arguments{
\item{optimizationObject}{An Optimization object}

\item{design}{A Design object}
}
\value{
An Evaluation object ready to be run
}
\description{
Helper: Build an Evaluation from an Optimization and Design
}
\keyword{internal}
