% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{Evaluate}
\alias{Evaluate}
\title{Evaluate an \code{StatisticalModel} object.}
\usage{
Evaluate(object, administrations, sampling_times, cond_init, fim)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{administrations}{An \code{Administration} object.}

\item{sampling_times}{A \code{SamplingTimes} object.}

\item{cond_init}{A list for the initial conditions of the \code{StatisticalModel} object.}

\item{fim}{\code{FIM} object.}
}
\value{
A \code{fim} object giving the Fisher Information Matrix of the \code{StatisticalModel} object.
}
\description{
Evaluate an \code{StatisticalModel} object.
}
