% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R
\name{evaluateAnalyticCore}
\alias{evaluateAnalyticCore}
\title{evaluateAnalyticCore: core function to evaluate analytic model}
\usage{
evaluateAnalyticCore(model, arm)
}
\arguments{
\item{model}{An object of class \code{ModelAnalytic} that defines the model.}

\item{arm}{An object of class \code{Arm} that defines the arm.}
}
\value{
A list of dataframes containing evaluation results
}
\description{
evaluateAnalyticCore: core function to evaluate analytic model
}
\keyword{internal}
