% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{EvaluateBayesianFIM}
\alias{EvaluateBayesianFIM}
\title{Evaluate design for each arm for a Bayesian FIM.}
\usage{
EvaluateBayesianFIM(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
The \code{PFIMProject} object with the list \code{designs} that contains the evaluation of the Bayesian FIM of each design for each arm.
}
\description{
Evaluate design for each arm for a Bayesian FIM.
}
