% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{EvaluateDesignForEachArm}
\alias{EvaluateDesignForEachArm}
\title{Evaluate Design for each arm.}
\usage{
EvaluateDesignForEachArm(object, statistical_model, fim)
}
\arguments{
\item{object}{A \code{Design} object.}

\item{statistical_model}{A \code{statisticalModel} object.}

\item{fim}{A \code{fim} object.}
}
\value{
The object \code{Design} evaluated for each of its arm.
}
\description{
Evaluate Design for each arm.
}
