% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{evaluateErrorModelDerivatives}
\alias{evaluateErrorModelDerivatives}
\title{evaluate the derivatives of the model error.}
\arguments{
\item{modelError}{An object \code{ModelError} that defines the model error.}

\item{evaluationModel}{A dataframe giving the outputs for the model evaluation.}
}
\value{
The matrices sigmaDerivatives and errorVariance.
}
\description{
evaluate the derivatives of the model error.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
