% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{evaluateFim}
\alias{evaluateFim}
\title{Evaluate the Fim.}
\arguments{
\item{fim}{An object of class \code{IndividualFim}.}

\item{model}{An object of class \code{Model} containing the structural and error parameters.}

\item{arm}{An object of class \code{Arm} representing the individual's design (doses and samplings).}
}
\value{
The \code{IndividualFim} object populated with the calculated \code{fisherMatrix}.
}
\description{
Computes the Fisher Information Matrix for an individual design. This method
calculates the structural information (fixed effects) and the variance
information (residual error components) to assemble the final FIM.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
