% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{EvaluateFIMsAndDesigns}
\alias{EvaluateFIMsAndDesigns}
\title{Evaluate the FIMs and the Designs.}
\usage{
EvaluateFIMsAndDesigns(
  object,
  responseNumber,
  responseNames,
  doses,
  times,
  admin,
  samp,
  statistical_model,
  cond_init,
  totalNumberOfFIMs,
  typeFim
)
}
\arguments{
\item{object}{An \code{Optimization} object.}

\item{responseNumber}{A numeric giving the number of responses.}

\item{responseNames}{A character string giving the name of the response}

\item{doses}{A vector of numeric values giving the doses.}

\item{times}{A vector of numeric values giving the times doses.}

\item{admin}{An \code{Administration} object giving the administration parameters.}

\item{samp}{An \code{SamplingTimes} object giving the sampling times parameters.}

\item{statistical_model}{A \code{statisticalModel} object.}

\item{cond_init}{cond_init}

\item{totalNumberOfFIMs}{A numeric giving the total number of FIMs.}

\item{typeFim}{A character strgin gving the type of the FIM.}
}
\value{
An \code{Optimization} object giving the results of the evaluation of the FIMs and the Designs.
}
\description{
Evaluate the FIMs and the Designs.
}
