% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{EvaluateIndividualFIM}
\alias{EvaluateIndividualFIM}
\title{Evaluate design for each arm for a Individual FIM.}
\usage{
EvaluateIndividualFIM(object)
}
\arguments{
\item{object}{\code{PFIMProject} object.}
}
\value{
The \code{PFIMProject} object with the list \code{designs} that contains the evaluation of the Individual FIM of each design for each arm.
}
\description{
Evaluate design for each arm for a Individual FIM.
}
