% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODE.R, R/ModelODEBolus.R,
%   R/ModelODEInfusion.R
\name{evaluateInitialConditions}
\alias{evaluateInitialConditions}
\title{evaluate the initial conditions.}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{model}{A object of class \code{Model} giving the model.}

\item{doseEvent}{A data frame giving the dose event for the ode solver.}
}
\description{
evaluate the initial conditions.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
