% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{evaluateModelGradient}
\alias{evaluateModelGradient}
\title{evaluate the gradient of the model}
\arguments{
\item{model}{An object \code{Model} that defines the model.}

\item{arm}{A object \code{Arm} giving the arm}
}
\value{
A data frame that contains the gradient of the model.
}
\description{
Computes the numerical gradient of the model response with respect to the
structural parameters using the finite difference method.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
