% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfusionEquations.R
\name{EvaluateModelInfusion}
\alias{EvaluateModelInfusion}
\title{Evaluate an analytic model in infusion.}
\usage{
EvaluateModelInfusion(object, samplingTimesModel, inputsModel, computeFIM)
}
\arguments{
\item{object}{A \code{ModelInfusionEquations} object.}

\item{samplingTimesModel}{A vector giving the sampling times of the model}

\item{inputsModel}{A list containing the inputs used for the model evaluation.}

\item{computeFIM}{A boolean if the FIM is computed or not.}
}
\value{
A list containing the evaluated responses and the gradients.
}
\description{
Evaluate an analytic model in infusion.
}
