% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODEquations.R
\name{EvaluateModelODE}
\alias{EvaluateModelODE}
\title{Evaluate an ODE model.}
\usage{
EvaluateModelODE(
  object,
  samplingTimesModel,
  cond_init_ode,
  inputsModel,
  parametersGradient,
  computeFIM
)
}
\arguments{
\item{object}{An \code{ModelODEquations} object.}

\item{samplingTimesModel}{A vector containing the sampling times.}

\item{cond_init_ode}{A vector containing the initial conditions.}

\item{inputsModel}{A list containing the inputs of the models.}

\item{parametersGradient}{A list containing the gradients of the model.}

\item{computeFIM}{A boolean for computing the FIM or not.}
}
\value{
A list containing the evaluated responses and the gradients.
}
\description{
Evaluate an ODE model.
}
