% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{EvaluatePopulationFIM}
\alias{EvaluatePopulationFIM}
\title{Evaluate a design for each arm for a Population FIM.}
\usage{
EvaluatePopulationFIM(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
The \code{PFIMProject} object with the list \code{designs} that contains the evaluation of the Population FIM of each design for each arm.
}
\description{
Evaluate a design for each arm for a Population FIM.
}
