% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{EvaluateStatisticalModel}
\alias{EvaluateStatisticalModel}
\title{Evaluate a statistical model for all the administrations and all the sampling times of an arm.}
\usage{
EvaluateStatisticalModel(object, statistical_model, fim)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{statistical_model}{An object from the class \code{StatisticalModel} or \code{ODEStatisticalModel}.}

\item{fim}{Character string giving the type of the Fisher Information Matrix: "PopulationFim", "IndividualFim", "BayesianFim".}
}
\value{
A list giving the evaluated Fisher Information Matrix, the concentration,
the sensitivity indices and to sampling times used for plotting the outputs.
}
\description{
Evaluate a statistical model for all the administrations and all the sampling times of an arm.
}
