% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{evaluateVarianceFIM}
\alias{evaluateVarianceFIM}
\title{evaluateVarianceFIM: compute the variance matrix V and MFbeta for the Bayesian FIM}
\usage{
evaluateVarianceFIM(fim, model, arm, isFixedMu = NULL, isFixedOmega = NULL)
}
\arguments{
\item{fim}{An object of class \code{PopulationFim}.}

\item{model}{An object of class \code{Model}.}

\item{arm}{An object of class \code{Arm}.}

\item{isFixedMu}{Optional logical vector; \code{TRUE} for parameters with fixed or zero mu.
Computed internally if \code{NULL} (default).}

\item{isFixedOmega}{Optional logical vector; \code{TRUE} for parameters with fixed or zero omega.
Computed internally if \code{NULL} (default).}
}
\value{
List with \code{MFbeta} and \code{V}.

List with \code{MFVar} and \code{V}.

List with MFbeta and MFVar.
}
\description{
evaluateVarianceFIM: compute the variance matrix V and MFbeta for the Bayesian FIM

evaluateVarianceFIM: compute V and MFVar for the individual FIM

evaluateVarianceFIM: optimised computation of MFbeta and MFVar
}
