% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\docType{class}
\name{Evaluation-class}
\alias{Evaluation-class}
\alias{Evaluation}
\title{Class "Evaluation"}
\description{
A class storing information concerning the evaluation of a design.
}
\section{Objects from the class}{

Objects form the class \code{Evaluation} can be created by calls of the form \code{Evaluation(...)} where (...) are the parameters for the \code{Evaluation} objects.
}

\section{Slots for the \code{Evaluation} objects}{

\describe{
\item{\code{name}:}{A string giving the name of the project.}
\item{\code{model}:}{A object of class \linkS4class{Model} giving the model.}
\item{\code{modelEquations}:}{A list giving the model equations.}
\item{\code{modelParameters}:}{A list giving the model parameters.}
\item{\code{modelError}:}{A list giving the model error for each outcome of the model.}
\item{\code{outcomes}:}{A list giving the model outcomes.}
\item{\code{designs}:}{A list giving the designs to be evaluated.}
\item{\code{fim}:}{An object of the class \code{Fim} containing the Fisher Information Matrix of the design.}
\item{\code{odeSolverParameters}:}{}
}
}

