% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{EvaluationModel}
\alias{EvaluationModel}
\title{Evaluation for the model, analytic, ode, infusion}
\usage{
EvaluationModel(
  object,
  samplingTimesModel,
  cond_init_ode,
  inputsModel,
  parametersGradient,
  computeFIM
)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{samplingTimesModel}{A vector containing the sampling times of the model.}

\item{cond_init_ode}{A vector containing the initial conditions for an ODE model.}

\item{inputsModel}{A list containing the parameters used for the model evaluation.}

\item{parametersGradient}{A list containing the parameters used for the evaluation of the gradient of the model}

\item{computeFIM}{A boolean giving TRUE id the FIM is compute, FALSE otherwise.}
}
\value{
A list contaning a dataframe giving the results of the evaluation and a list giving the gradient of the model.
}
\description{
Evaluation for the model, analytic, ode, infusion
}
