% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{Exponential}
\alias{Exponential}
\title{Exponential}
\usage{
Exponential(beta = 0, combinedEffect = 0, value = 0)
}
\arguments{
\item{beta}{beta coefficient}

\item{combinedEffect}{combined effect}
}
\description{
The class \code{Exponential} is used to define Exponential model for covariates.
}
\details{
Exponential
}
