% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\docType{class}
\name{FedorovWynnAlgorithm-class}
\alias{FedorovWynnAlgorithm-class}
\alias{FedorovWynnAlgorithm}
\title{Class "FedorovWynnAlgorithm"}
\description{
Class \code{FedorovWynnAlgorithm} represents an initial variable for ODE model.
}
\section{Objects from the class \code{FedorovWynnAlgorithm}}{

Objects form the class \code{FedorovWynnAlgorithm} can be created by calls of the form \code{FedorovWynnAlgorithm(...)}
where (...) are the parameters for the \code{FedorovWynnAlgorithm} objects.
}

\section{Slots for \code{FedorovWynnAlgorithm} objects}{

\describe{
\item{\code{elementaryProtocols}:}{A list of vector for the initial elementary protocols.}
\item{\code{numberOfSubjects}:}{A vector for the number of subjects.}
\item{\code{proportionsOfSubjects}:}{A vector for the number of subjects.}
\item{\code{OptimalDesign}:}{A object Design giving the optimal Design.}
\item{\code{showProcess}:}{A boolean to show the process or not.}
\item{\code{FisherMatrix}:}{A vector giving the Fisher Information }
\item{\code{optimalFrequencies}:}{A vector of the optimal frequencies.}
\item{\code{optimalSamplingTimes}:}{A list of vectors for the optimal sampling times.}
\item{\code{optimalDoses}:}{A vector for the optimal doses.}
}
}

