% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{FedorovWynnAlgorithm}
\alias{FedorovWynnAlgorithm}
\title{FedorovWynnAlgorithm}
\usage{
FedorovWynnAlgorithm(
  name = character(0),
  modelEquations = list(),
  modelCovariatesEquation = character(0),
  modelFromLibrary = list(),
  modelParameters = list(),
  modelCovariates = list(),
  modelError = list(),
  optimizer = character(0),
  optimizerParameters = list(),
  outputs = list(),
  designs = list(),
  fimType = character(0),
  fim = Fim(),
  numberOfOccasions = 1,
  odeSolverParameters = list(),
  optimisationDesign = list(),
  optimisationAlgorithmOutputs = list(),
  elementaryProtocols = list(),
  numberOfSubjects = 0,
  proportionsOfSubjects = 0,
  showProcess = FALSE,
  FedorovWynnAlgorithmOutputs = list()
)

FedorovWynnAlgorithm(
  name = character(0),
  modelEquations = list(),
  modelCovariatesEquation = character(0),
  modelFromLibrary = list(),
  modelParameters = list(),
  modelCovariates = list(),
  modelError = list(),
  optimizer = character(0),
  optimizerParameters = list(),
  outputs = list(),
  designs = list(),
  fimType = character(0),
  fim = Fim(),
  numberOfOccasions = 1,
  odeSolverParameters = list(),
  optimisationDesign = list(),
  optimisationAlgorithmOutputs = list(),
  elementaryProtocols = list(),
  numberOfSubjects = 0,
  proportionsOfSubjects = 0,
  showProcess = FALSE,
  FedorovWynnAlgorithmOutputs = list()
)
}
\arguments{
\item{name}{A string giving the name of the design evaluation.}

\item{modelEquations}{A list giving the model equations.}

\item{modelCovariatesEquation}{modelCovariatesEquation}

\item{modelFromLibrary}{A list giving the model equations from the library of model.}

\item{modelParameters}{A list giving the model parameters.}

\item{modelCovariates}{A list giving the model covariates.}

\item{modelError}{A list giving the model error.}

\item{optimizer}{A string giving the name of the optimization algorithm being used.}

\item{optimizerParameters}{A list giving the parameters of the optimization algorithm.}

\item{outputs}{A list giving the model outputs.}

\item{designs}{A list giving the designs to be evaluated.}

\item{fimType}{A string giving the type of Fim being evaluated.}

\item{fim}{A object \code{Fim} giving the Fim.}

\item{numberOfOccasions}{numberOfOccasions}

\item{odeSolverParameters}{A list giving the atol and rtol parameters for the ode solver.}

\item{optimisationDesign}{A list giving the evaluation of initial and optimal design.}

\item{optimisationAlgorithmOutputs}{A list giving the outputs of the optimization process.}

\item{elementaryProtocols}{List of elementary protocols.}

\item{numberOfSubjects}{Total number of subjects.}

\item{proportionsOfSubjects}{Initial proportion vector for each protocol.}

\item{showProcess}{Logical; print iteration progress.}

\item{FedorovWynnAlgorithmOutputs}{Output list from the Rcpp routine.}
}
\description{
The class \code{FedorovWynnAlgorithm} implements the FedorovWynn algorithm.

Implements the Fedorov-Wynn algorithm for optimal design.
}
\details{
FedorovWynnAlgorithm: S7 class
}
