% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillLibraryOfModels.R
\docType{data}
\name{FillLibraryOfModels}
\alias{FillLibraryOfModels}
\alias{PFIMLibraryOfModels}
\title{Function "FillLibraryOfModels"}
\format{
An object of class \code{LibraryOfModels} of length 1.
}
\usage{
PFIMLibraryOfModels
}
\description{
This function is used to define the models PK, PD and PKPD models and load these models in the library of models.
}
\keyword{datasets}
