% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\docType{class}
\name{Fim-class}
\alias{Fim-class}
\alias{Fim}
\title{Class "Fim"}
\description{
A class storing information regarding the Fisher matrix.
Type of the Fisher information: population ("PopulationFIM"), individual ("IndividualFIM") or Bayesian ("BayesianFIM").
}
\section{Objects from the class}{

Objects form the class \code{Fim} can be created by calls of the form \code{Fim(...)} where
(...) are the parameters for the \code{Fim} objects.
}

\section{Slots for \code{Fim} objects}{

\describe{
\item{\code{fisherMatrix}:}{A matrix giving the Fisher matrix.}
\item{\code{fixedEffects}:}{A matrix giving the fixed effects of the Fisher matrix.}
\item{\code{varianceEffects}:}{A matrix giving the variance effects of the Fisher matrix.}
\item{\code{shrinkage}:}{A vector giving the shrinkage value of the parameters.}
}
}

