% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/PopulationFim.R
\docType{methods}
\name{FinalizeFIMForOneElementaryDesign}
\alias{FinalizeFIMForOneElementaryDesign}
\alias{FinalizeFIMForOneElementaryDesign,PopulationFim-method}
\title{FinalizeFIMForOneElementaryDesign}
\usage{
FinalizeFIMForOneElementaryDesign(object, arm)

\S4method{FinalizeFIMForOneElementaryDesign}{PopulationFim}(object, arm)
}
\arguments{
\item{object}{A \code{Fim} object.}

\item{arm}{A \code{Arm} object.}
}
\value{
A matrix of numeric \code{mfisher} giving the Fisher Information Matrix.

The Fim times size of the arm.
}
\description{
FinalizeFIMForOneElementaryDesign

Finalize the Fim for one elementary design.
}
