% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexAlgorithm.R
\name{fisher.simplex}
\alias{fisher.simplex}
\title{Compute the fisher.simplex}
\usage{
fisher.simplex(simplex, optimizationObject, outcomes)
}
\arguments{
\item{simplex}{A list giving the parameters of the simplex.}

\item{optimizationObject}{An object from the class \linkS4class{Optimization}.}

\item{outcomes}{A vector giving the outcomes of the arms.}
}
\value{
A list giving the results of the optimization.
}
\description{
Compute the fisher.simplex
}
