% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexAlgorithm.R
\name{fisherSimplex}
\alias{fisherSimplex}
\title{Compute the fisher.simplex}
\arguments{
\item{simplex}{A list giving the parameters of the simplex.}

\item{optimizationObject}{An object \code{Optimization}.}

\item{outcomes}{A vector giving the outcomes of the arms.}
}
\value{
A list giving the results of the optimization.
}
\description{
Compute the fisher.simplex
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
