% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraint.R
\name{fixedDoses}
\alias{fixedDoses}
\title{Set the value for the fixed doses in the administration constraints.}
\usage{
fixedDoses(object, value)
}
\arguments{
\item{object}{An object \code{AdministrationConstraint} from the class \linkS4class{AdministrationConstraint}.}

\item{value}{A numeric vector giving the value of the fixed dose.}
}
\value{
The \code{AdministrationConstraint} object with its new value of the fixed dose.
}
\description{
Set the value for the fixed doses in the administration constraints.
}
