% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\name{FixTimeValues}
\alias{FixTimeValues}
\title{Set the value for the fixed times.}
\usage{
FixTimeValues(object, value)
}
\arguments{
\item{object}{A \code{SamplingConstraint} object.}

\item{value}{A vector of numeric giving the values for the fixed times.}
}
\value{
The \code{SamplingConstraint} object with the values for the fixed times.
}
\description{
Set the value for the fixed times.
}
