% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{g}
\alias{g}
\title{Evaluation of the model error.}
\usage{
g(object, f_x_i_theta)
}
\arguments{
\item{object}{\code{ModelError} object.}

\item{f_x_i_theta}{the nonlinear structural model \code{f_x_i_theta}.}
}
\value{
A numeric giving the evaluation of the error model.
}
\description{
Evaluation of the model error.
}
