% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{generateDosesCombination}
\alias{generateDosesCombination}
\title{Generate dose combinations for optimization}
\arguments{
\item{design}{An object \code{Design}.}
}
\value{
A list containing the combinations of doses and the total count.
}
\description{
Generate dose combinations for optimization
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
