% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{generateReportOptimization}
\alias{generateReportOptimization}
\title{Generate the HTML Report for Design Optimization}
\arguments{
\item{fim}{An object of class \code{\link{PopulationFim}} containing the
Fisher Information Matrix of the final optimized design.}

\item{optimizationAlgorithm}{An object of class \code{\link{MultiplicativeAlgorithm},\link{FedorovWynnAlgorithm}
\link{SimplexAlgorithm},\link{PSOAlgorithm},\link{PGBOAlgorithm}}.}

\item{tablesForReport}{A \code{list} of data frames (as returned by
\code{\link{tablesForReport}}) containing the final optimized statistical summaries.}
}
\value{
An HTML report file (or a path to the file) containing the detailed
optimization results and diagnostic plots.
}
\description{
The \code{generateReportOptimization} method compiles the results of a design
optimization into a professional HTML document. It specifically handles the
output of the \code{\link{MultiplicativeAlgorithm}}, documenting the
transition from the initial design to the optimal sampling schedule.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
