% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{generateSamplingTimesCombination}
\alias{generateSamplingTimesCombination}
\title{Generate sampling time combinations}
\arguments{
\item{design}{An object \code{Design}.}
}
\value{
A list of possible sampling time combinations for each arm.
}
\description{
Generate sampling time combinations
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
