% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{generateTables}
\alias{generateTables}
\title{Generate the tables for the report.}
\usage{
generateTables(object, plotOptions)
}
\arguments{
\item{object}{An object from the class \linkS4class{PFIMProject}.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
A list giving the kable able for the report ( evaluation and optimization).
}
\description{
Generate the tables for the report.
}
