% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R, R/LogNormal.R, R/Normal.R
\name{getAdjustedGradient}
\alias{getAdjustedGradient}
\alias{getAdjustedGradient,LogNormal-method}
\alias{getAdjustedGradient,Normal-method}
\title{getAdjustedGradient}
\usage{
getAdjustedGradient(object, outcomesGradient)

\S4method{getAdjustedGradient}{LogNormal}(object, outcomesGradient)

\S4method{getAdjustedGradient}{Normal}(object, outcomesGradient)
}
\arguments{
\item{object}{An object \code{distribution} from the class \linkS4class{Distribution}.}

\item{outcomesGradient}{A list containing the evaluation of the outcome gradients.}
}
\value{
A list giving the adjusted gradient.
}
\description{
Get the adjusted gradient.
}
