% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getAdministrationByOutcome}
\alias{getAdministrationByOutcome}
\title{Get the parameters of the administration for an arm given the response of the model.}
\usage{
getAdministrationByOutcome(object, outcome)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{outcome}{A character string giving the the name of the response of the model.}
}
\value{
A list of objects from \code{Administration} class giving the parameters of the
administration for the object \code{Arm}.
}
\description{
Get the parameters of the administration for an arm given the response of the model.
}
