% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\name{getAdministrationConstraint}
\alias{getAdministrationConstraint}
\title{Get the constraints on the administration for a  \code{DesignConstraint} object.}
\usage{
getAdministrationConstraint(object)
}
\arguments{
\item{object}{A \code{DesignConstraint} object.}
}
\value{
The list of constraints on the administration given by \code{administrationConstraint} for a  \code{DesignConstraint} object.
}
\description{
Get the constraints on the administration for a  \code{DesignConstraint} object.
}
