% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getAdministrations}
\alias{getAdministrations}
\title{Get all the administration for an arm.}
\usage{
getAdministrations(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
A list \code{administrations} of objects from the class \code{Administration} class giving
the parameters of the administration for the object \code{Arm}.
}
\description{
Get all the administration for an arm.
}
