% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getAdministrationsConstraints}
\alias{getAdministrationsConstraints}
\alias{getAdministrationsConstraints,Arm-method}
\title{getAdministrationsConstraints}
\usage{
getAdministrationsConstraints(object)

\S4method{getAdministrationsConstraints}{Arm}(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
The list \code{administrationsConstraints}.
}
\description{
Get the administrations constraints.
}
