% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\name{getallowedContinuousSamplingTimes}
\alias{getallowedContinuousSamplingTimes}
\title{Get the allowed Continuous SamplingTimes}
\usage{
getallowedContinuousSamplingTimes(object)
}
\arguments{
\item{object}{A \code{SamplingConstraint} object.}
}
\value{
A list giving the allowed Continuous SamplingTimes for the \code{SamplingConstraint} object.
}
\description{
Get the allowed Continuous SamplingTimes
}
