% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\name{getallowedDiscretSamplingTimes}
\alias{getallowedDiscretSamplingTimes}
\title{Get the allowed discret sampling simes}
\usage{
getallowedDiscretSamplingTimes(object)
}
\arguments{
\item{object}{\code{SamplingConstraint} object.}
}
\value{
The allowed discret sampling times of the \code{SamplingConstraint} object.
}
\description{
Get the allowed discret sampling simes
}
