% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{getAllowedDose}
\alias{getAllowedDose}
\title{Get the constraints on allowed dose}
\usage{
getAllowedDose(object)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}
}
\value{
The vector \code{allowed_dose} of the numeric values of the constraints on allowed dose.
}
\description{
Get the constraints on allowed dose
}
