% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraint.R
\name{getAllowedDoses}
\alias{getAllowedDoses}
\title{Get the vector of allowed amount of dose.}
\usage{
getAllowedDoses(object)
}
\arguments{
\item{object}{An object \code{AdministrationConstraint} from the class \linkS4class{AdministrationConstraint}.}
}
\value{
A vector \code{AllowedDoses} giving the allowed amount of dose.
}
\description{
Get the vector of allowed amount of dose.
}
