% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{getAllowedTime}
\alias{getAllowedTime}
\title{Get the constraints on allowed times.}
\usage{
getAllowedTime(object)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}
}
\value{
The numeric vector \code{allowed_time} giving the constraints on allowed times.
}
\description{
Get the constraints on allowed times.
}
