% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{getAllowedTinf}
\alias{getAllowedTinf}
\title{Get the constraints on Tinf.}
\usage{
getAllowedTinf(object)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}
}
\value{
The vector \code{allowed_tinf} giving the constraints on Tinf.
}
\description{
Get the constraints on Tinf.
}
