% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getAmountOfArms}
\alias{getAmountOfArms}
\title{Get the amount of arms in a Design.}
\usage{
getAmountOfArms(object)
}
\arguments{
\item{object}{A \code{Design} object.}
}
\value{
A numeric \code{amountOfArm} giving the number of arms in the design.
}
\description{
Get the amount of arms in a Design.
}
