% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getArmConstraints}
\alias{getArmConstraints}
\title{Get arm constraints for optimization algorithms}
\usage{
getArmConstraints(arm, optimizationAlgorithm, ...)
}
\arguments{
\item{arm}{An object of class \code{\link{Arm}}.}

\item{optimizationAlgorithm}{An object of class \code{MultiplicativeAlgorithm},
\code{FedorovWynnAlgorithm}, \code{SimplexAlgorithm}, \code{PSOAlgorithm}, or \code{PGBOAlgorithm}.}

\item{...}{Additional arguments}
}
\value{
A list of constraint entries, one per sampling outcome.
}
\description{
Extracts administration and sampling time constraints formatted for the
\code{MultiplicativeAlgorithm}, \code{FedorovWynnAlgorithm},
\code{SimplexAlgorithm}, \code{PSOAlgorithm}, or \code{PGBOAlgorithm}.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
