% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getArmData}
\alias{getArmData}
\title{Extract arm data for reporting}
\usage{
getArmData(arm, ...)
}
\arguments{
\item{arm}{An object of class \code{Arm}.}

\item{...}{Additional arguments.}
}
\value{
A list of named lists, one per sampling outcome.
}
\description{
Extracts arm-level design information (name, size, outcomes, doses, sampling
times) formatted for inclusion in HTML reports.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
