% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Design.R,
%   R/OptimizationAlgorithm.R
\name{getArms}
\alias{getArms}
\alias{getArms,Design-method}
\alias{getArms,OptimizationAlgorithm-method}
\title{Get the arms of an object.}
\usage{
getArms(object)

\S4method{getArms}{Design}(object)

\S4method{getArms}{OptimizationAlgorithm}(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
A list containing the arms of the object.
}
\description{
Get the arms of an object.
}
