% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getArmSize}
\alias{getArmSize}
\title{Get the size of an arm.}
\usage{
getArmSize(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
A numeric \code{arm_size} giving the size of the object \code{Arm}.
}
\description{
Get the size of an arm.
}
