% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate.R
\name{getCategoryOfReference}
\alias{getCategoryOfReference}
\title{getCategoryOfReference: get the category of reference}
\arguments{
\item{categoricalCovariate}{A object of class \code{CategoricalCovariate} giving the categorical covariate.}
}
\value{
A string giving the category of reference.
}
\description{
getCategoryOfReference: get the category of reference
}
