% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getcError}
\alias{getcError}
\title{Get the parameter c.}
\usage{
getcError(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelError}.}
}
\value{
A numeric giving the parameter c.
}
\description{
Get the parameter c.
}
