% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{getColumnAndParametersNamesFIM}
\alias{getColumnAndParametersNamesFIM}
\alias{getColumnAndParametersNamesFIM,BayesianFim-method}
\alias{getColumnAndParametersNamesFIM,IndividualFim-method}
\alias{getColumnAndParametersNamesFIM,PopulationFim-method}
\title{Get the names of the names of the parameters associated to each column of the fim.}
\usage{
getColumnAndParametersNamesFIM(object, model)

\S4method{getColumnAndParametersNamesFIM}{BayesianFim}(object, model)

\S4method{getColumnAndParametersNamesFIM}{IndividualFim}(object, model)

\S4method{getColumnAndParametersNamesFIM}{PopulationFim}(object, model)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{model}{An object from the class \linkS4class{Model}.}
}
\value{
A list giving the names of the parameters associated to each column of the fim.
}
\description{
Get the names of the names of the parameters associated to each column of the fim.
}
