% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getColumnAndParametersNamesFIMInLatex}
\alias{getColumnAndParametersNamesFIMInLatex}
\title{#' Get the names of the names of the parameters associated to each column of the fim in Latex format.}
\usage{
getColumnAndParametersNamesFIMInLatex(object, model)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{model}{An object from the class \linkS4class{Model}.}
}
\value{
A list giving the names of the parameters associated to each column of the fim in Latex format.
}
\description{
#' Get the names of the names of the parameters associated to each column of the fim in Latex format.
}
