% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getCondInit}
\alias{getCondInit}
\title{Get the initial conditions in a arm for an ODE model}
\usage{
getCondInit(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}}
}
\value{
A list \code{cond_init} giving the initial conditions for ODE model in the object \code{Arm}
}
\description{
Get the initial conditions in a arm for an ODE model
}
