% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getConditionNumberMatrix}
\alias{getConditionNumberMatrix}
\title{Get the Condition Number Matrix of the Fisher Information Matrix for a \code{Fim} object..}
\usage{
getConditionNumberMatrix(object, FixedEffectParameterNumber)
}
\arguments{
\item{object}{A \code{Fim} object.}

\item{FixedEffectParameterNumber}{A numerical giving the number of Fixed Effect Parameters.}
}
\value{
A matrix \code{conditionNumbers} of numerical values giving the Condition Number Matrix the min, max and min/max for the FixedEffects and VarianceComponents.
}
\description{
Get the Condition Number Matrix of the Fisher Information Matrix for a \code{Fim} object..
}
