% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{getContent}
\alias{getContent}
\title{Get content of a library of models.}
\usage{
getContent(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{LibraryOfModels}.}
}
\value{
A list giving the content of the library of models.
}
\description{
Get content of a library of models.
}
