% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{getContentsLibraryOfModels}
\alias{getContentsLibraryOfModels}
\title{Get the content of the \code{LibraryOfModels} object.}
\usage{
getContentsLibraryOfModels(object)
}
\arguments{
\item{object}{A \code{LibraryOfModels} object.}
}
\value{
A list \code{contentsLibraryOfModels} giving the two lists that respectively corresponds to the two librairies of the PK and PD models contained in the \code{\link{LibraryOfModels}}.
}
\description{
Get the content of the \code{LibraryOfModels} object.
}
