% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getCorr}
\alias{getCorr}
\title{Get the correlation matrix of the Fisher Information Matrix for a \code{Fim} object.}
\usage{
getCorr(object)
}
\arguments{
\item{object}{A \code{Fim} object.}
}
\value{
A matrix of numerical values \code{corr_mat} giving the correlation matrix from the Fisher Information Matrix.
}
\description{
Get the correlation matrix of the Fisher Information Matrix for a \code{Fim} object.
}
