% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\title{getCorrelationMatrix: return the correlation matrix of the FIM}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The Dcriterion
}
\description{
getCorrelationMatrix: return the correlation matrix of the FIM

getCorrelationMatrix : get the correlation matrix
}
