% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalCovariate.R,
%   R/CategoricalCovariateWithIOV.R
\name{getCovariateEffects}
\alias{getCovariateEffects}
\title{Compute the effects of a covariate without iov.}
\arguments{
\item{categoricalCovariate}{A object \code{CategoricalCovariate} giving the catecovariate}

\item{covariate}{A object \code{CategoricalCovariateWithIOV} giving the covariate with iov}
}
\value{
The effect of a covariate without iov.

The effect of a covariate with iov.
}
\description{
Compute the effects of a covariate without iov.

Compute the effects of a covariate with iov.
}
