% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getDataForArmEvaluation}
\alias{getDataForArmEvaluation}
\alias{getDataForArmEvaluation,Arm-method}
\title{getDataForArmEvaluation}
\usage{
getDataForArmEvaluation(object)

\S4method{getDataForArmEvaluation}{Arm}(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
A list containing the data for arm evaluation.
}
\description{
getDataForArmEvaluation
}
