% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getDcriterion}
\alias{getDcriterion}
\title{getDcriterion: return the D-criterion of the FIM}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The Dcriterion
}
\description{
getDcriterion: return the D-criterion of the FIM

getDcriterion : get the Dcriterion
}
