% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEquations.R
\docType{methods}
\name{getDerivate}
\alias{getDerivate}
\alias{getDerivate,ModelEquations-method}
\title{Get the derivate of an equation of a \code{ModelEquations} object.}
\usage{
getDerivate(object, equationName, parameter)

\S4method{getDerivate}{ModelEquations}(object, equationName, parameter)
}
\arguments{
\item{object}{A \code{ModelEquations} object.}

\item{equationName}{A character string giving the name of the response of the equations.}

\item{parameter}{An \code{ModelParameter} object.}
}
\value{
A list of expression of the derivate of an equation of a \code{ModelEquations}
}
\description{
Get the derivate of an equation of a \code{ModelEquations} object.
}
