% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{getDerivatesAdjustedByDistribution}
\alias{getDerivatesAdjustedByDistribution}
\title{Get the derivates adjusted by distribution of a \code{ModelParameter} object.}
\usage{
getDerivatesAdjustedByDistribution(object, df_total)
}
\arguments{
\item{object}{\code{ModelParameter} object.}

\item{df_total}{df_total}
}
\value{
A list of expression giving the derivates adjusted by distribution \code{distribution},
the mu \code{distribution} and the dftotal \code{distribution} of a \code{ModelParameter} object.
}
\description{
Get the derivates adjusted by distribution of a \code{ModelParameter} object.
}
