% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODEquations.R
\name{getDerivatives}
\alias{getDerivatives}
\title{Get the derivatives of a \code{ModelODEquations} object.}
\usage{
getDerivatives(object)
}
\arguments{
\item{object}{An \code{ModelODEquations} object.}
}
\value{
A list of expression \code{derivatives} giving the derivatives of a \code{ModelODEquations} object.
}
\description{
Get the derivatives of a \code{ModelODEquations} object.
}
