% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/IndividualFim.R, R/BayesianFim.R,
%   R/PopulationFim.R
\docType{methods}
\name{getDescription}
\alias{getDescription}
\alias{getDescription,IndividualFim-method}
\alias{getDescription,BayesianFim-method}
\alias{getDescription,PopulationFim-method}
\title{Get the description of FIM.}
\usage{
getDescription(object)

\S4method{getDescription}{IndividualFim}(object)

\S4method{getDescription}{BayesianFim}(object)

\S4method{getDescription}{PopulationFim}(object)
}
\arguments{
\item{object}{A \code{Fim} object.}
}
\value{
A character string that tells you that is a Fisher information matrix.

Return a string giving the type of the Fim.

A string giving the description of the object \code{BayesianFim}.

A string giving the type of the Fim.
}
\description{
Get the description of FIM.

Get the description \code{BayesianFim} object.

Get the type of the Fim.
}
