% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getDesign}
\alias{getDesign}
\title{Get the design of \code{PFIMProject} object.}
\usage{
getDesign(object)
}
\arguments{
\item{object}{\code{PFIMProject} object.}
}
\value{
The list \code{design} of the designs in the \code{PFIMProject} object.
}
\description{
Get the design of \code{PFIMProject} object.
}
