% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getDesigns}
\alias{getDesigns}
\title{Get the designs.}
\usage{
getDesigns(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{PFIMProject}.}
}
\value{
A list giving the designs of the object.
}
\description{
Get the designs.
}
