% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getDeterminant}
\alias{getDeterminant}
\title{getDeterminant: return the determinant of the FIM}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The determinant
}
\description{
getDeterminant: return the determinant of the FIM

getDeterminant: get the determinant
}
