% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteConstraint.R
\name{getDiscret}
\alias{getDiscret}
\title{Get the set of possible values for a \code{DiscreteConstraint} object.}
\usage{
getDiscret(object)
}
\arguments{
\item{object}{A \code{DiscreteConstraint} object.}
}
\value{
A numeric vector \code{discret} giving the set of possible values.
}
\description{
Get the set of possible values for a \code{DiscreteConstraint} object.
}
