% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{getDistribution}
\alias{getDistribution}
\title{Get the distribution of a \code{ModelParameter} object.}
\usage{
getDistribution(object)
}
\arguments{
\item{object}{\code{ModelParameter} object.}
}
\value{
The distribution given by \code{distribution} of a \code{ModelParameter} object.
}
\description{
Get the distribution of a \code{ModelParameter} object.
}
