% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R, R/AdministrationConstraints.R
\name{getDose}
\alias{getDose}
\alias{getDose,Administration-method}
\alias{getDose,AdministrationConstraints-method}
\title{getDose}
\usage{
getDose(object)

\S4method{getDose}{Administration}(object)

\S4method{getDose}{AdministrationConstraints}(object)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}
}
\value{
The numeric \code{amount_dose} giving the amount of doses.
}
\description{
Get the amount of doses.
}
