% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraint.R
\name{getDoseOptimisability}
\alias{getDoseOptimisability}
\title{Get the boolean \code{Optimisability} for optimizable dose.}
\usage{
getDoseOptimisability(object)
}
\arguments{
\item{object}{An object \code{AdministrationConstraint} from the class \linkS4class{AdministrationConstraint}.}
}
\value{
The boolean \code{Optimisability} giving FALSE for fixed dose, TRUE for an optimizable dose.
}
\description{
Get the boolean \code{Optimisability} for optimizable dose.
}
