% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getDVSigma}
\alias{getDVSigma}
\title{Get the DV Sigma of a \code{ModelError} object.}
\usage{
getDVSigma(object, parameter)
}
\arguments{
\item{object}{A \code{ModelError} object.}

\item{parameter}{An string giving a parameter of the model error.}
}
\value{
A list giving the \code{derivates} Sigma for a paramerer.
}
\description{
Get the DV Sigma of a \code{ModelError} object.
}
