% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getEigenValue}
\alias{getEigenValue}
\title{Get the eigen values of the Fisher Information Matrix for a \code{Fim} object.}
\usage{
getEigenValue(object)
}
\arguments{
\item{object}{A \code{Fim} object.}
}
\value{
A vector of numerical values \code{EV} giving the eigen values of the \code{Fim} object.
}
\description{
Get the eigen values of the Fisher Information Matrix for a \code{Fim} object.
}
