% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{getElementaryProtocols}
\alias{getElementaryProtocols}
\title{Get the matrix of all the combination of the elementary protocols.}
\usage{
getElementaryProtocols(object)
}
\arguments{
\item{object}{An \code{Optimization} object.}
}
\value{
A matrix giving  all the combination of the elementary protocols.
}
\description{
Get the matrix of all the combination of the elementary protocols.
}
