% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R, R/ModelEquations.R
\name{getEquation}
\alias{getEquation}
\alias{getEquation,ModelEquations-method}
\title{Get the equation of a \code{ModelError} object by their names.}
\usage{
getEquation(object, equationName)

\S4method{getEquation}{ModelEquations}(object, equationName)
}
\arguments{
\item{object}{A \code{ModelEquations} object.}

\item{equationName}{A character string giving the name of the response of the equations.}
}
\value{
An expression \code{equation} giving the equation of the model error.

An expression  \code{equations} giving the equation of a model with respect to its name.
}
\description{
Get the equation of a \code{ModelError} object by their names.

Get the equation of a \code{ModelEquations} object with respect to its name.
}
