% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEquations.R, R/Model.R, R/PKModel.R,
%   R/PKPDModel.R
\name{getEquations}
\alias{getEquations}
\alias{getEquations,Model-method}
\alias{getEquations,PKModel-method}
\alias{getEquations,PKPDModel-method}
\title{Get the equations of a \code{ModelEquations} object.}
\usage{
getEquations(object)

\S4method{getEquations}{Model}(object)

\S4method{getEquations}{PKModel}(object)

\S4method{getEquations}{PKPDModel}(object)
}
\arguments{
\item{object}{A \code{PKPDModel} object.}
}
\value{
A list of expression \code{equations} giving the equations of the model after the change of variable in the model.

A list of expression giving the equations of a \code{Model} object after the changing variable in PK model.

A list of expressions giving the equations of a \code{PKModel} object.

A list of expressions giving the equations of a \code{PKPDModel} object.
}
\description{
Get the equations of a \code{ModelEquations} object.

Get the equations of a \code{Model} object after changing variable in PK model.

Get the equations of a \code{PKModel} object.

Get the equations of a \code{PKPDModel} object.
}
