% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfusion.R
\name{getEquationsAfterInfusion}
\alias{getEquationsAfterInfusion}
\title{Get the equations after infusion.}
\usage{
getEquationsAfterInfusion(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
A list giving the equations after the infusion.
}
\description{
Get the equations after infusion.
}
