% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getEquationsModel}
\alias{getEquationsModel}
\title{Get the equations of a \code{Model} object.}
\usage{
getEquationsModel(object)
}
\arguments{
\item{object}{A \code{Model} object.}
}
\value{
A list of expression \code{equationsModel} giving the equations of a \code{Model} object.
}
\description{
Get the equations of a \code{Model} object.
}
