% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEquations.R, R/ModelODEquations.R
\name{getEquationsModelPKPD}
\alias{getEquationsModelPKPD}
\title{Get the equations of the PK and PD models of a \code{ModelEquations} object.}
\usage{
getEquationsModelPKPD(modelEquationsPKmodel, modelEquationsPDmodel)

getEquationsModelPKPD(modelEquationsPKmodel, modelEquationsPDmodel)
}
\arguments{
\item{modelEquationsPKmodel}{An expression giving the equation of the PK model.}

\item{modelEquationsPDmodel}{An expression giving the equation of the PD model.}
}
\value{
A list \code{output} giving:
\itemize{
\item{expressions for the equations of the PK and PD models}{, for an analytic PK and PD models.}
\item{expressions for the equations of the PK and the infusion equations PD models}{, for a PK model in infusion.}
\item{expressions for the equations of the PK and PD models}{, for an ODE PK and PD models.}
}
}
\description{
Get the equations of the PK and PD models of a \code{ModelEquations} object.

getEquationsModelPKPD
}
