% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{getEquationsStatisticalModel}
\alias{getEquationsStatisticalModel}
\title{Get the equations of a statistical model.}
\usage{
getEquationsStatisticalModel(object)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}
}
\value{
A \code{ModelEquationsg} object giving the equations of the \code{StatisticalModel} object.
}
\description{
Get the equations of a statistical model.
}
