% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getErrorModelParameters}
\alias{getErrorModelParameters}
\title{Get parameters of the error model of a \code{ModelError} object.}
\usage{
getErrorModelParameters(object)
}
\arguments{
\item{object}{A \code{ModelError} object.}
}
\value{
A list of string giving the parameters of the error model.
}
\description{
Get parameters of the error model of a \code{ModelError} object.
}
