% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{getErrorModelStandardErrors}
\alias{getErrorModelStandardErrors}
\title{Get the SE and RSE of the parameters.}
\usage{
getErrorModelStandardErrors(object, fim)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{fim}{A \code{Fim} object giving the Fisher Information Matrix.}
}
\value{
A dataframe giving he SE and RSE of the parameters.
}
\description{
Get the SE and RSE of the parameters.
}
