% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getEvaluationDesign}
\alias{getEvaluationDesign}
\title{Get the evaluated concentration and sensitivity indices of a design.}
\usage{
getEvaluationDesign(object)
}
\arguments{
\item{object}{A \code{Design} object.}
}
\value{
The object \code{Design} evaluated for each of its arm.
}
\description{
Get the evaluated concentration and sensitivity indices of a design.
}
