% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{getEvaluationInitialDesignResults}
\alias{getEvaluationInitialDesignResults}
\title{Get the evaluation results of the initial design.}
\usage{
getEvaluationInitialDesignResults(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Optimization}.}
}
\value{
The object from the class \linkS4class{Evaluation} giving the results of the evaluation of the initial design.
}
\description{
Get the evaluation results of the initial design.
}
