% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getEvaluationResponses}
\alias{getEvaluationResponses}
\title{Get the evaluated responses of the model.}
\usage{
getEvaluationResponses(object)
}
\arguments{
\item{object}{A \code{Design} object.}
}
\value{
The object \code{Design} evaluated for each of its arm.
}
\description{
Get the evaluated responses of the model.
}
