% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{getFim}
\alias{getFim}
\title{getFim}
\usage{
getFim(evaluation, ...)
}
\arguments{
\item{evaluation}{An object of class \code{Evaluation}.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with \code{fisherMatrix}, \code{fixedEffects},
  and \code{varianceEffects}.
}
\description{
Extract the Fisher Information Matrix from an \code{Evaluation} object.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{
Fim = getFim( evaluationPopulationFIMResults )
print( Fim )
}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
