% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getFimOfDesign}
\alias{getFimOfDesign}
\title{Get the Fisher Information Matrix of a design.}
\usage{
getFimOfDesign(object)
}
\arguments{
\item{object}{A \code{Design} object.}
}
\value{
A \code{Fim} object giving the Fisher Information Matrix of a design.
}
\description{
Get the Fisher Information Matrix of a design.
}
