% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getFims}
\alias{getFims}
\title{Get the Fisher Information Matrices.}
\usage{
getFims(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
A list \code{fimList} giving the Fisher Information Matrices for all the designs of a PFIMProject project.
}
\description{
Get the Fisher Information Matrices.
}
