% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getFisherMatrices}
\alias{getFisherMatrices}
\title{Get the fim matrices from all designs of a PFIMProject object.}
\usage{
getFisherMatrices(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
A list of matrices \code{fimList} giving the Fisher Information Matrix of all the designs of a PFIMProject project.
}
\description{
Get the fim matrices from all designs of a PFIMProject object.
}
