% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getFixedEffects}
\alias{getFixedEffects}
\title{Get the matrix of fixed effects.}
\usage{
getFixedEffects(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
The matrix of the fixed effects.
}
\description{
Get the matrix of fixed effects.
}
