% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{getFixedMu}
\alias{getFixedMu}
\title{Get the fixed effect.}
\usage{
getFixedMu(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelParameter}.}
}
\value{
A boolean giving the fixed mu.
}
\description{
Get the fixed effect.
}
