% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{getFixedOmega}
\alias{getFixedOmega}
\alias{getFixedOmega,ModelParameter-method}
\title{Get the fixed variance.}
\usage{
getFixedOmega(object)

\S4method{getFixedOmega}{ModelParameter}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelParameter}.}
}
\value{
A boolean giving the fixed omega.
}
\description{
Get the fixed variance.
}
