% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfusionEquations.R
\name{getInfusionEquations}
\alias{getInfusionEquations}
\title{Get the Infusion Equations.}
\usage{
getInfusionEquations(object)
}
\arguments{
\item{object}{A \code{ModelInfusionEquations} object.}
}
\value{
A list \code{equations} of the expressions giving the infusion equations of the  \code{ModelInfusionEquations} object
}
\description{
Get the Infusion Equations.
}
