% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R, R/Model.R
\name{getInitialConditions}
\alias{getInitialConditions}
\alias{getInitialConditions,Arm-method}
\alias{getInitialConditions,Model-method}
\title{getInitialConditions}
\usage{
getInitialConditions(object)

\S4method{getInitialConditions}{Arm}(object)

\S4method{getInitialConditions}{Model}(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
The list \code{initialConditions} for the object \code{Arm}.
}
\description{
Get the initial condition for the evaluation of an ode model.
}
