% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\name{getInitialTime}
\alias{getInitialTime}
\title{Get the initial time of a \code{SamplingTimes} object.}
\usage{
getInitialTime(object)
}
\arguments{
\item{object}{\code{SamplingTimes} object.}
}
\value{
A numeric \code{initialTime} giving the inital time of a \code{SamplingTimes} object.
}
\description{
Get the initial time of a \code{SamplingTimes} object.
}
