% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationAlgorithm.R
\name{getIterationAndCriteria}
\alias{getIterationAndCriteria}
\title{Get the iteration with the convergence criteria.}
\usage{
getIterationAndCriteria(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{OptimizationAlgorithm}.}
}
\value{
A dataframe giving the iteration with the convergence criteria.
}
\description{
Get the iteration with the convergence criteria.
}
