% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{getLibraryPDModels}
\alias{getLibraryPDModels}
\alias{getLibraryPDModels,LibraryOfModels-method}
\title{Get the library of PD models.}
\usage{
getLibraryPDModels(object)

\S4method{getLibraryPDModels}{LibraryOfModels}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{LibraryOfModels}.}
}
\value{
A list giving the PD models.
}
\description{
Get the library of PD models.
}
