% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{getLibraryPKModels}
\alias{getLibraryPKModels}
\title{Get the library of PK models.}
\usage{
getLibraryPKModels(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{LibraryOfModels}.}
}
\value{
A list giving the PK models.
}
\description{
Get the library of PK models.
}
