% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getListLastName}
\alias{getListLastName}
\title{getListLastName: Get names of the deepest elements in a nested list}
\usage{
getListLastName(list)
}
\arguments{
\item{list}{A \code{list} (potentially nested).}
}
\value{
A \code{character} vector containing the names of the last elements.
}
\description{
Recursively traverses a nested list to extract the names of the elements
at the lowest level of the hierarchy.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
