% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getMfisher}
\alias{getMfisher}
\title{Get the Fisher Information Matrix.}
\usage{
getMfisher(object)
}
\arguments{
\item{object}{A \code{Fim} object.}
}
\value{
A matrix of numeric \code{mfisher} giving the Fisher Information Matrix.
}
\description{
Get the Fisher Information Matrix.
}
