% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getModelEquations}
\alias{getModelEquations}
\alias{getModelEquations,PFIMProject-method}
\title{Get the model equations.}
\usage{
getModelEquations(object)

\S4method{getModelEquations}{PFIMProject}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{PFIMProject}.}
}
\value{
A list giving the model equations.
}
\description{
Get the model equations.
}
