% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response.R
\name{getModelError}
\alias{getModelError}
\title{Get the model error.}
\usage{
getModelError(object)
}
\arguments{
\item{object}{A \code{Response} object.}
}
\value{
The object
}
\description{
Get the model error.
}
