% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getModelErrorData}
\alias{getModelErrorData}
\title{get the parameters sigma slope and sigma inter (used for the report).}
\arguments{
\item{modelError}{An object \code{ModelError} that defines the model error.}
}
\value{
A list of dataframe with outcome, type of model error and sigma slope and inter.
}
\description{
get the parameters sigma slope and sigma inter (used for the report).
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
