% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getModelFromLibrary}
\alias{getModelFromLibrary}
\title{Get a model from the library of models.}
\usage{
getModelFromLibrary(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
Return a model from the the library of models.
}
\description{
Get a model from the library of models.
}
