% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getModelName}
\alias{getModelName}
\title{Get the name of the \code{Model} object.}
\usage{
getModelName(object)
}
\arguments{
\item{object}{A \code{Model} object.}
}
\value{
A character string \code{modelName} giving the name of the \code{Model} object.
}
\description{
Get the name of the \code{Model} object.
}
