% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{getModelNameList}
\alias{getModelNameList}
\title{Get the list of all the models in the \code{LibraryOfModels} object.}
\usage{
getModelNameList(object)
}
\arguments{
\item{object}{A \code{LibraryOfModels} object.}
}
\value{
The list \code{ModelNameList} of the names of the PK, PD and PKPD models in the \code{LibraryOfModels} object.
}
\description{
Get the list of all the models in the \code{LibraryOfModels} object.
}
