% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{getModelParameters}
\alias{getModelParameters}
\title{Get the model parameters of a statistical model.}
\usage{
getModelParameters(object)
}
\arguments{
\item{object}{\code{getModelParameters} object.}
}
\value{
A list giving the model parameters.
}
\description{
Get the model parameters of a statistical model.
}
