% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getModelParametersValues}
\alias{getModelParametersValues}
\alias{getModelParametersValues,Model-method}
\title{Get the values of the model parameters.}
\usage{
getModelParametersValues(object)

\S4method{getModelParametersValues}{Model}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
A list giving the values of the model parameters.
}
\description{
Get the values of the model parameters.
}
