% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getMu}
\alias{getMu}
\title{Get the fixed effect of an object.}
\usage{
getMu(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
The object with the updated fixed effect.
}
\description{
Get the fixed effect of an object.
}
