% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getName}
\alias{getName}
\title{Get the name of an object}
\usage{
getName(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
A character string \code{name} giving the name of the object.
}
\description{
Get the name of an object
}
