% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{getNameAdministration}
\alias{getNameAdministration}
\title{Get the name of the outcome of an object \code{Administration}.}
\usage{
getNameAdministration(object)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}
}
\value{
A character string giving the name for the response of the object \code{Administration}.
}
\description{
Get the name of the outcome of an object \code{Administration}.
}
