% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getNameArm}
\alias{getNameArm}
\title{Get the name of the arm.}
\usage{
getNameArm(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
A character string \code{name} giving the name of the object \code{Arm}.
}
\description{
Get the name of the arm.
}
