% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getNameDesign}
\alias{getNameDesign}
\title{Get the name of the design.}
\usage{
getNameDesign(object)
}
\arguments{
\item{object}{\code{Design} object.}
}
\value{
A character string \code{name} giving the name of design.
}
\description{
Get the name of the design.
}
