% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\name{getNameDesignConstraint}
\alias{getNameDesignConstraint}
\title{Get the name of the  \code{DesignConstraint} object.}
\usage{
getNameDesignConstraint(object)
}
\arguments{
\item{object}{A \code{DesignConstraint} object.}
}
\value{
A character string \code{name} giving the name of \code{DesignConstraint} object.
}
\description{
Get the name of the  \code{DesignConstraint} object.
}
