% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{getNameModelParameter}
\alias{getNameModelParameter}
\title{Get the name of a \code{ModelParameter} object.}
\usage{
getNameModelParameter(object)
}
\arguments{
\item{object}{\code{ModelParameter} object.}
}
\value{
A character string \code{name} giving the name of a \code{ModelParameter} object.
}
\description{
Get the name of a \code{ModelParameter} object.
}
