% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelVariable.R
\name{getNameModelVariable}
\alias{getNameModelVariable}
\title{Get the name of the initial variable for an ODE model.}
\usage{
getNameModelVariable(object)
}
\arguments{
\item{object}{\code{ModelVariable} object.}
}
\value{
A character string \code{name} giving the name of the initial variable for an ODE model.
}
\description{
Get the name of the initial variable for an ODE model.
}
