% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getNamePFIMProject}
\alias{getNamePFIMProject}
\title{Get the name of a PFIMProject project.}
\usage{
getNamePFIMProject(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
The character string \code{name} giving the name of a PFIMProject project.
}
\description{
Get the name of a PFIMProject project.
}
