% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response.R
\name{getNameResponse}
\alias{getNameResponse}
\title{Get the name of the response of the model.}
\usage{
getNameResponse(object)
}
\arguments{
\item{object}{A \code{Response} object.}
}
\value{
A character string \code{name} giving the name of the response of the model.
}
\description{
Get the name of the response of the model.
}
