% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getNames}
\alias{getNames}
\title{Get the names of an object.}
\usage{
getNames(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
A vector giving the names of the object.
}
\description{
Get the names of an object.
}
