% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\name{getNameSampleTime}
\alias{getNameSampleTime}
\title{Get the name of the response of the \code{SamplingTimes} object.}
\usage{
getNameSampleTime(object)
}
\arguments{
\item{object}{A \code{SamplingTimes} object.}
}
\value{
A character string \code{outcome} giving the name of the response of the model.
}
\description{
Get the name of the response of the \code{SamplingTimes} object.
}
