% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getNumberOfArms}
\alias{getNumberOfArms}
\title{Get the number of arms in a design.}
\usage{
getNumberOfArms(object)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}
}
\value{
A numeric \code{numberOfArms} giving the number of arms in the design.
}
\description{
Get the number of arms in a design.
}
