% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraint.R
\name{getNumberOfDoses}
\alias{getNumberOfDoses}
\title{Get the vector \code{AllowedDoses} of allowed amount of dose.}
\usage{
getNumberOfDoses(object)
}
\arguments{
\item{object}{An object \code{AdministrationConstraint} from the class \linkS4class{AdministrationConstraint}.}
}
\value{
The numeric \code{AllowedDoses} giving the number of allowed amount of doses in the object \code{AdministrationConstraint}.
}
\description{
Get the vector \code{AllowedDoses} of allowed amount of dose.
}
