% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{getNumberOfIterations}
\alias{getNumberOfIterations}
\alias{getNumberOfIterations,MultiplicativeAlgorithm-method}
\title{Get the number of iterations.}
\usage{
getNumberOfIterations(object)

\S4method{getNumberOfIterations}{MultiplicativeAlgorithm}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{MultiplicativeAlgorithm}.}
}
\value{
A numeric giving the number of iterations.
}
\description{
Get the number of iterations.
}
