% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getNumberOfParameter}
\alias{getNumberOfParameter}
\title{Get the number of parameters of a \code{ModelError} object.}
\usage{
getNumberOfParameter(object)
}
\arguments{
\item{object}{A \code{ModelError} object.}
}
\value{
A numeric giving the number of parameters.
}
\description{
Get the number of parameters of a \code{ModelError} object.
}
