% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEquations.R
\name{getNumberOfParameters}
\alias{getNumberOfParameters}
\title{Get the number of parameters of a \code{ModelEquations} object.}
\usage{
getNumberOfParameters(object)
}
\arguments{
\item{object}{A \code{ModelEquations} object.}
}
\value{
A numeric \code{allParameters} giving the number of parameters.
}
\description{
Get the number of parameters of a \code{ModelEquations} object.
}
