% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getNumberOfSamplings}
\alias{getNumberOfSamplings}
\title{Get the number of sampling times in a arm.}
\usage{
getNumberOfSamplings(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
A numeric giving the number of sampling times in the \code{Arm} object.
}
\description{
Get the number of sampling times in a arm.
}
