% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{getNumberOfsamplingsOptimisable}
\alias{getNumberOfsamplingsOptimisable}
\alias{getNumberOfsamplingsOptimisable,SamplingTimeConstraints-method}
\title{Get the number of sampling times that are optimisable.}
\usage{
getNumberOfsamplingsOptimisable(object)

\S4method{getNumberOfsamplingsOptimisable}{SamplingTimeConstraints}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{SamplingTimeConstraints}.}
}
\value{
A vector giving the number of sampling times that are optimisable.
}
\description{
Get the number of sampling times that are optimisable.
}
