% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\name{getnumberOfSamplingTimes}
\alias{getnumberOfSamplingTimes}
\title{Get the number of sampling times.}
\usage{
getnumberOfSamplingTimes(object)
}
\arguments{
\item{object}{A \code{SamplingConstraint} object.}
}
\value{
A numeric giving the number of sampling times in the \code{SamplingConstraint} object.
}
\description{
Get the number of sampling times.
}
