% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{getNumberOfTimesByWindows}
\alias{getNumberOfTimesByWindows}
\title{Get the number of sampling times by windows.}
\usage{
getNumberOfTimesByWindows(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{SamplingTimeConstraints}.}
}
\value{
A vector giving the number of sampling times by windows.
}
\description{
Get the number of sampling times by windows.
}
