% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getNumberSamples}
\alias{getNumberSamples}
\title{Get the number of sampled in a Design.}
\usage{
getNumberSamples(object)
}
\arguments{
\item{object}{\code{Design} object.}
}
\value{
A numeric \code{number_samples} giving the number of sample of the design.
}
\description{
Get the number of sampled in a Design.
}
