% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\name{getNumberTime}
\alias{getNumberTime}
\title{Get the number of times in a \code{SamplingTimes} object.}
\usage{
getNumberTime(object)
}
\arguments{
\item{object}{A \code{SamplingTimes} object.}
}
\value{
A numeric  giving the number of times.
}
\description{
Get the number of times in a \code{SamplingTimes} object.
}
