% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getOdeSolverParameters}
\alias{getOdeSolverParameters}
\title{Get the parameters for the ode solvers of an object.}
\usage{
getOdeSolverParameters(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
The list giving the parameters for the ode solvers.
}
\description{
Get the parameters for the ode solvers of an object.
}
