% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{getOmega}
\alias{getOmega}
\title{Get \code{Omega} of a \code{ModelParameter} object.}
\usage{
getOmega(object)
}
\arguments{
\item{object}{\code{ModelParameter} object.}
}
\value{
A numeric \code{omega} giving the variance \code{Omega} of a \code{ModelParameter} object.
}
\description{
Get \code{Omega} of a \code{ModelParameter} object.
}
