% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{getOptimalDesign}
\alias{getOptimalDesign}
\title{Get the optimal design.}
\usage{
getOptimalDesign(object)
}
\arguments{
\item{object}{An \code{Design} object.}
}
\value{
An \code{Design} object giving the optimal design.
}
\description{
Get the optimal design.
}
