% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{getOptimalFrequencies}
\alias{getOptimalFrequencies}
\alias{getOptimalFrequencies,FedorovWynnAlgorithm-method}
\title{Get the optimal frequencies}
\usage{
getOptimalFrequencies(object)

\S4method{getOptimalFrequencies}{FedorovWynnAlgorithm}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{FedorovWynnAlgorithm}.}
}
\value{
A vector giving the optimal frequencies
}
\description{
Get the optimal frequencies
}
