% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{getOptimalWeights}
\alias{getOptimalWeights}
\alias{getOptimalWeights,MultiplicativeAlgorithm-method}
\title{Get the optimal weights.}
\usage{
getOptimalWeights(object)

\S4method{getOptimalWeights}{MultiplicativeAlgorithm}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{MultiplicativeAlgorithm}.}
}
\value{
A vector giving the optimal weights.
}
\description{
Get the optimal weights.
}
