% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\name{getOptimisability}
\alias{getOptimisability}
\title{Get the optimisability of a \code{SamplingConstraint} object.}
\usage{
getOptimisability(object)
}
\arguments{
\item{object}{\code{SamplingConstraint} object.}
}
\value{
A boolean giving the optimisability of a \code{SamplingConstraint} object.
}
\description{
Get the optimisability of a \code{SamplingConstraint} object.
}
