% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getOptimizationResult}
\alias{getOptimizationResult}
\title{Get the results of the optimization process.}
\usage{
getOptimizationResult(object)
}
\arguments{
\item{object}{\code{Design} object.}
}
\value{
A \code{Optimization} object giving the results of the optimization process.
}
\description{
Get the results of the optimization process.
}
