% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{getOptimizationResults}
\alias{getOptimizationResults}
\alias{getOptimizationResults,Optimization-method}
\title{Get the optimization results.}
\usage{
getOptimizationResults(object)

\S4method{getOptimizationResults}{Optimization}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Optimization}.}
}
\value{
An object from the class \linkS4class{OptimizationAlgorithm} giving the optimization results.
}
\description{
Get the optimization results.
}
