% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Administration.R,
%   R/AdministrationConstraints.R, R/ModelError.R, R/SamplingTimeConstraints.R,
%   R/SamplingTimes.R
\name{getOutcome}
\alias{getOutcome}
\alias{getOutcome,Administration-method}
\alias{getOutcome,AdministrationConstraints-method}
\alias{getOutcome,ModelError-method}
\alias{getOutcome,SamplingTimeConstraints-method}
\alias{getOutcome,SamplingTimes-method}
\title{getOutcome}
\usage{
getOutcome(object)

\S4method{getOutcome}{Administration}(object)

\S4method{getOutcome}{AdministrationConstraints}(object)

\S4method{getOutcome}{ModelError}(object)

\S4method{getOutcome}{SamplingTimeConstraints}(object)

\S4method{getOutcome}{SamplingTimes}(object)
}
\arguments{
\item{object}{An object defined from a class of PFIM.}
}
\value{
A string giving the outcome of the object.
}
\description{
Get the outcome of an object.
}
