% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getOutcomes}
\alias{getOutcomes}
\title{Get the outcomes of a model.}
\usage{
getOutcomes(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
A list giving the outcomes of the model.
}
\description{
Get the outcomes of a model.
}
