% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getOutcomesEvaluation}
\alias{getOutcomesEvaluation}
\alias{getOutcomesEvaluation,Design-method}
\title{getOutcomesEvaluation}
\usage{
getOutcomesEvaluation(object)

\S4method{getOutcomesEvaluation}{Design}(object)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}
}
\value{
The list \code{outcomesEvaluation} containing the results of the design evaluation for the outcomes.
}
\description{
Get the results of the evaluation of the outcomes.
}
