% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getOutcomesForEvaluation}
\alias{getOutcomesForEvaluation}
\alias{getOutcomesForEvaluation,Model-method}
\title{Get the outcomes of a model used for the evaluation (is scales outcomes).}
\usage{
getOutcomesForEvaluation(object)

\S4method{getOutcomesForEvaluation}{Model}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
A list giving the outcomes of a model used for the evaluation (is scales outcomes).
}
\description{
Get the outcomes of a model used for the evaluation (is scales outcomes).
}
