% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEquations.R
\name{getParameters}
\alias{getParameters}
\title{Get the parameters of a \code{ModelEquations} object.}
\usage{
getParameters(object)
}
\arguments{
\item{object}{A \code{ModelEquations} object.}
}
\value{
A vector \code{allParameters} giving the parameters of the model.
}
\description{
Get the parameters of a \code{ModelEquations} object.
}
