% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{getParametersOdeSolver}
\alias{getParametersOdeSolver}
\title{Get parameters for the ode solver}
\usage{
getParametersOdeSolver(object)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}
}
\value{
The parameters for the ode solver.
}
\description{
Get parameters for the ode solver
}
