% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKPDModel.R
\name{getPDModel}
\alias{getPDModel}
\title{Get a PD model from a \code{PKPDModel} object.}
\usage{
getPDModel(object)
}
\arguments{
\item{object}{\code{PKPDModel} object.}
}
\value{
A \code{Model} object giving the \code{pdModel} from the PKPD model.
}
\description{
Get a PD model from a \code{PKPDModel} object.
}
