% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKPDModel.R
\name{getPKModel}
\alias{getPKModel}
\title{Get a PK model from a \code{PKPDModel} object.}
\usage{
getPKModel(object)
}
\arguments{
\item{object}{\code{PKPDModel} object.}
}
\value{
A \code{Model} object giving the \code{pkModel} from the PKPD model.
}
\description{
Get a PK model from a \code{PKPDModel} object.
}
