% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{getPKPDModel}
\alias{getPKPDModel}
\title{Get a PKPD model of the \code{LibraryOfModels} object.}
\usage{
getPKPDModel(object, namePKModel, namePDModel)
}
\arguments{
\item{object}{A \code{LibraryOfModels} object.}

\item{namePKModel}{A character string giving the name of the PK model.}

\item{namePDModel}{A character string giving the name of the PD model.}
}
\value{
Return a \code{Model} giving the PKPD model consisting of the PK and PD models named namePKModel and namePDModel respectively.
}
\description{
Get a PKPD model of the \code{LibraryOfModels} object.
}
