% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getPlotOptions}
\alias{getPlotOptions}
\title{Get the plot options for graphs responses and SI}
\usage{
getPlotOptions(plotOptions, outcomesNames)
}
\arguments{
\item{plotOptions}{A list giving the plots options.}

\item{outcomesNames}{A list giving the output names.}
}
\value{
The list containing the plot options.
}
\description{
Get the plot options for graphs responses and SI
}
