% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuousConstraint.R
\name{getRange}
\alias{getRange}
\title{Get the range of a \code{ContinuousConstraint} object.}
\usage{
getRange(object)
}
\arguments{
\item{object}{A\code{ContinuousConstraint} object.}
}
\value{
A numeric \code{range} giving the range of a \code{ContinuousConstraint} object.
}
\description{
Get the range of a \code{ContinuousConstraint} object.
}
