% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEquations.R
\name{getResponseIndice}
\alias{getResponseIndice}
\title{Get the index of the response of a \code{ModelEquations} object.}
\usage{
getResponseIndice(object, equationName)
}
\arguments{
\item{object}{A \code{ModelEquations} object.}

\item{equationName}{A character string giving the name of the response of the equations.}
}
\value{
A numeric giving the index of the equation in the model.
}
\description{
Get the index of the response of a \code{ModelEquations} object.
}
