% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraint.R,
%   R/SamplingConstraint.R
\name{getResponseName}
\alias{getResponseName}
\alias{getResponseName,SamplingConstraint-method}
\title{Get the name of the response for the administration constraints.}
\usage{
getResponseName(object)

\S4method{getResponseName}{SamplingConstraint}(object)
}
\arguments{
\item{object}{\code{SamplingConstraint} object.}
}
\value{
The character string \code{response} giving the name of the response of the object \code{AdministrationConstraint} object.

The character string \code{response} giving the name of the response for the \code{SamplingConstraint} object.
}
\description{
Get the name of the response for the administration constraints.

Get the name of the response for the SamplingConstraint.
}
