% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getResponseNameByIndice}
\alias{getResponseNameByIndice}
\title{Get the response name given the indice of the response.}
\usage{
getResponseNameByIndice(object, outcomeIndice)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{outcomeIndice}{A numeric giving the indice of the response in the \code{Arm} object.}
}
\value{
A character string giving the name of the response.
}
\description{
Get the response name given the indice of the response.
}
