% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{getResponsesStatisticalModel}
\alias{getResponsesStatisticalModel}
\title{Get the responses of a statistical model.}
\usage{
getResponsesStatisticalModel(object)
}
\arguments{
\item{object}{A \code{getResponsesStatisticalModel} object.}
}
\value{
A list giving the responses of a statistical model.
}
\description{
Get the responses of a statistical model.
}
