% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getRSE}
\alias{getRSE}
\title{getRSE: return the RSE data frame for the optimal design}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The RSE
}
\description{
getRSE: return the RSE data frame for the optimal design

getRSE: get the RSE
}
