% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\name{getSampleTime}
\alias{getSampleTime}
\title{Get the sample time of the response of the \code{SamplingTimes} object.}
\usage{
getSampleTime(object)
}
\arguments{
\item{object}{A \code{getSampleTime} object.}
}
\value{
A vector \code{sample_time} giving the sample time.
}
\description{
Get the sample time of the response of the \code{SamplingTimes} object.
}
