% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\name{getSamplingConstraints}
\alias{getSamplingConstraints}
\title{Get the constraints on the sampling for a \code{DesignConstraint} object.}
\usage{
getSamplingConstraints(object, responseName)
}
\arguments{
\item{object}{\code{DesignConstraint} object.}

\item{responseName}{A character string giving the name of the response.}
}
\value{
The lists of constraints \code{samplingConstraint} for a \code{DesignConstraint} object.
}
\description{
Get the constraints on the sampling for a \code{DesignConstraint} object.
}
