% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getSamplingConstraintsInArm}
\alias{getSamplingConstraintsInArm}
\title{Get the sampling constraints of an arm.}
\usage{
getSamplingConstraintsInArm(object, responseName)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{responseName}{A character string giving the name of the response.}
}
\value{
An object \code{constraintsOfTheResponse} from the class \code{SamplingConstraint}.
giving the sampling constraints of the object \code{Arm}.
}
\description{
Get the sampling constraints of an arm.
}
