% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getSamplingData}
\alias{getSamplingData}
\title{Extract sampling times and maximum sampling time}
\usage{
getSamplingData(arm, ...)
}
\arguments{
\item{arm}{An object of class \code{Arm}.}

\item{...}{Additional arguments.}
}
\value{
A list with \code{samplingTimes}, \code{samplings} (named list of
  numeric vectors), and \code{samplingMax} (numeric scalar).
}
\description{
Returns structured sampling information from an arm, used internally
by plotting methods.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
