% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getSamplings}
\alias{getSamplings}
\title{Get the sampling of an object.}
\usage{
getSamplings(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
A list of the samplings of the object.
}
\description{
Get the sampling of an object.
}
