% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{getSamplingsWindows}
\alias{getSamplingsWindows}
\title{Get the windows for the sampling times.}
\usage{
getSamplingsWindows(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{SamplingTimeConstraints}.}
}
\value{
A list giving the vector of the windows for the sampling times.
}
\description{
Get the windows for the sampling times.
}
