% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getSamplingTimes}
\alias{getSamplingTimes}
\alias{getSamplingTimes,Arm-method}
\title{getSamplingTimes}
\usage{
getSamplingTimes(object)

\S4method{getSamplingTimes}{Arm}(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
The list \code{samplingTimes} for the object \code{Arm}.
}
\description{
Get the vectors of sampling times for an arm.
}
