% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getSamplingTimesConstraints}
\alias{getSamplingTimesConstraints}
\title{Get the sampling times constraints.}
\usage{
getSamplingTimesConstraints(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
The list \code{getSamplingTimesConstraints}.
}
\description{
Get the sampling times constraints.
}
