% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getSE}
\alias{getSE}
\title{getSE: return the SE data frame for the optimal design}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The SE.
}
\description{
getSE: return the SE data frame for the optimal design

getSE: get the SE
}
