% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getShrinkage}
\alias{getShrinkage}
\title{getShrinkage: return the shrinkage of the FIM}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The shrinkage
}
\description{
getShrinkage: return the shrinkage of the FIM

getShrinkage: get the shrinkage
}
