% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getSig}
\alias{getSig}
\title{Get the values for Sigma derivatives \code{DVSigma} for the \code{ModelError} object.}
\usage{
getSig(object, f_x_i_theta)
}
\arguments{
\item{object}{\code{ModelError} object.}

\item{f_x_i_theta}{the nonlinear structural model \code{f_x_i_theta}}
}
\value{
A list indexed with the parameters giving the values for the derivatives with respect to each parameters of the model error.
}
\description{
Get the values for Sigma derivatives \code{DVSigma} for the \code{ModelError} object.
}
