% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getSigmaInter}
\alias{getSigmaInter}
\title{Get the\code{sigma_inter} of a \code{ModelError} object.}
\usage{
getSigmaInter(object)
}
\arguments{
\item{object}{A \code{ModelError} object.}
}
\value{
A numeric \code{sigma_inter} giving the \code{sigma_inter}.
}
\description{
Get the\code{sigma_inter} of a \code{ModelError} object.
}
