% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R, R/Combined1.R, R/Combined1c.R,
%   R/Combined2c.R, R/Combined2.R, R/Constant.R, R/Response.R
\name{getSigmaNames}
\alias{getSigmaNames}
\alias{getSigmaNames,Combined1-method}
\alias{getSigmaNames,Combined1c-method}
\alias{getSigmaNames,Combined2c-method}
\alias{getSigmaNames,Combined2-method}
\alias{getSigmaNames,Constant-method}
\alias{getSigmaNames,Response-method}
\title{Get the names for the error sigma inter.}
\usage{
getSigmaNames(object)

\S4method{getSigmaNames}{Combined1}(object)

\S4method{getSigmaNames}{Combined1c}(object)

\S4method{getSigmaNames}{Combined2c}(object)

\S4method{getSigmaNames}{Combined2}(object)

\S4method{getSigmaNames}{Constant}(object)

\S4method{getSigmaNames}{Response}(object)
}
\arguments{
\item{object}{A \code{Response} object.}
}
\value{
A character giving the names for the sigma inter of the error model.

The character string \code{sigmaNames} giving the names of the variances.

The character string \code{sigmaNames} giving the names of the variances.

The character string \code{sigmaNames} giving the names of the variances.

The character string \code{sigmaNames} giving the names of the variances.

The variances \code{sigma_{inter}} and \code{sigma_{slope}}.

A character string \code{sigmaNames} giving the names of the sigma.
}
\description{
Get the names for the error sigma inter.

Get the names of the variances.

Get the names of the variances.

Get the names of the variances.

Get the names of the variances.

Get the variances \code{sigma_{inter}} and \code{sigma_{slope}}.

Get the Sigma Names.
}
