% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getSigmaSlope}
\alias{getSigmaSlope}
\title{Get the \code{sigma_slope} of a \code{ModelError} object.}
\usage{
getSigmaSlope(object)
}
\arguments{
\item{object}{An \code{ModelError} object.}
}
\value{
The numeric \code{sigma_slope} giving the \code{sigma_slope}.
}
\description{
Get the \code{sigma_slope} of a \code{ModelError} object.
}
