% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R, R/Combined1.R, R/Combined1c.R,
%   R/Combined2c.R, R/Combined2.R, R/Constant.R
\name{getSigmaValues}
\alias{getSigmaValues}
\alias{getSigmaValues,Combined1-method}
\alias{getSigmaValues,Combined1c-method}
\alias{getSigmaValues,Combined2c-method}
\alias{getSigmaValues,Combined2-method}
\alias{getSigmaValues,Constant-method}
\title{Get the values of the variances \code{sigma_inter} and \code{sigma_slope}.}
\usage{
getSigmaValues(object)

\S4method{getSigmaValues}{Combined1}(object)

\S4method{getSigmaValues}{Combined1c}(object)

\S4method{getSigmaValues}{Combined2c}(object)

\S4method{getSigmaValues}{Combined2}(object)

\S4method{getSigmaValues}{Constant}(object)
}
\arguments{
\item{object}{A \code{Combined1} object.}
}
\value{
A numeric giving the values for the sigma inter of the error model.

A numeric vector giving the values of the variances \code{sigma_inter} and \code{sigma_slope}.

A numeric vector giving the values of the variances \code{sigma_inter} and \code{sigma_slope}.

A numeric vector giving the values of the variances \code{sigma_inter} and \code{sigma_slope}.

A numeric vector giving the values of the variances \code{sigma_inter} and \code{sigma_slope}.

A vector giving the values of the variances \code{sigma_{inter}} and \code{sigma_{slope}}.
}
\description{
Get the values of the variances \code{sigma_inter} and \code{sigma_slope}.

Get the values of the variances \code{sigma_inter} and \code{sigma_slope}.

Get the values of the variances \code{sigma_inter} and \code{sigma_slope}.

Get the values of the variances \code{sigma_inter} and \code{sigma_slope}.

Get the values of the variances \code{sigma_inter} and \code{sigma_slope}.

Get the values of the variances \code{sigma_{inter}} and \code{sigma_{slope}}.
}
