% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getSize}
\alias{getSize}
\title{Get the size of an object.}
\usage{
getSize(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
A numeric giving the size of the object.
}
\description{
Get the size of an object.
}
