% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getStatisticalModel}
\alias{getStatisticalModel}
\title{Get the \code{StatisticalModel} object of the PFIMProject object.}
\usage{
getStatisticalModel(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
Return the object \code{statistical_model} of the \linkS4class{StatisticalModel} of the PFIMProject object.
}
\description{
Get the \code{StatisticalModel} object of the PFIMProject object.
}
