% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/IndividualFim.R, R/PopulationFim.R
\name{getStatisticalModelStandardErrors}
\alias{getStatisticalModelStandardErrors}
\alias{getStatisticalModelStandardErrors,IndividualFim-method}
\alias{getStatisticalModelStandardErrors,PopulationFim-method}
\title{Get the SE of \code{IndividualFim} object.}
\usage{
getStatisticalModelStandardErrors(object, modelParameters)

\S4method{getStatisticalModelStandardErrors}{IndividualFim}(object, modelParameters)

\S4method{getStatisticalModelStandardErrors}{PopulationFim}(object, modelParameters)
}
\arguments{
\item{object}{A \code{Fim} object.}

\item{modelParameters}{A character string giving the model parameters.}
}
\value{
A data frame giving the expected standard error.

A list giving the fixed effects of \code{IndividualFim} object.

A data frame giving the expected standard error.
}
\description{
Get the SE of \code{IndividualFim} object.

Compute expected standard error data frame.
}
