% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{getTau}
\alias{getTau}
\alias{getTau,Administration-method}
\title{getTau}
\usage{
getTau(object)

\S4method{getTau}{Administration}(object)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}
}
\value{
The numeric \code{tau} giving the frequency \code{tau}.
}
\description{
Get the frequency \code{tau}.
}
