% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\name{getTotalNumberOfIndividuals}
\alias{getTotalNumberOfIndividuals}
\title{get the total number of individuals in a \code{DesignConstraint} object.}
\usage{
getTotalNumberOfIndividuals(object)
}
\arguments{
\item{object}{\code{DesignConstraint} object.}
}
\value{
The \code{DesignConstraint} object with the total number of individual.
}
\description{
get the total number of individuals in a \code{DesignConstraint} object.
}
