% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getTotalSize}
\alias{getTotalSize}
\title{Get the total size of a design.}
\usage{
getTotalSize(object)
}
\arguments{
\item{object}{\code{Design} object.}
}
\value{
A numeric \code{total_size} giving the size of a design.
}
\description{
Get the total size of a design.
}
