% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getVarianceEffects}
\alias{getVarianceEffects}
\title{Get the matrix of the variance effects.}
\usage{
getVarianceEffects(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
The matrix of the variance effects.
}
\description{
Get the matrix of the variance effects.
}
