% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{getWeightFrame}
\alias{getWeightFrame}
\title{Get the frame with weight vector after optimisation.}
\usage{
getWeightFrame(object)
}
\arguments{
\item{object}{A \code{MultiplicativeAlgorithm} object.}
}
\value{
The data frame \code{armFrame} with weight vector after optimisation.
}
\description{
Get the frame with weight vector after optimisation.
}
