% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getWeights}
\alias{getWeights}
\title{Get the weights for the optimal designs.}
\usage{
getWeights(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
A data frame \code{weights} giving the weights of the optimal designs.
}
\description{
Get the weights for the optimal designs.
}
