% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{getWeightThreshold}
\alias{getWeightThreshold}
\alias{getWeightThreshold,MultiplicativeAlgorithm-method}
\title{Get the parameter weightThreshold}
\usage{
getWeightThreshold(object)

\S4method{getWeightThreshold}{MultiplicativeAlgorithm}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{MultiplicativeAlgorithm}.}
}
\value{
A numeric giving the WeightThreshold.
}
\description{
Get the parameter weightThreshold
}
