% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualFim.R
\docType{class}
\name{IndividualFim-class}
\alias{IndividualFim-class}
\alias{IndividualFim}
\title{Class "Fim"}
\description{
A class storing information regarding the individual Fisher matrix.
The class \code{IndividualFim} inherits from the class \code{Fim}.
}
