% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{initialize,Administration-method}
\alias{initialize,Administration-method}
\title{initialize}
\usage{
\S4method{initialize}{Administration}(.Object, outcome, timeDose, dose, Tinf, tau)
}
\arguments{
\item{.Object}{.Object}

\item{outcome}{outcome}

\item{timeDose}{timeDose}

\item{dose}{dose}

\item{Tinf}{Tinf}

\item{tau}{tau}
}
\value{
Administration
}
\description{
initialize
}
