% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraints.R
\name{initialize,AdministrationConstraints-method}
\alias{initialize,AdministrationConstraints-method}
\title{initialize}
\usage{
\S4method{initialize}{AdministrationConstraints}(.Object, outcome, doses)
}
\arguments{
\item{.Object}{.Object}

\item{outcome}{outcome}

\item{doses}{doses}
}
\description{
initialize
}
