% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{initialize,Arm-method}
\alias{initialize,Arm-method}
\title{initialize}
\usage{
\S4method{initialize}{Arm}(
  .Object,
  name,
  size,
  administrations,
  initialConditions,
  samplingTimes,
  administrationsConstraints,
  samplingTimesConstraints,
  dataForArmEvaluation
)
}
\arguments{
\item{.Object}{.Object}

\item{name}{name}

\item{size}{size}

\item{administrations}{administrations}

\item{initialConditions}{initialConditions}

\item{samplingTimes}{samplingTimes}

\item{administrationsConstraints}{administrationsConstraints}

\item{samplingTimesConstraints}{samplingTimesConstraints}

\item{dataForArmEvaluation}{dataForArmEvaluation}
}
\value{
Arm
}
\description{
initialize
}
