% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{initialize,Design-method}
\alias{initialize,Design-method}
\title{initialize}
\usage{
\S4method{initialize}{Design}(
  .Object,
  name,
  size,
  arms,
  outcomesEvaluation,
  outcomesGradient,
  numberOfArms,
  fim
)
}
\arguments{
\item{.Object}{.Object}

\item{name}{name}

\item{size}{size}

\item{arms}{arms}

\item{outcomesEvaluation}{outcomesEvaluation}

\item{outcomesGradient}{outcomesGradient}

\item{numberOfArms}{numberOfArms}

\item{fim}{fim}
}
\value{
Design
}
\description{
initialize
}
