% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{initialize,Distribution-method}
\alias{initialize,Distribution-method}
\title{initialize}
\usage{
\S4method{initialize}{Distribution}(.Object, parameters)
}
\arguments{
\item{.Object}{.Object}

\item{parameters}{parameters}
}
\value{
Distribution
}
\description{
initialize
}
