% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{initialize,Evaluation-method}
\alias{initialize,Evaluation-method}
\title{initialize}
\usage{
\S4method{initialize}{Evaluation}(
  .Object,
  name,
  model,
  modelEquations,
  modelParameters,
  modelError,
  outcomes,
  designs,
  fim,
  odeSolverParameters
)
}
\arguments{
\item{.Object}{.Object}

\item{name}{name}

\item{model}{model}

\item{modelEquations}{modelEquations}

\item{modelParameters}{modelParameters}

\item{modelError}{modelError}

\item{outcomes}{outcomes}

\item{designs}{designs}

\item{fim}{fim}

\item{odeSolverParameters}{odeSolverParameters}
}
\value{
Evaluation
}
\description{
initialize
}
