% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{initialize,FedorovWynnAlgorithm-method}
\alias{initialize,FedorovWynnAlgorithm-method}
\title{initialize}
\usage{
\S4method{initialize}{FedorovWynnAlgorithm}(
  .Object,
  elementaryProtocols,
  numberOfSubjects,
  proportionsOfSubjects,
  showProcess
)
}
\arguments{
\item{.Object}{.Object}

\item{elementaryProtocols}{elementaryProtocols}

\item{numberOfSubjects}{numberOfSubjects}

\item{proportionsOfSubjects}{proportionsOfSubjects}

\item{showProcess}{showProcess}
}
\value{
FedorovWynnAlgorithm
}
\description{
initialize
}
