% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{initialize,Fim-method}
\alias{initialize,Fim-method}
\title{initialize}
\usage{
\S4method{initialize}{Fim}(.Object, fisherMatrix, fixedEffects, varianceEffects, shrinkage)
}
\arguments{
\item{.Object}{.Object}

\item{fisherMatrix}{fisherMatrix}

\item{fixedEffects}{fixedEffects}

\item{varianceEffects}{varianceEffects}

\item{shrinkage}{shrinkage}
}
\value{
Fim
}
\description{
initialize
}
