% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{initialize,LibraryOfModels-method}
\alias{initialize,LibraryOfModels-method}
\title{initialize}
\usage{
\S4method{initialize}{LibraryOfModels}(.Object, name, content)
}
\arguments{
\item{.Object}{.Object}

\item{name}{fisherMatrix}

\item{content}{fixedEffects}
}
\value{
LibraryOfModels
}
\description{
initialize
}
