% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKPDModels.R
\name{initialize,LibraryOfPKPDModels-method}
\alias{initialize,LibraryOfPKPDModels-method}
\title{initialize}
\usage{
\S4method{initialize}{LibraryOfPKPDModels}(.Object)
}
\arguments{
\item{.Object}{.Object}
}
\value{
LibraryOfPKPDModels
}
\description{
initialize
}
