% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{initialize,LogNormal-method}
\alias{initialize,LogNormal-method}
\title{initialize}
\usage{
\S4method{initialize}{LogNormal}(.Object, ...)
}
\arguments{
\item{.Object}{.Object}

\item{...}{args}
}
\value{
LogNormal
}
\description{
initialize
}
