% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{initialize,Model-method}
\alias{initialize,Model-method}
\title{initialize}
\usage{
\S4method{initialize}{Model}(
  .Object,
  name,
  description,
  equations,
  outcomes,
  outcomesForEvaluation,
  parameters,
  modelError,
  initialConditions,
  odeSolverParameters,
  modelFromLibrary
)
}
\arguments{
\item{.Object}{.Object}

\item{name}{name}

\item{description}{description}

\item{equations}{equations}

\item{outcomes}{outcomes}

\item{outcomesForEvaluation}{outcomesForEvaluation}

\item{parameters}{parameters}

\item{modelError}{modelError}

\item{initialConditions}{initialConditions}

\item{odeSolverParameters}{odeSolverParameters}

\item{modelFromLibrary}{modelFromLibrary}
}
\value{
Model
}
\description{
initialize
}
