% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R
\name{initialize,ModelAnalytic-method}
\alias{initialize,ModelAnalytic-method}
\title{initialize}
\usage{
\S4method{initialize}{ModelAnalytic}(
  .Object,
  name,
  description,
  equations,
  outcomes,
  parameters,
  modelError
)
}
\arguments{
\item{.Object}{.Object}

\item{name}{name}

\item{description}{description}

\item{equations}{equations}

\item{outcomes}{outcomes}

\item{parameters}{parameters}

\item{modelError}{modelError}
}
\value{
ModelAnalytic
}
\description{
initialize
}
