% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfusion.R
\name{initialize,ModelInfusion-method}
\alias{initialize,ModelInfusion-method}
\title{initialize}
\usage{
\S4method{initialize}{ModelInfusion}(
  .Object,
  name,
  description,
  equations,
  outcomes,
  parameters,
  modelError,
  initialConditions,
  odeSolverParameters
)
}
\arguments{
\item{.Object}{.Object}

\item{name}{name}

\item{description}{description}

\item{equations}{equations}

\item{outcomes}{outcomes}

\item{parameters}{parameters}

\item{modelError}{modelError}

\item{initialConditions}{initialConditions}

\item{odeSolverParameters}{odeSolverParameters}
}
\value{
ModelInfusion
}
\description{
initialize
}
