% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{initialize,ModelParameter-method}
\alias{initialize,ModelParameter-method}
\title{initialize}
\usage{
\S4method{initialize}{ModelParameter}(.Object, name, distribution, fixedMu, fixedOmega)
}
\arguments{
\item{.Object}{.Object}

\item{name}{name}

\item{distribution}{distribution}

\item{fixedMu}{fixedMu}

\item{fixedOmega}{fixedOmega}
}
\value{
ModelParameter
}
\description{
initialize
}
