% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{initialize,MultiplicativeAlgorithm-method}
\alias{initialize,MultiplicativeAlgorithm-method}
\title{initialize}
\usage{
\S4method{initialize}{MultiplicativeAlgorithm}(
  .Object,
  arms,
  lambda,
  delta,
  numberOfIterations,
  weightThreshold,
  optimalWeights,
  optimalDesign,
  showProcess
)
}
\arguments{
\item{.Object}{.Object}

\item{arms}{arms}

\item{lambda}{lambda}

\item{delta}{delta}

\item{numberOfIterations}{numberOfIterations}

\item{weightThreshold}{weightThreshold}

\item{optimalWeights}{optimalWeights}

\item{optimalDesign}{optimalDesign}

\item{showProcess}{showProcess}
}
\value{
MultiplicativeAlgorithm
}
\description{
initialize
}
