% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{initialize,Normal-method}
\alias{initialize,Normal-method}
\title{initialize}
\usage{
\S4method{initialize}{Normal}(.Object, ...)
}
\arguments{
\item{.Object}{.Object}

\item{...}{args}
}
\value{
Normal
}
\description{
initialize
}
