% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{initialize,Optimization-method}
\alias{initialize,Optimization-method}
\title{initialize}
\usage{
\S4method{initialize}{Optimization}(
  .Object,
  name,
  model,
  modelEquations,
  modelParameters,
  modelError,
  optimizer,
  optimizerParameters,
  outcomes,
  designs,
  fim,
  odeSolverParameters,
  optimizationResults,
  evaluationFIMResults,
  evaluationInitialDesignResults
)
}
\arguments{
\item{.Object}{.Object}

\item{name}{name}

\item{model}{model}

\item{modelEquations}{modelEquations}

\item{modelParameters}{modelParameters}

\item{modelError}{modelError}

\item{optimizer}{optimizer}

\item{optimizerParameters}{optimizerParameters}

\item{outcomes}{outcomes}

\item{designs}{designs}

\item{fim}{fim}

\item{odeSolverParameters}{odeSolverParameters}

\item{optimizationResults}{optimizationResults}

\item{evaluationFIMResults}{evaluationFIMResults}

\item{evaluationInitialDesignResults}{evaluationInitialDesignResults}
}
\value{
Optimization
}
\description{
initialize
}
