% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationAlgorithm.R
\name{initialize,OptimizationAlgorithm-method}
\alias{initialize,OptimizationAlgorithm-method}
\title{initialize}
\usage{
\S4method{initialize}{OptimizationAlgorithm}(.Object, name, parameters)
}
\arguments{
\item{.Object}{.Object}

\item{name}{name}

\item{parameters}{parameters}
}
\value{
OptimizationAlgorithm
}
\description{
initialize
}
