% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{initialize,PFIMProject-method}
\alias{initialize,PFIMProject-method}
\title{initialize}
\usage{
\S4method{initialize}{PFIMProject}(.Object, name, description)
}
\arguments{
\item{.Object}{.Object}

\item{name}{name}

\item{description}{description}
}
\value{
PFIMProject
}
\description{
initialize
}
