% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGBOAlgorithm.R
\name{initialize,PGBOAlgorithm-method}
\alias{initialize,PGBOAlgorithm-method}
\title{initialize}
\usage{
\S4method{initialize}{PGBOAlgorithm}(
  .Object,
  N,
  muteEffect,
  maxIteration,
  purgeIteration,
  seed,
  showProcess,
  optimalDesign,
  iterationAndCriteria
)
}
\arguments{
\item{.Object}{.Object}

\item{N}{N}

\item{muteEffect}{muteEffect}

\item{maxIteration}{maxIteration}

\item{purgeIteration}{purgeIteration}

\item{seed}{seed}

\item{showProcess}{showProcess}

\item{optimalDesign}{optimalDesign}

\item{iterationAndCriteria}{iterationAndCriteria}
}
\value{
PGBOAlgorithm
}
\description{
initialize
}
