% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportional.R
\name{initialize,Proportional-method}
\alias{initialize,Proportional-method}
\title{initialize}
\usage{
\S4method{initialize}{Proportional}(
  .Object,
  outcome,
  equation,
  derivatives,
  sigmaInter,
  sigmaSlope,
  cError
)
}
\arguments{
\item{.Object}{.Object}

\item{outcome}{outcome}

\item{equation}{equation}

\item{derivatives}{derivatives}

\item{sigmaInter}{sigmaInter}

\item{sigmaSlope}{sigmaSlope}

\item{cError}{cError}
}
\value{
Proportional
}
\description{
initialize
}
