% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOAlgorithm.R
\name{initialize,PSOAlgorithm-method}
\alias{initialize,PSOAlgorithm-method}
\title{initialize}
\usage{
\S4method{initialize}{PSOAlgorithm}(
  .Object,
  maxIteration,
  populationSize,
  personalLearningCoefficient,
  globalLearningCoefficient,
  seed,
  showProcess,
  optimalDesign,
  iterationAndCriteria
)
}
\arguments{
\item{.Object}{.Object}

\item{maxIteration}{maxIteration}

\item{populationSize}{populationSize}

\item{personalLearningCoefficient}{personalLearningCoefficient}

\item{globalLearningCoefficient}{globalLearningCoefficient}

\item{seed}{seed}

\item{showProcess}{showProcess}

\item{optimalDesign}{optimalDesign}

\item{iterationAndCriteria}{iterationAndCriteria}
}
\value{
PSOAlgorithm
}
\description{
initialize
}
