% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{initialize,SamplingTimeConstraints-method}
\alias{initialize,SamplingTimeConstraints-method}
\title{initialize}
\usage{
\S4method{initialize}{SamplingTimeConstraints}(
  .Object,
  outcome,
  initialSamplings,
  fixedTimes,
  numberOfsamplingsOptimisable,
  samplingsWindows,
  numberOfTimesByWindows,
  minSampling
)
}
\arguments{
\item{.Object}{.Object}

\item{outcome}{outcome}

\item{initialSamplings}{initialSamplings}

\item{fixedTimes}{fixedTimes}

\item{numberOfsamplingsOptimisable}{numberOfsamplingsOptimisable}

\item{samplingsWindows}{samplingsWindows}

\item{numberOfTimesByWindows}{numberOfTimesByWindows}

\item{minSampling}{minSampling}
}
\value{
SamplingTimeConstraints
}
\description{
initialize
}
