% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\name{initialize,SamplingTimes-method}
\alias{initialize,SamplingTimes-method}
\title{initialize}
\usage{
\S4method{initialize}{SamplingTimes}(.Object, outcome, samplings)
}
\arguments{
\item{.Object}{.Object}

\item{outcome}{outcome}

\item{samplings}{samplings}
}
\value{
SamplingTimes
}
\description{
initialize
}
