% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexAlgorithm.R
\name{initialize,SimplexAlgorithm-method}
\alias{initialize,SimplexAlgorithm-method}
\title{initialize}
\usage{
\S4method{initialize}{SimplexAlgorithm}(
  .Object,
  pctInitialSimplexBuilding,
  maxIteration,
  tolerance,
  optimalDesigns,
  iterationAndCriteria,
  showProcess
)
}
\arguments{
\item{.Object}{.Object}

\item{pctInitialSimplexBuilding}{pctInitialSimplexBuilding}

\item{maxIteration}{maxIteration}

\item{tolerance}{tolerance}

\item{optimalDesigns}{optimalDesigns}

\item{iterationAndCriteria}{iterationAndCriteria}

\item{showProcess}{showProcess}
}
\value{
SimplexAlgorithm
}
\description{
initialize
}
