% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{is.multidose}
\alias{is.multidose}
\title{Test if an object \code{Administration}for a model is multi-doses or not.}
\usage{
is.multidose(object)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}
}
\value{
A boolean that gives TRUE if the administration is multi-doses, FALSE otherwise.
}
\description{
Test if an object \code{Administration}for a model is multi-doses or not.
}
