% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{isFixed}
\alias{isFixed}
\title{Boolean to set if a model parameters is fixed or not.}
\usage{
isFixed(object)
}
\arguments{
\item{object}{\code{ModelParameter} object.}
}
\value{
A boolean \code{fixed} giving TRUE if the model parameters is fixed, or FALSE is this parameters remain to be estimated.
}
\description{
Boolean to set if a model parameters is fixed or not.
}
