% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{isFixedMu}
\alias{isFixedMu}
\title{Boolean to set if \code{mu} is fixed or not.}
\usage{
isFixedMu(object)
}
\arguments{
\item{object}{\code{ModelParameter} object.}
}
\value{
A boolean \code{isFixedMu} giving TRUE if \code{mu} is fixed, FALSE otherwise.
}
\description{
Boolean to set if \code{mu} is fixed or not.
}
