% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\name{isLessThanDelay}
\alias{isLessThanDelay}
\title{Set the constraint on minimal time delay in sampling times.}
\usage{
isLessThanDelay(object, samplingTimes)
}
\arguments{
\item{object}{A \code{SamplingConstraint} object.}

\item{samplingTimes}{A \code{SamplingTimes} object.}
}
\value{
A boolean that give TRUE/FALSE if the constraint on minimal delay is satisfied.
}
\description{
Set the constraint on minimal time delay in sampling times.
}
