% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{isModelAnalytic}
\alias{isModelAnalytic}
\alias{isModelAnalytic,Model-method}
\title{Test if a mode is analytic.}
\usage{
isModelAnalytic(object)

\S4method{isModelAnalytic}{Model}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
Return a Boolean giving if the mode is analytic or not.
}
\description{
Test if a mode is analytic.
}
