% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{isModelODE}
\alias{isModelODE}
\title{Test if a mode is ode.}
\usage{
isModelODE(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
Return a Boolean giving if the mode is ode or not.
}
\description{
Test if a mode is ode.
}
