% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{isNotFixed}
\alias{isNotFixed}
\title{Boolean to set if a model parameters is not fixed or not.}
\usage{
isNotFixed(object)
}
\arguments{
\item{object}{\code{ModelParameter} object.}
}
\value{
A boolean \code{isNotFixed} giving TRUE if the model parameters is not fixed, FALSE otherwise.
}
\description{
Boolean to set if a model parameters is not fixed or not.
}
