% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{isNotFixedMu}
\alias{isNotFixedMu}
\title{Boolean to set if \code{mu} is not fixed or not.}
\usage{
isNotFixedMu(object)
}
\arguments{
\item{object}{\code{ModelParameter} object.}
}
\value{
A boolean \code{isNotFixedMu} giving TRUE if \code{mu} is not fixed, FALSE otherwise.
}
\description{
Boolean to set if \code{mu} is not fixed or not.
}
