% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\name{isTimeInBetweenBounds}
\alias{isTimeInBetweenBounds}
\title{Set the constraint on the sampling times bounds.}
\usage{
isTimeInBetweenBounds(object, time)
}
\arguments{
\item{object}{A \code{SamplingConstraint} object.}

\item{time}{A \code{SamplingTimes} object.}
}
\value{
A boolean that give TRUE/FALSE if the constraint on on the sampling times bounds are satisfied.
}
\description{
Set the constraint on the sampling times bounds.
}
