% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportAndPlots.R
\name{knitrAdministrationParameters}
\alias{knitrAdministrationParameters}
\title{Set the table knitrAdministrationParameters.}
\usage{
knitrAdministrationParameters(object)
}
\arguments{
\item{object}{An object \code{knitrAdministrationParameters} from the class \linkS4class{ReportAndPlots}.}
}
\value{
The table knitrAdministrationParameters.
}
\description{
Set the table knitrAdministrationParameters.
}
