% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportAndPlots.R
\name{knitrFIM}
\alias{knitrFIM}
\title{Set the table knitrFIM.}
\usage{
knitrFIM(object)
}
\arguments{
\item{object}{An object \code{knitrFIM} from the class \linkS4class{ReportAndPlots}.}
}
\value{
The table knitrFIM.
}
\description{
Set the table knitrFIM.
}
