% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportAndPlots.R
\name{knitrInitialDesigns}
\alias{knitrInitialDesigns}
\title{Set the table knitrInitialDesigns.}
\usage{
knitrInitialDesigns(object)
}
\arguments{
\item{object}{An object \code{knitrInitialDesigns} from the class \linkS4class{ReportAndPlots}.}
}
\value{
The table kknitrInitialDesigns.
}
\description{
Set the table knitrInitialDesigns.
}
