% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportAndPlots.R
\name{knitrModelEquations}
\alias{knitrModelEquations}
\title{Set the table knitrModelEquations.}
\usage{
knitrModelEquations(object)
}
\arguments{
\item{object}{An object \code{knitrModelEquations} from the class \linkS4class{ReportAndPlots}.}
}
\value{
The table knitrModelEquations.
}
\description{
Set the table knitrModelEquations.
}
