% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportAndPlots.R
\name{knitrModelError}
\alias{knitrModelError}
\title{Set the table knitrModelError.}
\usage{
knitrModelError(object)
}
\arguments{
\item{object}{An object \code{knitrModelError} from the class \linkS4class{ReportAndPlots}.}
}
\value{
The table knitrModelError.
}
\description{
Set the table knitrModelError.
}
