% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportAndPlots.R
\name{knitrModelParameters}
\alias{knitrModelParameters}
\title{Set the table knitrModelParameters.}
\usage{
knitrModelParameters(object)
}
\arguments{
\item{object}{An object \code{knitrModelParameters} from the class \linkS4class{ReportAndPlots}.}
}
\value{
The table knitrModelParameters.
}
\description{
Set the table knitrModelParameters.
}
