% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportAndPlots.R
\name{knitrOptimalDesign}
\alias{knitrOptimalDesign}
\title{Set the table knitrOptimalDesign.}
\usage{
knitrOptimalDesign(object)
}
\arguments{
\item{object}{An object \code{knitrFIM} from the class \linkS4class{ReportAndPlots}.}
}
\value{
The table knitrOptimalDesign.
}
\description{
Set the table knitrOptimalDesign.
}
