% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPDModels.R
\docType{data}
\name{LibraryOfPDModels}
\alias{LibraryOfPDModels}
\title{LibraryOfPDModels Class}
\format{
An object of class \code{PFIM::LibraryOfPDModels} (inherits from \code{PFIM::LibraryOfModels}, \code{S7_object}) of length 1.
}
\usage{
LibraryOfPDModels
}
\description{
The \code{LibraryOfPDModels} class is a specialized container for managing
and storing Pharmacodynamic (PD) model definitions.
}
\details{
This class inherits from \code{\link{LibraryOfModels}} and provides a
dedicated structure for pharmacodynamic responses. It is designed to handle
various PD mechanisms, including direct effect models (Emax, Sigmoid Emax),
indirect response models (Turnover), and kinetic-pharmacodynamic (K-PD)
structures.
}
\section{Slots}{

\describe{
\item{\code{models}}{A named list of PD model structures (e.g., Emax, Indirect Response).}
}}

\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
\keyword{datasets}
