% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKPDModels.R
\docType{class}
\name{LibraryOfPKPDModels-class}
\alias{LibraryOfPKPDModels-class}
\alias{LibraryOfPKPDModels}
\title{Class "LibraryOfPKPDModels"}
\description{
The class \code{LibraryOfPKPDModels} represents the library of PKPD models.
The class \code{LibraryOfPKPDModels} inherits from the class \code{LibraryOfModels}.
}
