% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\docType{class}
\name{LogNormal-class}
\alias{LogNormal-class}
\alias{LogNormal}
\title{Class "LogNormal"}
\description{
The class defines all the required methods for a LogNormal distribution object.
The class \code{LogNormal} inherits from the class \code{Distribution}.
}
