% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormalDistribution.R
\docType{class}
\name{LogNormalDistribution-class}
\alias{LogNormalDistribution-class}
\alias{LogNormalDistribution}
\title{Class "LogNormalDistribution"}
\description{
Class \code{LogNormalDistribution} represent a Log-Normal distribution with
mean \code{mean_log_Gaussian} and standard deviation \code{sd_log_Gaussian}.
}
\section{Objects from the Class}{
 LogNormalDistribution objects
are typically created by calls to \code{\ {pfim}} and contain the following slots:

\describe{
\item{\code{mean_log_Gaussian}:}{A numeric givnig the mean of the log Normal Distribution.}
\item{\code{sd_log_Gaussian}:}{A numeric giving the standard deviation of the log Normal Distribution.}
}
}

