% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\alias{Model}
\title{Class "Model" representing a Model}
\description{
A class storing information concerning the models in the \code{LibraryOfModels}.
}
\section{Objects from the Class  \code{Model}}{

Objects form the Class \code{Model} can be created by calls of the form \code{Model(...)} where
(...) are the parameters for the \code{Model} object.
}

\section{Slots for the \code{Model} objects}{

\describe{
\item{\code{nameModel}:}{A character string giving the name of the model.}
\item{\code{descriptionModel}:}{A list of character string giving the characterisation of the model (name, administration, number of compartment)}
\item{\code{equationsModel}:}{A object \code{ModelEquations} giving the equations of the model.}
}
}

