% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R
\docType{class}
\name{ModelAnalytic-class}
\alias{ModelAnalytic-class}
\alias{ModelAnalytic}
\title{Class "ModelAnalytic"}
\description{
The class \code{Model} defines information concerning the construction of an analytical model.
The class \code{ModelAnalytic} inherits from the class \code{Model}.
}
