% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalyticBolus.R
\docType{class}
\name{ModelAnalyticBolus-class}
\alias{ModelAnalyticBolus-class}
\alias{ModelAnalyticBolus}
\title{Class "ModelAnalyticBolus"}
\description{
The class \code{Model} defines information concerning the construction of an analytical bolus model.
The class \code{ModelAnalyticBolus} inherits from the class \code{ModelAnalytic}.
}
