% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalyticBolusSteadyState.R
\docType{class}
\name{ModelAnalyticBolusSteadyState-class}
\alias{ModelAnalyticBolusSteadyState-class}
\alias{ModelAnalyticBolusSteadyState}
\title{Class "ModelAnalyticBolusSteadyState"}
\description{
The class \code{Model} defines information concerning the construction of an analytical model in steady state.
The class \code{ModelAnalyticBolusSteadyState} inherits from the class \code{ModelAnalyticSteadyState}.
}
