% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalyticInfusion.R
\docType{class}
\name{ModelAnalyticInfusion-class}
\alias{ModelAnalyticInfusion-class}
\alias{ModelAnalyticInfusion}
\title{Class "ModelAnalyticInfusion"}
\description{
The class \code{Model} defines information concerning the construction of an analytical model in infusion.
The class \code{ModelAnalyticInfusion} inherits from the class \code{ModelInfusion}.
}
