% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalyticInfusionSteadyState.R
\docType{class}
\name{ModelAnalyticInfusionSteadyState-class}
\alias{ModelAnalyticInfusionSteadyState-class}
\alias{ModelAnalyticInfusionSteadyState}
\title{Class "ModelAnalyticInfusionSteadyState"}
\description{
The class \code{Model} defines information concerning the construction of an analytical model in infusion in steady state.
The class \code{ModelAnalyticInfusionSteadyState} inherits from the class \code{ModelAnalyticInfusion}.
}
