% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalyticSteadyState.R
\docType{class}
\name{ModelAnalyticSteadyState-class}
\alias{ModelAnalyticSteadyState-class}
\alias{ModelAnalyticSteadyState}
\title{Class "ModelAnalyticSteadyState"}
\description{
The class \code{ModelAnalyticSteadyState} defines information concerning the construction of an analytical model steady state.
The class \code{ModelAnalyticSteadyState} inherits from the class \code{ModelAnalytic}.
}
