% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCovariateEquation.R
\name{ModelCovariateEquation}
\alias{ModelCovariateEquation}
\title{ModelCovariateEquation}
\usage{
ModelCovariateEquation(beta = 0, combinedEffect = 0, value = 0)
}
\arguments{
\item{beta}{beta}

\item{combinedEffect}{combinedEffect}

\item{value}{value}
}
\description{
The class \code{ModelCovariateEquation} is an abstract class used to define models for covariates.
}
\details{
ModelCovariateEquation
}
