% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEquations.R
\docType{class}
\name{ModelEquations-class}
\alias{ModelEquations-class}
\alias{ModelEquations}
\title{Class "ModelEquations" representing a the equations of a model.}
\description{
A class storing information concerning the model equations of the models in the \linkS4class{LibraryOfModels}.
}
\section{Objects from the class \code{ModelEquations}}{

Objects form the Class \code{ModelEquations} can be created by calls of the form \code{ModelEquations(...)} where
(...) are the parameters for the \code{ModelEquations} objects.
}

\section{Slots for  \code{ModelEquations} objects}{

\describe{
\item{\code{equations}:}{A list giving the equations of the model.}
\item{\code{allParameters}:}{A vector giving all the parameters of the model.}
}
}

