% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{ModelError}
\alias{ModelError}
\title{ModelError Class}
\usage{
ModelError(
  output = "output",
  equation = expression(),
  derivatives = list(),
  sigmaInter = 0.1,
  sigmaSlope = 0,
  sigmaInterFixed = FALSE,
  sigmaSlopeFixed = FALSE,
  cError = 1
)
}
\arguments{
\item{output}{A string giving the model error output.}

\item{equation}{A expression giving the model error equation.}

\item{derivatives}{A list giving the derivatives of the model error equation.}

\item{sigmaInter}{A double giving the sigma inter.}

\item{sigmaSlope}{A double giving the sigma slope}

\item{sigmaInterFixed}{A boolean giving if the  sigma inter is fixed or not.}

\item{sigmaSlopeFixed}{A boolean giving if the  sigma slope is fixed or not.}

\item{cError}{A integer giving the power parameter.}
}
\description{
The class \code{ModelError} is used to defined a model error.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
# 1. Define an additive error model
# sigmaInter: intercept (additive), sigmaSlope: slope (proportional)
additiveError = ModelError(
  output     = "RespPK",
  sigmaInter = 0.1,
  sigmaSlope = 0.0
)
print(additiveError)

# 2. Define a combined error model (Additive + Proportional)
combinedError = ModelError(
  output     = "RespPK",
  sigmaInter = 0.05,
  sigmaSlope = 0.15
)
print(combinedError)
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
