% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfusionEquations.R
\docType{class}
\name{ModelInfusionEquations-class}
\alias{ModelInfusionEquations-class}
\alias{ModelInfusionEquations}
\title{Class "ModelInfusionEquations" representing a model with infusion equations}
\description{
A class giving information on the infusion equations regarding the equations of the model.
}
\section{Objects from the class}{
 \code{{ModelInfusionEquations}} objects are typically created by calls to \code{{ModelInfusionEquations}} and contain the following slots:

\describe{
\item{\code{object}:}{An object from the class \code{ModelEquations}}
}
}

