% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfusionODEquations.R
\docType{class}
\name{ModelInfusionODEquations-class}
\alias{ModelInfusionODEquations-class}
\alias{ModelInfusionODEquations}
\title{Class "ModelInfusionODEquations" representing a model with infusion equations in ODE model.}
\description{
A class giving information on the infusion equations regarding the equations of the model.
}
\section{Objects from the class}{
 \code{ModelInfusionODEquations} objects are typically created by calls to \code{ModelInfusionODEquations} and contain the following slots:

\describe{
\item{\code{duringInfusionResponsesEquations}:}{A list containing the equations during the infusion.}
\item{\code{afterInfusionResponsesEquations}:}{A list containing the equations after the infusion.}
\item{\code{duringInfusionDerivatives}:}{A list containing the derivatives during the infusion.}
\item{\code{afterInfusionDerivatives}:}{A list containing the derivatives after the infusion.}
\item{\code{derivatives}:}{A list contaning the derivatives of the model.}
}
}

