% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODE.R
\docType{class}
\name{ModelODE-class}
\alias{ModelODE-class}
\alias{ModelODE}
\title{Class "ModelODE"}
\description{
The class \code{ModelODE} defines information concerning the construction of an ode model.
The class \code{ModelODE} inherits from the class \code{Model}.
}
