% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODEBolus.R
\docType{class}
\name{ModelODEBolus-class}
\alias{ModelODEBolus-class}
\alias{ModelODEBolus}
\title{Class "ModelODEBolus"}
\description{
The class \code{ModelODEBolus} defines information concerning the construction of an ode model bolus.
The class \code{ModelODEBolus} inherits from the class \code{ModelBolus}.
}
