% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODEDoseInEquations.R
\docType{class}
\name{ModelODEDoseInEquations-class}
\alias{ModelODEDoseInEquations-class}
\alias{ModelODEDoseInEquations}
\title{Class "ModelODEDoseInEquations"}
\description{
The class \code{ModelODEDoseInEquations} defines information concerning the construction of an ode model
where the dose is in the model equations. The class \code{ModelODEDoseInEquations} inherits from the class \code{ModelODE}.
}
