% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODEInfusion.R
\docType{class}
\name{ModelODEInfusion-class}
\alias{ModelODEInfusion-class}
\alias{ModelODEInfusion}
\title{Class "ModelODEInfusion"}
\description{
The class \code{ModelODEInfusion} defines information concerning the construction of an ode model in infusion.
The class \code{ModelODEInfusion} inherits from the class \code{ModelInfusion}.
}
