% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODEInfusion.R
\name{ModelODEInfusion}
\alias{ModelODEInfusion}
\title{ModelODEInfusion Class}
\usage{
ModelODEInfusion(
  name = character(0),
  modelParameters = list(),
  samplings = numeric(0),
  modelEquations = list(),
  wrapper = function() NULL,
  outputFormula = list(),
  outputNames = character(0),
  variableNames = character(0),
  outcomesWithAdministration = character(0),
  outcomesWithNoAdministration = character(0),
  modelError = list(),
  odeSolverParameters = list(),
  parametersForComputingGradient = list(),
  initialConditions = numeric(0),
  functionArguments = character(0),
  functionArgumentsSymbol = list()
)
}
\arguments{
\item{name}{A \code{character} vector specifying the name of the model.}

\item{modelParameters}{A \code{list} of objects defining the model parameters.}

\item{samplings}{A \code{numeric} vector specifying the planned sampling times.}

\item{modelEquations}{A \code{list} containing the system of equations (analytical or ODEs).}

\item{wrapper}{A \code{function} wrapper used to interface the model (defaults to \code{function() NULL}).}

\item{outputFormula}{A \code{list} of mathematical formulas for the model outputs.}

\item{outputNames}{A \code{character} vector defining the names of the output variables.}

\item{variableNames}{A \code{character} vector defining the names of the state variables.}

\item{outcomesWithAdministration}{A \code{character} vector specifying outcomes associated with drug administration.}

\item{outcomesWithNoAdministration}{A \code{character} vector specifying outcomes without drug administration.}

\item{modelError}{A \code{list} defining the residual error model structure.}

\item{odeSolverParameters}{A \code{list} of parameters for the ODE solver (e.g., \code{atol}, \code{rtol}).}

\item{parametersForComputingGradient}{A \code{list} of parameters required for numerical gradient computation.}

\item{initialConditions}{A \code{numeric} vector specifying the initial state of the system.}

\item{functionArguments}{A \code{character} vector of arguments required by the model function.}

\item{functionArgumentsSymbol}{A \code{list} of symbols representing the function arguments.}
}
\description{
The \code{ModelODEInfusion} class is specifically designed to define ODE-based
models for infusion-type administrations. It extends the \code{ModelInfusion}
properties to handle continuous drug delivery through differential equations.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
