% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODEInfusionDoseInEquations.R
\docType{class}
\name{ModelODEInfusionDoseInEquations-class}
\alias{ModelODEInfusionDoseInEquations-class}
\alias{ModelODEInfusionDoseInEquations}
\title{Class "ModelODEInfusionDoseInEquations"}
\description{
The class \code{ModelODEInfusionDoseInEquations} defines information concerning the construction of an ode model
in infusion where the dose is in the model equations. The class \code{ModelODEInfusionDoseInEquations} inherits from the class \code{ModelODEInfusion}.
}
