% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODEquations.R
\docType{class}
\name{ModelODEquations-class}
\alias{ModelODEquations-class}
\alias{ModelODEquations}
\title{Class "ModelODEquations" representing the equations of an ODE model}
\description{
A class storing information concerning the equations for the ODE models in the \code{LibraryOfModels}.
}
\section{Objects from the class }{
 \code{ModelODEquations} objects are typically created by calls to \code{ModelODEquations} and contain the following slots:

\describe{
\item{\code{derivatives}:}{A list of expression giving the the derivatives of the model.}
}
}

