% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{ModelParameter}
\alias{ModelParameter}
\title{ModelParameter}
\usage{
ModelParameter(
  name = character(0),
  gamma = 0,
  distribution = Distribution(),
  fixedMu = FALSE,
  fixedOmega = FALSE,
  value = numeric(0)
)
}
\arguments{
\item{name}{A string giving the name of the parameter.}

\item{gamma}{A numeric giving the value of the variance for the inter occasion variability (iov).}

\item{distribution}{A string giving the distribution of the parameter.}

\item{fixedMu}{A Boolean setting TRUE/FALSE if the mu is estimated or not.}

\item{fixedOmega}{A Boolean setting TRUE/FALSE if the omega is estimated or not.}

\item{value}{A numeric giving the value of the covariates effects.}
}
\description{
The class \code{ModelParameter} is used to defined the model parameters.
}
\details{
ModelParameter
}
