% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelVariable.R
\docType{class}
\name{ModelVariable-class}
\alias{ModelVariable-class}
\alias{ModelVariable}
\title{Class "ModelVariable"}
\description{
Class "ModelVariable" represents an initial variable for ODE model.
}
\section{Objects from the class}{
 \code{ModelVariable} objects are typically created by calls to \code{ModelVariable} and contain the following slots:
}

\section{Slots for \code{ModelVariable} objects}{

\describe{
\item{\code{name}:}{A character string giving the name of the initial variable of an ODE model.}
\item{\code{value}:}{A numeric giving the value of the initial variable of an ODE model.}
}
}

