% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{modifyArm}
\alias{modifyArm}
\title{Modify an arm of a design.}
\usage{
modifyArm(object, name, arm)
}
\arguments{
\item{object}{A \code{Design} object.}

\item{name}{A character string giving the name of the \code{Arm} object to be modified in the \code{Design} object.}

\item{arm}{An \code{Arm} object.}
}
\value{
The \code{Design} object with the modified arm.
}
\description{
Modify an arm of a design.
}
