% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{modifySamplingTimes}
\alias{modifySamplingTimes}
\title{Modify the sampling times of an arm.}
\usage{
modifySamplingTimes(object, outcome, samplingTimes)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{outcome}{A character string giving the name of the outcome ie the name of the response.}

\item{samplingTimes}{A vector of numeric giving the new sampling times.}
}
\value{
The object \code{Arm} object with its new sampling times.
}
\description{
Modify the sampling times of an arm.
}
