% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\docType{class}
\name{Normal-class}
\alias{Normal-class}
\alias{Normal}
\title{Class "Normal"}
\description{
The class defines all the required methods for a Normal distribution object.
The class \code{Normal} inherits from the class \code{Distribution}.
}
