% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalDistribution.R
\docType{class}
\name{NormalDistribution-class}
\alias{NormalDistribution-class}
\alias{NormalDistribution}
\title{Class "NormalDistribution"}
\description{
Class \code{LogNormalDistribution} represent a Normal distribution
}
\section{Objects from the class}{
 \code{LogNormalDistribution} objects
are typically created by calls to the class \code{NormalDistribution}.
}

