% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\name{numberOfSamplingTimesIsOptimisable}
\alias{numberOfSamplingTimesIsOptimisable}
\title{Set the number of sampling times that are optimisable.}
\usage{
numberOfSamplingTimesIsOptimisable(object, FixedNumberTimes)
}
\arguments{
\item{object}{A \code{SamplingConstraint} object.}

\item{FixedNumberTimes}{A numeric giving the number of sampling times to be fixed.}
}
\value{
Set the number of sampling times that are optimisable in the constraints.
}
\description{
Set the number of sampling times that are optimisable.
}
