% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\docType{class}
\name{Optimization-class}
\alias{Optimization-class}
\alias{Optimization}
\title{Class "Optimization"}
\description{
A class storing information concerning Optimization.
}
\section{Objects from the class \code{Optimization}}{

Objects form the class \code{Optimization} can be created by calls of the form \code{Optimization(...)} where
(...) are the parameters for the \code{Optimization} objects.
}

\section{Slots for \code{Optimization} objects}{

 \describe{
   \item{\code{showProcess}:}{A logical if the optimization process is shown or not.}
   \item{\code{FisherMatrices}:}{A list of all the Fisher matrices used in the optimization process.}
   \item{\code{combinedTimes}:}{A list giving all he combinaison of n elements for a vector of times. }
   \item{\code{arms}:}{A list giving all the arms on which the optimization processis done.}
 }
}

