% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationAlgorithm.R
\docType{class}
\name{OptimizationAlgorithm-class}
\alias{OptimizationAlgorithm-class}
\alias{OptimizationAlgorithm}
\title{Class "OptimizationAlgorithm"}
\description{
A class storing information concerning the optimization algorithm.
}
\section{Objects from the class}{

Objects form the class \code{OptimizationAlgorithm} can be created by calls of the form \code{OptimizationAlgorithm(...)} where
(...) are the parameters for the \code{OptimizationAlgorithm} objects.
}

\section{Slots for \code{Administration} objects}{

 \describe{
   \item{\code{name}:}{A character string giving the name of the optimization algorithm.}
   \item{\code{parameters}:}{A list giving the parameters of the optimization algorithm.}
 }
}

