% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationAlgorithm.R
\name{optimize}
\alias{optimize}
\title{Optimize a design.}
\usage{
optimize(object, optimizerParameters, optimizationObject)
}
\arguments{
\item{object}{An object from the class \linkS4class{OptimizationAlgorithm}.}

\item{optimizerParameters}{A list giving the optimization parameters.}

\item{optimizationObject}{An object giving the optimization algorithm.}
}
\value{
A list giving the results if the optimization.
}
\description{
Optimize a design.
}
