% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/PGBOAlgorithm.R,
%   R/PSOAlgorithm.R, R/SimplexAlgorithm.R, R/FedorovWynnAlgorithm.R,
%   R/MultiplicativeAlgorithm.R
\name{optimizeDesign}
\alias{optimizeDesign}
\title{Optimize Study Design}
\usage{
optimizeDesign(optimizationObject, optimizationAlgorithm, ...)
}
\arguments{
\item{optimizationObject}{An object of class \code{\link{Optimization}}
containing the design space, model, and objective function settings.}

\item{optimizationAlgorithm}{An object of the algorithm class to use
(\code{MultiplicativeAlgorithm}, \code{FedorovWynnAlgorithm}, \code{PSOAlgorithm}, etc.).}

\item{\dots}{Additional arguments passed to methods}
}
\value{
The updated \code{optimizationObject} with optimized designs,
final FIM value, and algorithm-specific outputs stored in the relevant slots.
}
\description{
Dispatches the optimization of a study design to the appropriate algorithm.
Methods are available for all PFIM optimization algorithms:
\code{\link{MultiplicativeAlgorithm}}, \code{\link{FedorovWynnAlgorithm}},
\code{\link{SimplexAlgorithm}}, \code{\link{PSOAlgorithm}},
and \code{\link{PGBOAlgorithm}}.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
