% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDModel.R
\docType{class}
\name{PDModel-class}
\alias{PDModel-class}
\alias{PDModel}
\title{Class "PDModel" representing a PD model.}
\description{
A class storing information concerning the PD models in the \linkS4class{LibraryOfModels}.
}
\section{Objects from the class}{
 \code{PDModel} objects are typically created by calls to \code{PDModel}.
}

\section{Slots for the \code{PDModel} objects, that are heritated from the class \linkS4class{Model}}{

\describe{
\item{\code{nameModel}:}{A character string giving the name of the model.}
\item{\code{descriptionModel}:}{A list of character string giving the characterisation of the model (name, administration, number of compartment)}
\item{\code{equationsModel}:}{A object \code{ModelEquations} giving the equations of the model.}
}
}

