% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{PFIMProject}
\alias{PFIMProject}
\title{PFIMProject Class}
\usage{
PFIMProject(
  name = character(0),
  modelEquations = list(),
  modelFromLibrary = list(),
  modelParameters = list(),
  modelError = list(),
  optimizer = character(0),
  optimizerParameters = list(),
  outputs = list(),
  designs = list(),
  fimType = character(0),
  fim = Fim(),
  odeSolverParameters = list()
)
}
\arguments{
\item{name}{A string representing the name of the project or evaluation study.}

\item{modelEquations}{A list containing the mathematical equations of the model.}

\item{modelFromLibrary}{A list specifying the pre-defined model selected from the PFIM library.}

\item{modelParameters}{A list defining the fixed effects and random effects (variances) of the model.}

\item{modelError}{A list specifying the residual error model (e.g., constant, proportional, or combined).}

\item{optimizer}{A string identifying the optimization algorithm to be used (e.g., "Simplex", "Fedov").}

\item{optimizerParameters}{A list of settings for the chosen optimizer (e.g., iterations, tolerance).}

\item{outputs}{A list defining the observation variables or responses of the model.}

\item{designs}{A list of `Design` objects representing the experimental protocols to be evaluated.}

\item{fimType}{A string specifying the FIM calculation method (e.g., "Population", "Individual", "Bayesian").}

\item{fim}{An object of class \code{Fim} representing the computed Fisher Information Matrix.}

\item{odeSolverParameters}{A list containing technical settings for the ODE solver, such as \code{atol} and \code{rtol}.}
}
\description{
The `PFIMProject` class is the central orchestrator for a Population Fisher
Information Matrix (PFIM) analysis. It encapsulates all necessary components
for design evaluation or optimization, including structural models, statistical
parameters, experimental designs, and optimization settings.
}
\section{Slots}{

\describe{
\item{\code{name}}{\code{character}}

\item{\code{modelEquations}}{\code{list}}

\item{\code{modelFromLibrary}}{\code{list}}

\item{\code{modelParameters}}{\code{list}}

\item{\code{modelError}}{\code{list}}

\item{\code{optimizer}}{\code{character}}

\item{\code{optimizerParameters}}{\code{list}}

\item{\code{outputs}}{\code{list}}

\item{\code{designs}}{\code{list}}

\item{\code{fimType}}{\code{character}}

\item{\code{fim}}{\code{Fim}}

\item{\code{odeSolverParameters}}{\code{list}}
}}

\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
