% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportAndPlots.R
\name{PFIMProjectReportEvaluation}
\alias{PFIMProjectReportEvaluation}
\title{Generate the html report for the evaluation.}
\usage{
PFIMProjectReportEvaluation(object, inputPath, outputPath, plotOptions)
}
\arguments{
\item{object}{\code{PFIMProject} object.}

\item{inputPath}{A string giving the input path.}

\item{outputPath}{A string giving the output path.}

\item{plotOptions}{A list giving the options.}
}
\value{
The html report for the evaluation.
}
\description{
Generate the html report for the evaluation.
}
