% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKModel.R
\docType{class}
\name{PKModel-class}
\alias{PKModel-class}
\alias{PKModel}
\title{Class "PKModel" representing a PK model.}
\description{
A class storing information concerning the PK models in the \linkS4class{LibraryOfModels}.
}
\section{Objects from the class \code{PKModel}}{
 objects are typically created by calls to \code{PKModel}.
}

\section{Slots for the \code{PKModel} objects, that are heritated from the class \linkS4class{Model}}{

\describe{
\item{\code{nameModel}:}{A character string giving the name of the model.}
\item{\code{descriptionModel}:}{A list of character string giving the characterisation of the model (name, administration, number of compartment)}
\item{\code{equationsModel}:}{A object \code{ModelEquations} giving the equations of the model.}
}
}

