% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKPDModel.R
\docType{class}
\name{PKPDModel-class}
\alias{PKPDModel-class}
\alias{PKPDModel}
\alias{PKPDModelModel}
\title{Class "PKPDModel" representing a PKPDModel model.}
\description{
A class storing information concerning the PKPDModel models in the \linkS4class{LibraryOfModels}.
}
\section{Objects from the class \code{PKPDModel}}{
 objects are typically created by calls to \code{PKPDModel}.
}

\section{Slots for the \code{PKPDModel} objects, that are heritated from the class \linkS4class{Model}}{

\describe{
\item{\code{nameModel}:}{A character string giving the name of the model.}
\item{\code{descriptionModel}:}{A list of character string giving the characterisation of the model (name, administration, number of compartment)}
\item{\code{equationsModel}:}{A object \code{ModelEquations} giving the equations of the model.}
}
}

