% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotCriteria}
\alias{plotCriteria}
\title{Plot the D criteria over time.}
\usage{
plotCriteria(object, ...)
}
\arguments{
\item{object}{\code{PFIMProject} object.}

\item{...}{A list giving the plot options.}
}
\value{
A plot of the D criteria over iterations for a Design optimization.
}
\description{
Plot the D criteria over time.
}
