% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEvaluation.R
\docType{class}
\name{PlotEvaluation-class}
\alias{PlotEvaluation-class}
\alias{PlotEvaluation}
\title{Class "PlotEvaluation"}
\description{
A class storing information concerning the design evaluation.
The class \code{PlotEvaluation} inherits from the class \code{Evaluation}.
}
