% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{plotEvaluation}
\alias{plotEvaluation}
\title{plotEvaluation: Plot model responses for design evaluation}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject} containing the design.}

\item{plotOptions}{A \code{list} specifying graphical parameters (e.g., \code{xlim}, \code{ylim}, \code{main}, \code{col}).}
}
\value{
A series of plots representing the expected model responses and sampling schedules.
}
\description{
Generates graphical representations of the model responses based on the
design and parameters defined in the PFIM project.
}
\examples{
\dontrun{
# 1. Define custom plot options, e.g. form the Vignette 01:
myPlotOptions = list( unitTime = c( "hour" ), unitOutcomes= c( "mcg/mL" , "DI\%" ) )

# 2. Generate the plots for the project
plotEvaluation(myPFIMproject, plotOptions = myPlotOptions)
}
}
