% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{plotFrequencies}
\alias{plotFrequencies}
\title{plotFrequencies: plot optimal frequencies (FedorovWynnAlgorithm)}
\arguments{
\item{optimization}{An \code{Optimization} object.}
}
\description{
plotFrequencies: plot optimal frequencies (FedorovWynnAlgorithm)
}
