% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{plotFrequenciesFedorovWynnAlgorithm}
\alias{plotFrequenciesFedorovWynnAlgorithm}
\title{plotFrequenciesFedorovWynnAlgorithm}
\arguments{
\item{optimization}{An \code{Optimization} object.}

\item{optimizationAlgorithm}{A \code{FedorovWynnAlgorithm} object.}
}
\value{
A ggplot bar chart of optimal protocol frequencies.
}
\description{
plotFrequenciesFedorovWynnAlgorithm
}
