% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotFrequenciesOptimisation}
\alias{plotFrequenciesOptimisation}
\title{Plot the frequencies for the FedorovWynn algorithm.}
\usage{
plotFrequenciesOptimisation(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
A barplot of the the frequencies for the FedorovWynn algorithm..
}
\description{
Plot the frequencies for the FedorovWynn algorithm.
}
