% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{plotOutcomesEvaluation}
\alias{plotOutcomesEvaluation}
\alias{plotOutcomesEvaluation,Design-method}
\title{plotOutcomesEvaluation}
\usage{
plotOutcomesEvaluation(object, initialDesign, model, plotOptions)

\S4method{plotOutcomesEvaluation}{Design}(object, initialDesign, model, plotOptions)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}

\item{initialDesign}{An object \code{design} from the class \linkS4class{Design}.}

\item{model}{An object \code{model} from the class \linkS4class{Model}.}

\item{plotOptions}{A list containing the plot options.}
}
\value{
A list containing the plots the evaluation of the outcomes.
}
\description{
Plot the evaluation of the outcomes.
}
