% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{plotOutcomesGradient}
\alias{plotOutcomesGradient}
\alias{plotOutcomesGradient,Design-method}
\title{plotOutcomesGradient}
\usage{
plotOutcomesGradient(object, initialDesign, model, plotOptions)

\S4method{plotOutcomesGradient}{Design}(object, initialDesign, model, plotOptions)
}
\arguments{
\item{object}{An object \code{design} from the class \linkS4class{Design}.}

\item{initialDesign}{An object \code{design} from the class \linkS4class{Design}.}

\item{model}{An object \code{model} from the class \linkS4class{Model}.}

\item{plotOptions}{A list containing the plot options.}
}
\value{
A list containing the plots the evaluation of the outcome gradients..
}
\description{
Plot the evaluation of the outcome gradients.
}
