% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotResponse}
\alias{plotResponse}
\title{Plot the concentration over time of a model.}
\usage{
plotResponse(object, plotOptions)
}
\arguments{
\item{object}{\code{PFIMProject} object.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
A list containing the plots of the concentration over time of a model.
}
\description{
Plot the concentration over time of a model.
}
