% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{plotRSEFIM}
\alias{plotRSEFIM}
\title{Plot Relative Standard Errors (RSE\%) for Population FIM}
\arguments{
\item{fim}{An object of class \code{\link{PopulationFim}} containing the
calculated RSE values.}

\item{evaluation}{An object of class \code{Evaluation} providing the
context and parameter names for the plot.}
}
\value{
A plot object (typically \code{ggplot2} or \code{lattice}) displaying
the RSE\% for structural and variance parameters.
}
\description{
Plot Relative Standard Errors (RSE\%) for Population FIM
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
