% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{plotSEFIM}
\alias{plotSEFIM}
\title{Plot Standard Errors from the Population FIM}
\arguments{
\item{fim}{An object of class \code{\link{PopulationFim}} containing the
calculated SE and RSE values.}

\item{evaluation}{An object of class \code{Evaluation} providing the
context of the model being plotted.}
}
\value{
A \code{ggplot2} or base R plot object showing the bar plot of the
parameter uncertainties.
}
\description{
The \code{plotSEFIM} method generates a diagnostic bar plot representing the
Standard Errors (SE) or Relative Standard Errors (RSE\%) for all estimated
parameters. This visualization is crucial for comparing the precision
between structural parameters (fixed effects) and variance components.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
