% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotSensitivity}
\alias{plotSensitivity}
\title{Plot the sensitivity indices of a model over time.}
\usage{
plotSensitivity(object, plotOptions)
}
\arguments{
\item{object}{\code{PFIMProject} object.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
A list containing the plots of the sensitivity indices of a model over time.
}
\description{
Plot the sensitivity indices of a model over time.
}
