% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{plotSensitivityIndices}
\alias{plotSensitivityIndices}
\title{plotSensitivityIndices: plot sensitivity indices for the optimal design}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject} containing the design and model specifications.}

\item{plotOptions}{A \code{list} specifying graphical parameters for the sensitivity plots.}
}
\value{
A series of plots showing the sensitivity of model responses over time for each parameter.
}
\description{
Generates plots for the sensitivity indices (partial derivatives) of the model
responses with respect to the population parameters. This is essential for
identifying the optimal sampling times where the model is most sensitive to
parameter changes.
}
\examples{
\dontrun{
# 1. Define custom plot options, e.g. form the Vignette 01:
myPlotOptions = list( unitTime = c( "hour" ), unitOutcomes= c( "mcg/mL" , "DI\%" ) )

# 2. Generate sensitivity index plots for the current project
plotSensitivityIndices(myPFIMproject, plotOptions = myPlotOptions)
}
}
