% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotWeightOptimisation}
\alias{plotWeightOptimisation}
\title{Plot the optimal weights for the Multiplicative algorithm.}
\usage{
plotWeightOptimisation(object, threshold)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}

\item{threshold}{A numeric giving the threshold for the weights.}
}
\value{
A barplot of the optimal weights above the threshold.
}
\description{
Plot the optimal weights for the Multiplicative algorithm.
}
