% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{plotWeights}
\alias{plotWeights}
\title{plotWeights: plot algorithm weights (MultiplicativeAlgorithm)}
\arguments{
\item{optimization}{An \code{Optimization} object.}
}
\description{
plotWeights: plot algorithm weights (MultiplicativeAlgorithm)
}
