% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationFim.R
\docType{class}
\name{PopulationFim-class}
\alias{PopulationFim-class}
\alias{PopulationFim}
\title{Class "PopulationFim"}
\description{
A class storing information regarding the population Fisher matrix.
The class \code{PopulationFim} inherits from the class \code{Fim}.
}
