% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationFim.R
\name{PopulationFim}
\alias{PopulationFim}
\title{PopulationFim Class}
\usage{
PopulationFim(
  fisherMatrix = numeric(0),
  fixedEffects = numeric(0),
  varianceEffects = numeric(0),
  SEAndRSE = list(),
  condNumberFixedEffects = 0,
  condNumberVarianceEffects = 0,
  shrinkage = numeric(0)
)
}
\arguments{
\item{fisherMatrix}{A matrix giving the numerical values of the Fim.}

\item{fixedEffects}{A matrix giving the numerical values of the fixed effects of the Fim.}

\item{varianceEffects}{A matrix giving the numerical values of variance effects of the Fim.}

\item{SEAndRSE}{A data frame giving the calculated values of SE and RSE for parameters.}

\item{condNumberFixedEffects}{The condition number of the fixed effects portion of the Fim.}

\item{condNumberVarianceEffects}{The condition number of the variance effects portion of the Fim.}

\item{shrinkage}{A vector giving the shrinkage values for the random effects.}
}
\description{
The \code{PopulationFim} class is a child of the \code{Fim} class. It is
specifically designed to store and manage the Fisher Information Matrix
calculated for population-level analyses. Unlike individual FIMs, it
incorporates the variance-covariance components of the random effects,
providing a measure of the information content regarding both structural
and statistical parameters.
\itemize{
  \item Determining the Standard Errors (SE) of population parameters.
  \item Calculating the Relative Standard Errors (RSE\%).
  \item Evaluating and optimizing designs for clinical trials.
}
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
