% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response.R
\name{PopulationFIMEvaluateVariance}
\alias{PopulationFIMEvaluateVariance}
\title{Evaluate the Variance of a Population FIM}
\usage{
PopulationFIMEvaluateVariance(
  object,
  equations,
  model_parameters,
  administrations,
  sampling_times,
  df_total,
  errorVariances,
  sigmaDerivatives
)
}
\arguments{
\item{object}{A \code{Response} object.}

\item{equations}{An object of class \code{Response} containing the name of the reponse and the equation of the model error.}

\item{model_parameters}{An object of class \code{ModelParameters} containing the values and the distributions of the model parameters.}

\item{administrations}{An object of class \code{Administration} containing the parametrization for the administration of the model.}

\item{sampling_times}{An object of class \code{SamplingTimes} containing the parametrization for the sampling times of the model.}

\item{df_total}{parameter df_total}

\item{errorVariances}{parameter errorVariances}

\item{sigmaDerivatives}{parameter sigmaDerivatives}
}
\value{
A list giving  \code{VDist} and \code{MF_var}.
}
\description{
Evaluate the Variance of a Population FIM
}
