% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/FedorovWynnAlgorithm.R,
%   R/MultiplicativeAlgorithm.R
\name{PrepareFIMs}
\alias{PrepareFIMs}
\alias{PrepareFIMs,FedorovWynnAlgorithm-method}
\alias{PrepareFIMs,MultiplicativeAlgorithm-method}
\title{Prepare the FIMs for the optimization.}
\usage{
PrepareFIMs(object, statistical_model, cond_init, constraint, typeFim)

\S4method{PrepareFIMs}{FedorovWynnAlgorithm}(object, statistical_model, cond_init, constraint, typeFim)

\S4method{PrepareFIMs}{MultiplicativeAlgorithm}(object, statistical_model, cond_init, constraint, typeFim)
}
\arguments{
\item{object}{A \code{MultiplicativeAlgorithm} object.}

\item{statistical_model}{A \code{StatisticalModel} object.}

\item{cond_init}{: cond_init}

\item{constraint}{: A \code{Constraint} object.}

\item{typeFim}{: A character string giving the r=type of FIM : Population, Individual or Bayesian.}
}
\value{
A list \code{result} of all the FIMs.

A list \code{FIMs} of the Fisher Informations matrices.

A list \code{FIMs} of the Fisher Informations Matrices.
}
\description{
Prepare the FIMs for the optimization.

Prepare the Fisher Informations matrices.

Prepare the Fisher Informations Matrices.
}
