% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{processArmEvaluationResults}
\alias{processArmEvaluationResults}
\title{Process arm evaluation for response plots}
\usage{
processArmEvaluationResults(arm, model, fim, designName, plotOptions, ...)
}
\arguments{
\item{arm}{An object of class \code{Arm}.}

\item{model}{An object of class \code{Model}.}

\item{fim}{An object of class \code{Fim}.}

\item{designName}{A string giving the name of the design.}

\item{plotOptions}{A list with \code{unitTime} and \code{unitOutcomes}.}

\item{...}{Additional arguments}
}
\value{
A named list of \code{ggplot} objects per design and arm.
}
\description{
Evaluates the model on a dense time grid and generates response plots
for a given arm, overlaying the design sampling points.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
