% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportional.R
\docType{class}
\name{Proportional-class}
\alias{Proportional-class}
\alias{Proportional}
\title{Class "Proportional"}
\description{
The Class "Proportional" defines the the residual error variance according
to the formula  g(sigma_inter, sigma_slope, c_error, f(x, theta)) =  sigma_slope*f(x,theta).
}
\section{Objects from the Class \linkS4class{Proportional}}{

Objects are typically created by calls to \code{Proportional} and contain the following slots
that are inherited from the class \linkS4class{Combined1}:
}

\section{Slots for the \code{Proportional} objects}{

\describe{
\item{\code{.Object}:}{An object of the Class \code{Proportional}}
\item{\code{sigma_inter}:}{A numeric value giving the sigma inter of the error model}
\item{\code{sigma_slope}:}{A numeric value giving the sigma slope of the error model}
}
}

