% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOAlgorithm.R
\docType{class}
\name{PSOAlgorithm-class}
\alias{PSOAlgorithm-class}
\alias{PSOAlgorithm}
\title{Class "PSOAlgorithm"}
\description{
The Class "PSOAlgorithm" implements the PSO algortihm : Particle Swarm Optimization
}
\section{Objects from the class \code{PSOAlgorithm}}{

Objects form the xlass \code{PSOAlgorithm} can be created by calls of the form \code{PSOAlgorithm(...)} where
(...) are the parameters for the \code{PSOAlgorithm} objects.
}

\section{Slots for \code{PSOAlgorithm} objects}{

 \describe{
   \item{\code{maxIteration}:}{A numeric giving the maximum of iterations.}
   \item{\code{populationSize}:}{A numeric giving the mpopulation size.}
   \item{\code{intertiaWeight}:}{A numeric giving the inertial weight.}
   \item{\code{personalLearningCoefficient}:}{A numeric giving the personal learning coefficient.}
   \item{\code{globalLearningCoefficient}:}{A numeric giving the global learning coefficient.}
   \item{\code{resultsPSO}:}{A list giving the iteration and the results when a new best criteria is found.}
 }
}

