\name{readPFIM}
\alias{readPFIM}
\title{Read a PFIM3.2.2 model specification}
\usage{readPFIM(dir, model_created=FALSE)}
\description{Read a PFIM3.2.2 model specification and create the internal representation}
\value{a PFIMmod object}
\arguments{\item{dir}{directory containing files named 'stdin.r' and 'model.r'}
\item{model_created}{logical Is there another file called 'model_created.r'?}}
