% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-utils.R
\name{rebuildEvalModel}
\alias{rebuildEvalModel}
\title{Cached project model; arm administration applied later.}
\usage{
rebuildEvalModel(pfimproject, finiteDifference = FALSE)
}
\value{
The updated project object with rebuilt evaluation model.
}
\description{
Keys on model signature + plain/FD mode.
}
\keyword{internal}
