% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEquations.R
\name{remplaceDose}
\alias{remplaceDose}
\title{Function to remplace a dose.}
\usage{
remplaceDose(ex, progression, all)
}
\arguments{
\item{ex}{parameter ex}

\item{progression}{parameter progression}

\item{all}{parameter all}
}
\value{
expression of the equation with dose expression.
}
\description{
Function to remplace a dose.
}
