% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{replaceVariablesLibraryOfModels}
\alias{replaceVariablesLibraryOfModels}
\title{: replace variable in the LibraryOfModels}
\usage{
replaceVariablesLibraryOfModels(text, old, new)
}
\arguments{
\item{text}{the text}

\item{old}{old string}

\item{new}{new string}
}
\value{
text with new string
}
\description{
A utility function designed to rename or replace variable names within
model strings (e.g., library equations). It ensures safe replacement
by protecting reserved mathematical terms and keywords.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
