% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{Report}
\alias{Report}
\title{Report: generate the HTML optimization report}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation or Optimization.}

\item{outputPath}{A string giving the path where the output are saved.}

\item{outputFile}{A string giving the name of the output file.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
The html report of the design evaluation or optimization.
}
\description{
Report: generate the HTML optimization report

Report: generate the report.
}
