% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportAndPlots.R
\docType{class}
\name{ReportAndPlots-class}
\alias{ReportAndPlots-class}
\alias{ReportAndPlots}
\title{Class "ReportAndPlots"}
\description{
The class \code{ReportAndPlots} defines the htmal reports for the evaluation and the optimization.
}
\section{Objects from the class}{
 \code{ReportAndPlots} objects are typically created by calls to \code{{ReportAndPlots}} and contain the following slots:

\describe{
\item{\code{object}:}{An object from the class \code{ReportAndPlots}}
}
}

