% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{reportPFIMProject}
\alias{reportPFIMProject}
\title{Generate a htmal report html.}
\usage{
reportPFIMProject(object, ...)
}
\arguments{
\item{object}{\code{PFIMProject} object.}

\item{...}{A list giving options for the report.}
}
\value{
an html givin a report of the project for evaluation or optimization
}
\description{
Generate a htmal report html.
}
