% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{reportTablesDesign}
\alias{reportTablesDesign}
\alias{reportTablesDesign,Design-method}
\title{reportTablesDesign}
\usage{
reportTablesDesign(object)

\S4method{reportTablesDesign}{Design}(object)
}
\arguments{
\item{object}{An object \code{design} from the class \linkS4class{Design}.}
}
\value{
A table of the design parameters for the report.
}
\description{
Generate table for the report.
}
