% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{reportTablesModelError}
\alias{reportTablesModelError}
\alias{reportTablesModelError,Model-method}
\title{Generate the tables for model errors for the evaluation report.}
\usage{
reportTablesModelError(object)

\S4method{reportTablesModelError}{Model}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
A kable table for the evaluation report.
}
\description{
Generate the tables for model errors for the evaluation report.
}
