% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{reportTablesSamplingConstraints}
\alias{reportTablesSamplingConstraints}
\title{Generate table for the report.}
\usage{
reportTablesSamplingConstraints(object)
}
\arguments{
\item{object}{An object \code{design} from the class \linkS4class{Design}.}
}
\value{
A table of the sampling constraints parameters for the report.
}
\description{
Generate table for the report.
}
