% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{resizeFisherMatrix}
\alias{resizeFisherMatrix}
\alias{resizeFisherMatrix,ANY-method}
\title{Resize the fisher Matrix from a vector to a matrix.}
\usage{
resizeFisherMatrix(nbOfDimensions, fisherMatrix)

\S4method{resizeFisherMatrix}{ANY}(nbOfDimensions, fisherMatrix)
}
\arguments{
\item{nbOfDimensions}{: a numeric for the dimensions of the fisher matrix.}

\item{fisherMatrix}{: a vector that contain the low triangular Fisher matrix + its main diagonal.}
}
\value{
The Fisher matrix of size nbOfDimensions*nbOfDimensions
}
\description{
Resize the fisher Matrix from a vector to a matrix.
}
