% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response.R
\docType{class}
\name{Response-class}
\alias{Response-class}
\alias{Response}
\title{Class "Response"}
\description{
Class \code{Response} represents a structural model.
}
\section{Objects from the class}{

\code{Response} objects are typically created by calls to \code{Response} and contain the following slots
model_error = g(sigma_inter, sigma_slope , f(x, theta)), this part is considered in class \linkS4class{ModelError}.
There are different possibilities to calculate g.
}

\section{Slots for \code{Response} objects}{

\describe{
\item{\code{name}:}{A character string giving the name for model error.}
\item{\code{model_error}:}{An object \code{model_error} from the Class \linkS4class{ModelError}}.
}
}

