% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\docType{class}
\name{SamplingConstraint-class}
\alias{SamplingConstraint-class}
\alias{SamplingConstraint}
\title{Class "SamplingConstraint"}
\description{
Class "SamplingConstraint" storing information concerning sampling constraint.
}
\section{Objects from the class}{
 \code{SamplingConstraint} objects are typically created by calls to \code{SamplingConstraint} and contain the following slots:
}

\section{Slots for \code{SamplingConstraint}  objects}{

\describe{
\item{\code{response}:}{A character string for the name of the response of the model.}
\item{\code{numberOptimisability}:}{A boolean that gives TRUE for optimizing the number of times and FALSE for fixing the number of times.}
\item{\code{numberOfSamplingTimes}:}{A vector of the number of sampling times.}
\item{\code{fixedTimes}:}{A vector of the number of fixed times.}
\item{\code{continuousSamplingTimes}:}{A list of the continuous sampling times.}
\item{\code{discretSamplingTimes}:}{A list of the discrete sampling times.}
\item{\code{min_delay}:}{A numeric giving the minimal interval in the sampling times.}
}
}

