% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\docType{class}
\name{SamplingTimeConstraints-class}
\alias{SamplingTimeConstraints-class}
\alias{SamplingTimeConstraints}
\title{Class "SamplingTimeConstraints"}
\description{
The class "SamplingTimeConstraints" implements the constraints for the sampling times.
}
\section{Objects from the class \code{SamplingTimeConstraints}}{

Objects form the class \code{SamplingTimeConstraints} can be created by calls of the form \code{SamplingTimeConstraints(...)} where
(...) are the parameters for the \code{SamplingTimeConstraints} objects.
}

\section{Slots for \code{SamplingTimeConstraints} objects}{

 \describe{
   \item{\code{outcome}:}{A string giving the outcome.}
   \item{\code{initialSamplings}:}{A vector giving the sampling times.}
   \item{\code{fixedTimes}:}{A vector giving the fixed  sampling times.}
   \item{\code{numberOfsamplingsOptimisable}:}{A vector giving the sampling times to be optimized.}
  \item{\code{samplingsWindows}:}{A list giving the windows for the sampling times.}
   \item{\code{numberOfTimesByWindows}:}{A vector giving the number of sampling times by windows.}
   \item{\code{minSampling}:}{A numeric giving the minimal sampling times.}
 }
}

