% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\docType{class}
\name{SamplingTimes-class}
\alias{SamplingTimes-class}
\alias{SamplingTimes}
\alias{[,Sampling-method}
\alias{[<-,Sampling-method}
\title{Class "SamplingTimes"}
\description{
Class "SamplingTimes" stores information concerning sampling times.
}
\section{Objects from the Class}{

Objects form the Class \code{SamplingTimes} can be created by calls of the form \code{SamplingTimes(...)} where
(...) are the parameters for the \code{SamplingTimes} objects.
}

\section{Slots for the \code{SamplingTimes} objects}{

\describe{
\item{\code{outcome}:}{A character string giving either a compartment name or number (character or integer, TBD with model) (nombre de reponses "1", "2").}
\item{\code{sample_time}:}{A list of discrete vectors giving the times when sampling design is performed. }
\item{\code{initialTime}:}{A numeric giving the initial time of the vecotr of sampling times.}
}
}

