% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfusionODEquations.R
\name{scaleResponsesEvaluationODEInfusion}
\alias{scaleResponsesEvaluationODEInfusion}
\title{function to adjust Responses with variables values.}
\usage{
scaleResponsesEvaluationODEInfusion(
  out_variable,
  modelParameters,
  variablesNames,
  responseNames,
  inputsModel
)
}
\arguments{
\item{out_variable}{parameter out_variable.}

\item{modelParameters}{parameter modelParameters.}

\item{variablesNames}{parameter variablesNames.}

\item{responseNames}{parameter responseNames.}

\item{inputsModel}{parameter inputsModel.}
}
\value{
A dataframe giving the evaluated responses adjusted with variables values.
}
\description{
function to adjust Responses with variables values.
}
