% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{setAdministrations}
\alias{setAdministrations}
\alias{setAdministrations,Arm-method}
\title{setAdministrations}
\usage{
setAdministrations(object, administrations)

\S4method{setAdministrations}{Arm}(object, administrations)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{administrations}{A list \code{administrations} of objects from the class \code{Administration} class giving
the parameters of the administration for the object \code{Arm}.}
}
\value{
The object \code{Arm} with the list \code{administrations} of objects from the class \code{Administration} class giving
the parameters of the administration for the object \code{Arm}.
}
\description{
Set all the administration for an arm.
}
