% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{setAllowedDose<-}
\alias{setAllowedDose<-}
\alias{setAllowedDose}
\title{Set the constraints on allowed dose.}
\usage{
setAllowedDose(object) <- value
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}

\item{value}{A numeric value for the new dose value.}
}
\value{
The \code{Administration} object with the new constraints on the allowed dose.
}
\description{
Set the constraints on allowed dose.
}
