% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{setAllowedTime<-}
\alias{setAllowedTime<-}
\title{Set the constraints on allowed times.}
\usage{
setAllowedTime(object) <- value
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}

\item{value}{A vector of the numeric values for the new constraints on allowed times.}
}
\value{
The \code{Administration} object with the new constraints on allowed times.
}
\description{
Set the constraints on allowed times.
}
