% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{setAllowedTinf<-}
\alias{setAllowedTinf<-}
\title{Set the constraints on Tinf.}
\usage{
setAllowedTinf(object) <- value
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}

\item{value}{A numeric value for the new constraints on Tinf.}
}
\value{
The \code{Administration} object with the new constraints on Tinf.
}
\description{
Set the constraints on Tinf.
}
