% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{setAmountDose}
\alias{setAmountDose}
\title{Set the amount of dose}
\usage{
setAmountDose(object, value)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}

\item{value}{A numeric value of the amount of dose.}
}
\value{
The numeric \code{amount_dose} giving the new value of the amount of dose.
}
\description{
Set the amount of dose
}
