% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setAmountOfArms}
\alias{setAmountOfArms}
\title{Set the amount of arms in a Design.}
\usage{
setAmountOfArms(object, value)
}
\arguments{
\item{object}{A \code{Design} object.}

\item{value}{A numeric giving the new value of the amount of arms in the design.}
}
\value{
The \code{Design} object with the new vamue of amount of arms.
}
\description{
Set the amount of arms in a Design.
}
