% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\name{setAmountOfArmsAim}
\alias{setAmountOfArmsAim}
\title{Set amount of arms in a  \code{DesignConstraint} object for the case we aim to obtain a fixed amount of arms as result.}
\usage{
setAmountOfArmsAim(object, value)
}
\arguments{
\item{object}{A \code{DesignConstraint} object.}

\item{value}{A numeric.}
}
\value{
A numeric \code{amountOfArm} giving the amount of arms for the case we aim to obtain a fixed amount of arms as result.
}
\description{
Set amount of arms in a  \code{DesignConstraint} object for the case we aim to obtain a fixed amount of arms as result.
}
