% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setArms}
\alias{setArms}
\title{Set the arms of a design.}
\usage{
setArms(object, value)
}
\arguments{
\item{object}{A \code{Design} object.}

\item{value}{A \code{Arm} object.}
}
\value{
The design \code{Design} with the new arm.
}
\description{
Set the arms of a design.
}
