% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{setArmSize}
\alias{setArmSize}
\title{Set the size of an arm.}
\usage{
setArmSize(object, value)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{value}{A numeric giving the new size of the object \code{Arm}.}
}
\value{
The object \code{Arm} object with its new size.
}
\description{
Set the size of an arm.
}
