% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{setCError<-}
\alias{setCError<-}
\title{Set the CError of a \code{ModelError} object.}
\usage{
setCError(object) <- value
}
\arguments{
\item{object}{An \code{ModelError} object.}

\item{value}{The value for CError.}
}
\value{
The \code{ModelError} object with the new value of the CError.
}
\description{
Set the CError of a \code{ModelError} object.
}
