% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{setConstraint}
\alias{setConstraint}
\title{Set the constraint to the PFIMProject projet.}
\usage{
setConstraint(object, constraint)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}

\item{constraint}{The constraint to set}
}
\value{
The \code{PFIMProject} object with the \code{constraint}.
}
\description{
Set the constraint to the PFIMProject projet.
}
