% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{setContent}
\alias{setContent}
\alias{setContent,LibraryOfModels-method}
\title{Set content of a library of models.}
\usage{
setContent(object, content)

\S4method{setContent}{LibraryOfModels}(object, content)
}
\arguments{
\item{object}{An object from the class \linkS4class{LibraryOfModels}.}

\item{content}{A list giving the content of the library of models.}
}
\value{
The library of models with the updated content.
}
\description{
Set content of a library of models.
}
