% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{setDataForArmEvaluation}
\alias{setDataForArmEvaluation}
\alias{setDataForArmEvaluation,Arm-method}
\title{setDataForArmEvaluation}
\usage{
setDataForArmEvaluation(object, data)

\S4method{setDataForArmEvaluation}{Arm}(object, data)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{data}{A list containing the data for arm evaluation}
}
\value{
Set the list containing the data for arm evaluation.
}
\description{
setDataForArmEvaluation
}
