% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R, R/ModelAnalytic.R,
%   R/ModelAnalyticSteadyState.R, R/ModelAnalyticInfusion.R,
%   R/ModelAnalyticInfusionSteadyState.R, R/ModelODEBolus.R,
%   R/ModelODEDoseInEquations.R, R/ModelODEDoseNotInEquations.R,
%   R/ModelODEInfusionDoseInEquations.R
\name{setDataForModelEvaluation}
\alias{setDataForModelEvaluation}
\alias{setDataForModelEvaluation,ModelAnalytic-method}
\alias{setDataForModelEvaluation,ModelAnalyticSteadyState-method}
\alias{setDataForModelEvaluation,ModelAnalyticInfusion-method}
\alias{setDataForModelEvaluation,ModelAnalyticInfusionSteadyState-method}
\alias{setDataForModelEvaluation,ModelODEBolus-method}
\alias{setDataForModelEvaluation,ModelODEDoseInEquations-method}
\alias{setDataForModelEvaluation,ModelODEDoseNotInEquations-method}
\alias{setDataForModelEvaluation,ModelODEInfusionDoseInEquations-method}
\title{Generate the table of dose, time dose etc. for model evaluation}
\usage{
setDataForModelEvaluation(object, arm)

\S4method{setDataForModelEvaluation}{ModelAnalytic}(object, arm)

\S4method{setDataForModelEvaluation}{ModelAnalyticSteadyState}(object, arm)

\S4method{setDataForModelEvaluation}{ModelAnalyticInfusion}(object, arm)

\S4method{setDataForModelEvaluation}{ModelAnalyticInfusionSteadyState}(object, arm)

\S4method{setDataForModelEvaluation}{ModelODEBolus}(object, arm)

\S4method{setDataForModelEvaluation}{ModelODEDoseInEquations}(object, arm)

\S4method{setDataForModelEvaluation}{ModelODEDoseNotInEquations}(object, arm)

\S4method{setDataForModelEvaluation}{ModelODEInfusionDoseInEquations}(object, arm)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{arm}{An object from the class \linkS4class{Arm}.}
}
\value{
Return a dataframe with all the data for model evaluation
}
\description{
Generate the table of dose, time dose etc. for model evaluation
}
