% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{setDelta}
\alias{setDelta}
\title{Set the \code{delta} parameters for the Multiplicative algorithm.}
\usage{
setDelta(object, values)
}
\arguments{
\item{object}{\code{MultiplicativeAlgorithm} object.}

\item{values}{values}
}
\value{
The \code{MultiplicativeAlgorithm} object with the new value of \code{delta}.
}
\description{
Set the \code{delta} parameters for the Multiplicative algorithm.
}
