% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{setDerivatives}
\alias{setDerivatives}
\title{Set the derivatives of the model error equation.}
\usage{
setDerivatives(object, derivatives)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelError}.}

\item{derivatives}{The derivatives of the model error equation.}
}
\value{
The model error with the updated model error equation.
}
\description{
Set the derivatives of the model error equation.
}
