% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{setDesign}
\alias{setDesign}
\title{Set the design of \code{PFIMProject} object.}
\usage{
setDesign(object, value)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}

\item{value}{A \code{Design} object.}
}
\value{
The \code{PFIMProject} object with the new \code{Designs}.
}
\description{
Set the design of \code{PFIMProject} object.
}
