% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{setDesigns}
\alias{setDesigns}
\alias{setDesigns,Optimization-method}
\title{Set the designs.}
\usage{
setDesigns(object, designs)

\S4method{setDesigns}{Optimization}(object, designs)
}
\arguments{
\item{object}{An object from the class \linkS4class{Optimization}.}

\item{designs}{A list of objects from the class \linkS4class{Design}.}
}
\value{
The object with the new designs.
}
\description{
Set the designs.
}
