% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteConstraint.R
\name{setDiscret<-}
\alias{setDiscret<-}
\title{Set the possible values for a \code{DiscreteConstraint} object.}
\usage{
setDiscret(object) <- value
}
\arguments{
\item{object}{A \code{DiscreteConstraint} object.}

\item{value}{Value for the discrete constraint in the  \code{DiscreteConstraint} object.}
}
\value{
The \code{DiscreteConstraint} object with the set of new values.
}
\description{
Set the possible values for a \code{DiscreteConstraint} object.
}
