% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{setDistribution}
\alias{setDistribution}
\alias{setDistribution,ModelParameter-method}
\title{Set the distribution.}
\usage{
setDistribution(object, distribution)

\S4method{setDistribution}{ModelParameter}(object, distribution)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelParameter}.}

\item{distribution}{An object from the class \linkS4class{Distribution}.}
}
\value{
The model parameter with the updated distribution.
}
\description{
Set the distribution.
}
