% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfusion.R
\name{setEquationsAfterInfusion}
\alias{setEquationsAfterInfusion}
\alias{setEquationsAfterInfusion,Model-method}
\title{Set the equations after infusion.}
\usage{
setEquationsAfterInfusion(object, equations)

\S4method{setEquationsAfterInfusion}{Model}(object, equations)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{equations}{A list giving the equations after the infusion.}
}
\value{
The model with the updated equations after the infusion.
}
\description{
Set the equations after infusion.
}
