% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfusion.R
\name{setEquationsDuringInfusion}
\alias{setEquationsDuringInfusion}
\alias{setEquationsDuringInfusion,Model-method}
\title{Set the equations during infusion.}
\usage{
setEquationsDuringInfusion(object, equations)

\S4method{setEquationsDuringInfusion}{Model}(object, equations)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{equations}{A list giving the equations during the infusion.}
}
\value{
The model with the updated equations during the infusion.
}
\description{
Set the equations during infusion.
}
