% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{setEvaluationFIMResults}
\alias{setEvaluationFIMResults}
\alias{setEvaluationFIMResults,Optimization-method}
\title{Set the evaluation results.}
\usage{
setEvaluationFIMResults(object, value)

\S4method{setEvaluationFIMResults}{Optimization}(object, value)
}
\arguments{
\item{object}{An object from the class \linkS4class{Optimization}.}

\item{value}{An object from the class \linkS4class{Evaluation} giving the evaluation results.}
}
\value{
The object with the updated object from the class \linkS4class{Evaluation}.
}
\description{
Set the evaluation results.
}
