% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{setEvaluationInitialDesignResults}
\alias{setEvaluationInitialDesignResults}
\title{Set the evaluation results of the initial design.}
\usage{
setEvaluationInitialDesignResults(object, value)
}
\arguments{
\item{object}{An object from the class \linkS4class{Optimization}.}

\item{value}{An object from the class \linkS4class{Evaluation} giving the evaluation results of the initial design.}
}
\value{
The object with the updated object from the class \linkS4class{Evaluation}.
}
\description{
Set the evaluation results of the initial design.
}
