% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{setFimTypeToString}
\alias{setFimTypeToString}
\alias{setFimTypeToString,Fim-method}
\title{Convert the type of the object fim to a string.}
\usage{
setFimTypeToString(object)

\S4method{setFimTypeToString}{Fim}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
The type of the object fim convert as a string.
}
\description{
Convert the type of the object fim to a string.
}
