% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{setFisherMatrix}
\alias{setFisherMatrix}
\title{Set the FIM.}
\usage{
setFisherMatrix(object, value)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{value}{A matrix giving the FIM.}
}
\value{
The object from the class \linkS4class{Fim} with the FIM updated.
}
\description{
Set the FIM.
}
