% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{setFixedMu}
\alias{setFixedMu}
\alias{setFixedMu,ModelParameter-method}
\title{Set the mu as fixed or not.}
\usage{
setFixedMu(object, value)

\S4method{setFixedMu}{ModelParameter}(object, value)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelParameter}.}

\item{value}{A Boolean if fixed or not.}
}
\value{
The mode parameter with the the mu updated as fixed or not.
}
\description{
Set the mu as fixed or not.
}
