% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{setFixedOmega}
\alias{setFixedOmega}
\alias{setFixedOmega,ModelParameter-method}
\title{Set the omega as fixed of not.}
\usage{
setFixedOmega(object, value)

\S4method{setFixedOmega}{ModelParameter}(object, value)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelParameter}.}

\item{value}{A Boolean fixed or not.}
}
\value{
The model parameter with the omega updated as fixed or not.
}
\description{
Set the omega as fixed of not.
}
