% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R, R/Model.R
\name{setInitialConditions}
\alias{setInitialConditions}
\title{Set the initial condition for the evaluation of an ode model.}
\usage{
setInitialConditions(object, initialConditions)

setInitialConditions(object, initialConditions)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{initialConditions}{A list giving the initial conditions.}
}
\value{
An object \code{Arm} from the class \linkS4class{Arm} with the new initial conditions \code{initialConditions}.

The model with the updated initial conditions.
}
\description{
Set the initial condition for the evaluation of an ode model.

Set the initial conditions of a ode model.
}
