% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{setIteration}
\alias{setIteration}
\title{Set the number of iterations for the multiplicative algorithm.}
\usage{
setIteration(object, values)
}
\arguments{
\item{object}{\code{MultiplicativeAlgorithm} object.}

\item{values}{A numeric.}
}
\value{
The \code{MultiplicativeAlgorithm} object with the new values of the number of iterations.
}
\description{
Set the number of iterations for the multiplicative algorithm.
}
