% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationAlgorithm.R
\name{setIterationAndCriteria}
\alias{setIterationAndCriteria}
\alias{setIterationAndCriteria,OptimizationAlgorithm-method}
\title{Set the iteration with the convergence criteria.}
\usage{
setIterationAndCriteria(object, value)

\S4method{setIterationAndCriteria}{OptimizationAlgorithm}(object, value)
}
\arguments{
\item{object}{An object from the class \linkS4class{OptimizationAlgorithm}.}

\item{value}{A dataframe giving the iteration with the convergence criteria.}
}
\value{
A dataframe giving the iteration with the convergence criteria.
}
\description{
Set the iteration with the convergence criteria.
}
