% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{setMfisher<-}
\alias{setMfisher<-}
\title{Set a matrix value for the Fisher Information Matrix.}
\usage{
setMfisher(object) <- value
}
\arguments{
\item{object}{A \code{Fim} object.}

\item{value}{A matrix of numerical values.}
}
\value{
The \code{Fim} object with the Fisher Information Matrix with the new values.
}
\description{
Set a matrix value for the Fisher Information Matrix.
}
