% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response.R
\name{setModelError<-}
\alias{setModelError<-}
\title{Set the model error.}
\usage{
setModelError(object) <- value
}
\arguments{
\item{object}{A \code{Response} object.}

\item{value}{The new value for the model error.}
}
\value{
The \code{Response} object with the new value for the model error.
}
\description{
Set the model error.
}
