% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setModelFromLibrary}
\alias{setModelFromLibrary}
\alias{setModelFromLibrary,Model-method}
\title{Set  a model from the library of model}
\usage{
setModelFromLibrary(object, modelFromLibrary)

\S4method{setModelFromLibrary}{Model}(object, modelFromLibrary)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{modelFromLibrary}{An object from the class \linkS4class{Model}.}
}
\value{
The model with the updated model from library of models.
}
\description{
Set  a model from the library of model
}
