% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{setMu}
\alias{setMu}
\title{Set the value of the fixed effect mu of an object.}
\usage{
setMu(object, value)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{value}{The value of the fixed effect mu.}
}
\value{
The object with the updated fixed effect mu.
}
\description{
Set the value of the fixed effect mu of an object.
}
