% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{setName}
\alias{setName}
\title{Set the name of an object.}
\usage{
setName(object, name)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{name}{A string giving the name of the object.}
}
\value{
The object with the updated name.
}
\description{
Set the name of an object.
}
