% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setNameDesign}
\alias{setNameDesign}
\title{Set the name of the design.}
\usage{
setNameDesign(object, name)
}
\arguments{
\item{object}{\code{Design} object.}

\item{name}{A character string \code{name} giving the new name of design.}
}
\value{
The \code{Design} object with its new name.
}
\description{
Set the name of the design.
}
