% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{setNamePFIMProject}
\alias{setNamePFIMProject}
\title{Set the name of a PFIMProject projet.}
\usage{
setNamePFIMProject(object, value)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}

\item{value}{A character string giving the new name of the PFIMProject project.}
}
\value{
The \code{PFIMProject} object with a new name.
}
\description{
Set the name of a PFIMProject projet.
}
