% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setNumberOfArms}
\alias{setNumberOfArms}
\title{Set the number of arms in a design.}
\usage{
setNumberOfArms(object, numberOfArms)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}

\item{numberOfArms}{A numeric \code{numberOfArms} giving the new number of arms in the design.}
}
\value{
An object \code{Design} with the \code{numberOfArms} updated.
}
\description{
Set the number of arms in a design.
}
