% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setNumberSamples<-}
\alias{setNumberSamples<-}
\title{Set the number of Sample in a Design.}
\usage{
setNumberSamples(object) <- value
}
\arguments{
\item{object}{A \code{Design} object.}

\item{value}{A numeric giving the new value of samples.}
}
\value{
The \code{Design} object with the new number of samples.
}
\description{
Set the number of Sample in a Design.
}
