% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{setOmega}
\alias{setOmega}
\title{Set the Omega matrix.}
\usage{
setOmega(object, omega)
}
\arguments{
\item{object}{A \code{Fim} object.}

\item{omega}{A matrix omega giving the new values of the variances.}
}
\value{
The \code{Fim} object with the Omega matrix with the new values.
}
\description{
Set the Omega matrix.
}
