% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{setOptimalDesign<-}
\alias{setOptimalDesign<-}
\alias{setOptimalDesign}
\title{Set an optimal design.}
\usage{
setOptimalDesign(object) <- value
}
\arguments{
\item{object}{A \code{PFIM} object.}

\item{value}{A \code{Design} object.}
}
\value{
The \code{PFIM} object with the optimal design.
}
\description{
Set an optimal design.
}
