% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationAlgorithm.R
\name{setOptimalDesign}
\alias{setOptimalDesign}
\alias{setOptimalDesign,OptimizationAlgorithm-method}
\title{Set the optimal design.}
\usage{
setOptimalDesign(object, optimalDesign)

\S4method{setOptimalDesign}{OptimizationAlgorithm}(object, optimalDesign)
}
\arguments{
\item{object}{An object from the class \linkS4class{OptimizationAlgorithm}.}

\item{optimalDesign}{An object from the class \linkS4class{Design}.}
}
\value{
The object with the updated optimal design.
}
\description{
Set the optimal design.
}
