% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{setOptimalWeights}
\alias{setOptimalWeights}
\alias{setOptimalWeights,MultiplicativeAlgorithm-method}
\title{Set the optimal weights.}
\usage{
setOptimalWeights(object, optimalWeights)

\S4method{setOptimalWeights}{MultiplicativeAlgorithm}(object, optimalWeights)
}
\arguments{
\item{object}{An object from the class \linkS4class{MultiplicativeAlgorithm}.}

\item{optimalWeights}{A vector giving the optimal weights.}
}
\value{
The object with the updated optimal weights.
}
\description{
Set the optimal weights.
}
