% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{setOptimizationResults}
\alias{setOptimizationResults}
\alias{setOptimizationResults,Optimization-method}
\title{Set the optimization results.}
\usage{
setOptimizationResults(object, value)

\S4method{setOptimizationResults}{Optimization}(object, value)
}
\arguments{
\item{object}{An object from the class \linkS4class{Optimization}.}

\item{value}{An object from the class \linkS4class{OptimizationAlgorithm} giving the optimization results.}
}
\value{
The object with the updated object from the class \linkS4class{OptimizationAlgorithm}.
}
\description{
Set the optimization results.
}
