% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{setOutcome}
\alias{setOutcome}
\title{Set the outcome of an object.}
\usage{
setOutcome(object, outcome)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{outcome}{A string defined the outcome.}
}
\value{
A string giving the updated outcome of the object.
}
\description{
Set the outcome of an object.
}
