% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setOutcomes}
\alias{setOutcomes}
\alias{setOutcomes,Model-method}
\title{Set the outcomes of a model.}
\usage{
setOutcomes(object, outcomes)

\S4method{setOutcomes}{Model}(object, outcomes)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{outcomes}{A list giving the outcomes of the model.}
}
\value{
The model with the updated outcomes.
}
\description{
Set the outcomes of a model.
}
