% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setOutcomesEvaluation}
\alias{setOutcomesEvaluation}
\title{Set the results of the evaluation of the outcomes.}
\usage{
setOutcomesEvaluation(object, outcomesEvaluation)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}

\item{outcomesEvaluation}{A list containing the evaluation of the outcomes.}
}
\value{
An object \code{Design} with the list \code{outcomesEvaluation} updated.
}
\description{
Set the results of the evaluation of the outcomes.
}
