% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setOutcomesForEvaluation}
\alias{setOutcomesForEvaluation}
\title{Set the outcomes of a model used for the evaluation (is scales outcomes).}
\usage{
setOutcomesForEvaluation(object, outcomes)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{outcomes}{A list giving the outcomes of a model used for the evaluation (is scales outcomes).}
}
\value{
The model with the updated outcomes for the evaluation.
}
\description{
Set the outcomes of a model used for the evaluation (is scales outcomes).
}
