% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setOutcomesGradient}
\alias{setOutcomesGradient}
\alias{setOutcomesGradient,Design-method}
\title{setOutcomesGradient}
\usage{
setOutcomesGradient(object, outcomesGradient)

\S4method{setOutcomesGradient}{Design}(object, outcomesGradient)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}

\item{outcomesGradient}{A list containing the evaluation of the outcome gradients.}
}
\value{
An object \code{Design} with the list \code{outcomesGradient} updated.
}
\description{
Set the results of the evaluation of the outcomes.
}
