% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Distribution.R, R/Model.R,
%   R/FedorovWynnAlgorithm.R, R/MultiplicativeAlgorithm.R, R/PGBOAlgorithm.R,
%   R/PSOAlgorithm.R, R/SimplexAlgorithm.R
\name{setParameters}
\alias{setParameters}
\alias{setParameters,Distribution-method}
\alias{setParameters,Model-method}
\alias{setParameters,FedorovWynnAlgorithm-method}
\alias{setParameters,MultiplicativeAlgorithm-method}
\alias{setParameters,PGBOAlgorithm-method}
\alias{setParameters,PSOAlgorithm-method}
\alias{setParameters,SimplexAlgorithm-method}
\title{Set the parameters of an object.}
\usage{
setParameters(object, parameters)

\S4method{setParameters}{Distribution}(object, parameters)

\S4method{setParameters}{Model}(object, parameters)

\S4method{setParameters}{FedorovWynnAlgorithm}(object, parameters)

\S4method{setParameters}{MultiplicativeAlgorithm}(object, parameters)

\S4method{setParameters}{PGBOAlgorithm}(object, parameters)

\S4method{setParameters}{PSOAlgorithm}(object, parameters)

\S4method{setParameters}{SimplexAlgorithm}(object, parameters)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{parameters}{A list of parameters.}
}
\value{
The object with the updated list of parameters.
}
\description{
Set the parameters of an object.
}
