% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{setParametersForEvaluateModel}
\alias{setParametersForEvaluateModel}
\title{Set the parameters for the evaluation of the model.}
\usage{
setParametersForEvaluateModel(
  object,
  administrations,
  sampling_times,
  cond_init
)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{administrations}{An \code{Administration} object.}

\item{sampling_times}{A \code{SamplingTimes} object.}

\item{cond_init}{A list for the initial conditions of the \code{StatisticalModel} object.}
}
\value{
A list containing the parameters used for the model evaluation.
}
\description{
Set the parameters for the evaluation of the model.
}
