% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setParametersModel}
\alias{setParametersModel}
\title{Set the parameters of the \code{Model} object.}
\usage{
setParametersModel(object, parameters)
}
\arguments{
\item{object}{A \code{Model} object.}

\item{parameters}{The vector of character string giving the parameters names.}
}
\value{
The \code{Model} object with the  new parameters.
}
\description{
Set the parameters of the \code{Model} object.
}
