% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{setParametersOdeSolver}
\alias{setParametersOdeSolver}
\title{Set parameters for the ode solver}
\usage{
setParametersOdeSolver(object, value)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}

\item{value}{A list giving the values of the parameters.}
}
\value{
The \code{StatisticalModel} object with the new parameters for the ode solver.
}
\description{
Set parameters for the ode solver
}
