% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\name{setPossibleArms}
\alias{setPossibleArms}
\title{Set the possible arms in a Design or the case when lots of arms are defined and aim to optimise amoung several of them.}
\usage{
setPossibleArms(object, design, choice)
}
\arguments{
\item{object}{A \code{DesignConstraint} object.}

\item{design}{A \code{Design} object.}

\item{choice}{A vector of arm's serial number, to form an arm-space}
}
\value{
The \code{DesignConstraint} object with all the possible arms
}
\description{
Set the possible arms in a Design or the case when lots of arms are defined and aim to optimise amoung several of them.
}
