% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuousConstraint.R
\name{setRange<-}
\alias{setRange<-}
\title{Set the range of a \code{ContinuousConstraint} object.}
\usage{
setRange(object) <- value
}
\arguments{
\item{object}{A \code{ContinuousConstraint} object.}

\item{value}{A numeric.}
}
\value{
The \code{ContinuousConstraint} object with the new range.
}
\description{
Set the range of a \code{ContinuousConstraint} object.
}
