% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\name{setSampleTime}
\alias{setSampleTime}
\title{Set the sample time of the response of the \code{SamplingTimes} object.}
\usage{
setSampleTime(object, values)
}
\arguments{
\item{object}{A \code{SamplingTimes} object.}

\item{values}{A vector giving the new values of the sampling times.}
}
\value{
The \code{SamplingTimes} object with the new sample times.
}
\description{
Set the sample time of the response of the \code{SamplingTimes} object.
}
