% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setSamplingConstraintForOptimization}
\alias{setSamplingConstraintForOptimization}
\title{Initialize sampling constraints}
\arguments{
\item{design}{An object \code{Design}.}
}
\value{
The \code{Design} object with updated \code{samplingTimesConstraints}.
}
\description{
Initialize sampling constraints
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
