% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{setSamplings<-}
\alias{setSamplings<-}
\title{Set the sampling times for an arm.}
\usage{
setSamplings(object) <- value
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{value}{The sampling times given by the objects from the class \code{SamplingTimes}.}
}
\value{
The object \code{Arm} with its new sampling times.
}
\description{
Set the sampling times for an arm.
}
