% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\name{setSamplings}
\alias{setSamplings}
\alias{setSamplings,SamplingTimes-method}
\title{Set the sampling times.}
\usage{
setSamplings(object, samplings)

\S4method{setSamplings}{SamplingTimes}(object, samplings)
}
\arguments{
\item{object}{An object from the class \linkS4class{SamplingTimes}.}

\item{samplings}{A vector giving the sampling times.}
}
\value{
The updated sampling times.
}
\description{
Set the sampling times.
}
