% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{setSamplingTime}
\alias{setSamplingTime}
\alias{setSamplingTime,Arm-method}
\title{setSamplingTime}
\usage{
setSamplingTime(object, samplingTime)

\S4method{setSamplingTime}{Arm}(object, samplingTime)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{samplingTime}{An object \code{samplingTime} from the class \linkS4class{SamplingTimes}.}
}
\value{
An object \code{Arm} from the class \linkS4class{Arm} with the new sampling time \code{samplingTime}.
}
\description{
Set the sampling time of an arm.
}
