% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/MultiplicativeAlgorithm.R,
%   R/SimplexAlgorithm.R
\name{setShowProcess}
\alias{setShowProcess}
\alias{setShowProcess,MultiplicativeAlgorithm-method}
\alias{setShowProcess,SimplexAlgorithm-method}
\title{Show the process for the optimization.}
\usage{
setShowProcess(object, ifShow)

\S4method{setShowProcess}{MultiplicativeAlgorithm}(object, ifShow)

\S4method{setShowProcess}{SimplexAlgorithm}(object, ifShow)
}
\arguments{
\item{object}{A \code{SimplexAlgorithm} object.}

\item{ifShow}{A boolean.}
}
\value{
Show process for the optimization.

Shows the process for FIM computing.

Show \code{SimplexAlgorithm} object.
}
\description{
Show the process for the optimization.

Shows the process for FIM computing.

Show the process of optimization.
}
