% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{setShrinkage}
\alias{setShrinkage}
\alias{setShrinkage,BayesianFim-method}
\alias{setShrinkage,IndividualFim-method}
\alias{setShrinkage,PopulationFim-method}
\title{Set the shrinkage.}
\usage{
setShrinkage(object, value)

\S4method{setShrinkage}{BayesianFim}(object, value)

\S4method{setShrinkage}{IndividualFim}(object, value)

\S4method{setShrinkage}{PopulationFim}(object, value)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{value}{A vector giving the shrinkage of the Bayesian fim.}
}
\value{
The object with the updated shrinkage.
}
\description{
Set the shrinkage.
}
