% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{setSigmaInter<-}
\alias{setSigmaInter<-}
\title{Set the value for \code{sigma_inter} of a \code{ModelError} object.}
\usage{
setSigmaInter(object) <- value
}
\arguments{
\item{object}{An \code{ModelError} object.}

\item{value}{The value for \code{sigma_inter}}
}
\value{
The \code{ModelError} object with the new value for the \code{sigma_inter}.
}
\description{
Set the value for \code{sigma_inter} of a \code{ModelError} object.
}
