% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{setSigmaSlope<-}
\alias{setSigmaSlope<-}
\title{Set the value for \code{sigma_slope} of a \code{ModelError} object.}
\usage{
setSigmaSlope(object) <- value
}
\arguments{
\item{object}{An \code{ModelError} object.}

\item{value}{The value for \code{sigma_slope}.}
}
\value{
The \code{ModelError} object with the new value for the \code{sigma_slope}.
}
\description{
Set the value for \code{sigma_slope} of a \code{ModelError} object.
}
