% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{setSigmaSlope}
\alias{setSigmaSlope}
\title{Set the parameter sigma slope.}
\usage{
setSigmaSlope(object, sigmaSlope)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelError}.}

\item{sigmaSlope}{A numeric giving the parameter sigma slope.}
}
\value{
The model error with the updated sigma slope.
}
\description{
Set the parameter sigma slope.
}
