% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Arm.R, R/Design.R
\name{setSize}
\alias{setSize}
\alias{setSize,Arm-method}
\alias{setSize,Design-method}
\title{setSize}
\usage{
setSize(object, size)

setSize(object, size)

\S4method{setSize}{Arm}(object, size)

\S4method{setSize}{Design}(object, size)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{size}{A numeric giving the new size of the object \code{Arm}.}
}
\value{
The object with its size updated.

The object \code{Arm} object with its new size.
}
\description{
Set the size of an object.

Set the size of an arm.
}
