% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{setTau}
\alias{setTau}
\title{Set the frequency \code{tau}.}
\usage{
setTau(object, tau)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}

\item{tau}{A numeric value for the infusion lag tau.}
}
\value{
The object \code{Administration} object with its new value of the infusion lag tau.
}
\description{
Set the frequency \code{tau}.
}
