% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{setTimeDose}
\alias{setTimeDose}
\alias{setTimeDose,Administration-method}
\title{setTimeDose}
\usage{
setTimeDose(object, timeDose)

\S4method{setTimeDose}{Administration}(object, timeDose)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}

\item{timeDose}{A numeric value of the time dose.}
}
\value{
The object \code{Administration} with its new times vector for doses.
}
\description{
Set the times vector when doses are given.
}
