% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{setTinf}
\alias{setTinf}
\title{Set the infusion duration.}
\usage{
setTinf(object, Tinf)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}

\item{Tinf}{A numeric value for the infusion duration Tinf.}
}
\value{
The object \code{Administration} with its new value of the infusion duration Tinf.
}
\description{
Set the infusion duration.
}
