% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConstraint.R
\name{setTotalNumberOfIndividuals}
\alias{setTotalNumberOfIndividuals}
\title{Set the total number of individuals in a \code{DesignConstraint} object.}
\usage{
setTotalNumberOfIndividuals(object, totalNumberOfIndividual)
}
\arguments{
\item{object}{\code{DesignConstraint} object.}

\item{totalNumberOfIndividual}{Total number of individual to be set.}
}
\value{
The \code{DesignConstraint} object with the total number of individual.
}
\description{
Set the total number of individuals in a \code{DesignConstraint} object.
}
