% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setTotalSize<-}
\alias{setTotalSize<-}
\title{Set the total size of a Design.}
\usage{
setTotalSize(object) <- value
}
\arguments{
\item{object}{A \code{Design} object.}

\item{value}{A numeric giving the new value of the size of the design.}
}
\value{
The \code{Design} object with the new size.
}
\description{
Set the total size of a Design.
}
