% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{showArmData}
\alias{showArmData}
\title{Show the data of an arm for a design.}
\usage{
showArmData(object)
}
\arguments{
\item{object}{A \code{Design} object.}
}
\value{
Return a character string giving the the data summary of an arm for a design.
}
\description{
Show the data of an arm for a design.
}
