% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{showConstraints}
\alias{showConstraints}
\title{Show all the constraints of the \code{PFIMProject} object.}
\usage{
showConstraints(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
Show the all the objects \code{Constraints} in the \code{PFIMProject} object.
}
\description{
Show all the constraints of the \code{PFIMProject} object.
}
