% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{showDesigns}
\alias{showDesigns}
\title{Show all the Designs.}
\usage{
showDesigns(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
Show all the design \code{designs} in the \code{PFIMProject} object.
}
\description{
Show all the Designs.
}
