% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{showFims}
\alias{showFims}
\title{Show the Fisher Information Matrix for all the designs.}
\usage{
showFims(object)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}
}
\value{
Show the Fisher Information Matrix \code{fimOfDesign} for all the designs.
}
\description{
Show the Fisher Information Matrix for all the designs.
}
