% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/IndividualFim.R, R/PopulationFim.R
\name{showStatisticalModelStandardErrors}
\alias{showStatisticalModelStandardErrors}
\alias{showStatisticalModelStandardErrors,IndividualFim-method}
\alias{showStatisticalModelStandardErrors,PopulationFim-method}
\title{Show expected standard error data frame.}
\usage{
showStatisticalModelStandardErrors(object, modelParameters)

\S4method{showStatisticalModelStandardErrors}{IndividualFim}(object, modelParameters)

\S4method{showStatisticalModelStandardErrors}{PopulationFim}(object, modelParameters)
}
\arguments{
\item{object}{An \code{IndividualFim} object.}

\item{modelParameters}{A \code{modelParameters.} object.}
}
\value{
A data frame giving the standard error.

A dataframe giving the standard errors.

A dataframe giving the model standards errors.
}
\description{
Show expected standard error data frame.

Show the statistical model standard errors

Show the statistical model standards errors.
}
