% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexAlgorithm.R
\docType{class}
\name{SimplexAlgorithm-class}
\alias{SimplexAlgorithm-class}
\alias{SimplexAlgorithm}
\title{Class "SimplexAlgorithm"}
\description{
The Class "SimplexAlgorithm" implements the Nelder-Mead method (also downhill simplex method, amoeba method) [1].
}
\section{Objects from the class}{

Objects form the class \code{SimplexAlgorithm} can be created by calls of the form \code{SimplexAlgorithm(...)} where
(...) are the parameters for the \code{SimplexAlgorithm} objects.
}

\section{Slots for \code{SimplexAlgorithm} objects}{

 \describe{
   \item{\code{pct_initial_simplex_building}:}{A numeric giving the percentage of initial vertices for the simplex algorithm.}
   \item{\code{max_iteration}:}{A numeric giving the maximum of iterations.}
   \item{\code{tolerance}:}{A numeric giving the tolerance criteria for stopping the algorithm.}
   \item{\code{showProcess}:}{A boolean to show or not the process.}
   \item{\code{OptimalDesign}:}{A \code{Design} object giving the optimal design.}
 }
}

\references{
[1]
Nelder JA, Mead R (1965) A Simplex Method for Function Minimization. Comput J 7: 308-313.
}
