% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardDistribution.R
\docType{class}
\name{StandardDistribution-class}
\alias{StandardDistribution-class}
\alias{StandardDistribution}
\title{Class "StandardDistribution"}
\description{
Class \code{StandardDistribution} represents class for standard distributions
}
