% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\docType{class}
\name{StatisticalModel-class}
\alias{StatisticalModel-class}
\alias{StatisticalModel}
\title{Class "StatisticalModel"}
\description{
Class \code{StatisticalModel} represents a statistical model.
}
\section{Mathematical description of the statistical model}{

\describe{
\item{-}{\deqn{y = f(x, \theta) + g*\epsilon}, this part is considered in class \linkS4class{Response}}
\item{-}{\deqn{f(x, \theta)}, this part is considered in class \linkS4class{Response}}
\item{-}{\deqn{\theta={(\mu, \omega)}}, this part is considered in class \linkS4class{ModelParameter}}
\item{-}{\deqn{\epsilon}, this is a slot of the object \code{NormalDistribution} with mean = 0 and covariate_matrix = I}
}
}

\section{Objects from the class}{
 \code{StatisticalModel} objects are typically created by calls to \code{StatisticalModel} and contain the following slots:
}

\section{Slots for \code{StatisticalModel} objects}{

\describe{
\item{\code{modelEquations}:}{An object from the class \code{ModelEquations}}
\item{\code{responses}:}{A list of objects of type Responses -> f(x, theta)}
\item{\code{correlations}:}{A list giving all the covariables.}
\item{\code{model_parameters}:}{A list giving all the parameters of the models.}
}
}

