% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{summary,Design-method}
\alias{summary,Design-method}
\title{summary}
\usage{
\S4method{summary}{Design}(object)
}
\arguments{
\item{object}{A \code{Design} object.}
}
\value{
Return a list giving the name, the number of individuals, the total size of the design, the
the summary of all the parameters of the arms for a design and the amount of arm in the design.
}
\description{
summary
}
