% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{summaryArmData}
\alias{summaryArmData}
\title{Gives a summary of all the parameters of an arm for a design.}
\usage{
summaryArmData(object)
}
\arguments{
\item{object}{A \code{Design} object.}
}
\value{
Display a summary of all the parameters of the arms for a design.
}
\description{
Gives a summary of all the parameters of an arm for a design.
}
