% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateTest.R
\name{tost}
\alias{tost}
\title{tost: alias de covariateTest pour compatibilite ascendante}
\arguments{
\item{pfimproject}{Un objet \code{Evaluation}.}
}
\value{
Un objet \code{CovariateTest}.
}
\description{
tost: alias de covariateTest pour compatibilite ascendante
}
