% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{updateSamplingTimes}
\alias{updateSamplingTimes}
\title{Update sampling times for plotting}
\usage{
updateSamplingTimes(arm, samplingData, ...)
}
\arguments{
\item{arm}{An object of class \code{Arm}.}

\item{samplingData}{The list output from \code{getSamplingData}.}

\item{...}{Additional arguments}
}
\value{
The updated \code{Arm} object.
}
\description{
Replaces the arm's sampling times with a dense regular grid combined with
the original sampling points, enabling smooth curve rendering.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
