% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{addModel}
\alias{addModel}
\title{Add a model to a library of models.}
\usage{
addModel(object, model)
}
\arguments{
\item{object}{An object from the class \linkS4class{LibraryOfModels}.}

\item{model}{An object from the class \linkS4class{Model}.}
}
\value{
The library of models with the added model.
}
\description{
Add a model to a library of models.
}
