% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{Administration}
\alias{Administration}
\title{Administration}
\usage{
Administration(
  outcome = character(0),
  timeDose = numeric(0),
  dose = numeric(0),
  Tinf = numeric(0),
  tau = 0
)
}
\arguments{
\item{outcome}{A string giving the outcome for the administration.}

\item{timeDose}{A vector of double giving the time doses.}

\item{dose}{A vector of double giving the doses.}

\item{Tinf}{A vector of double giving the time for infusion Tinf.}

\item{tau}{An integer giving the tau value for repeated dose or steady state.}
}
\description{
The class \code{Administration} represents the administration and
stores information concerning the administration for the dosage regimen.
}
