% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraints.R
\name{AdministrationConstraints}
\alias{AdministrationConstraints}
\title{AdministrationConstraints}
\usage{
AdministrationConstraints(outcome = character(0), doses = list())
}
\arguments{
\item{outcome}{A string giving the outcome for the administration.}

\item{doses}{A vector of double giving the doses.}
}
\description{
The class \code{AdministrationConstraints} represents the constraint of an input to the system.
The class stores information concerning the constraints for the dosage regimen.
}
