% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\docType{class}
\name{Arm-class}
\alias{Arm-class}
\alias{Arm}
\title{Class "Arm"}
\description{
The class \code{Arm} combines the treatment and the sampling schedule.
}
\section{Objects from the class}{

Objects form the class \code{Arm} can be created by calls of the form \code{Arm(...)} where (...) are the parameters for the \code{Arm} objects.
}

\section{Slots for the \code{Arm} objects}{

\describe{
\item{\code{name}:}{A string giving the name of the arm.}
\item{\code{size}:}{An integer giving the number of subjects in the arm. By default set to 1.}
\item{\code{administrations}:}{A list of the administrations.}
\item{\code{initialConditions}:}{A list of the initial conditions.}
\item{\code{samplingTimes}:}{A list of the sampling times.}
\item{\code{administrationsConstraints}:}{A list of the administrations constraints.}
\item{\code{samplingTimesConstraints}:}{A list of the sampling times constraints.}
}
}

