% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{Arm}
\alias{Arm}
\title{Arm}
\usage{
Arm(
  name = character(0),
  size = numeric(0),
  administrations = list(),
  initialConditions = list(),
  samplingTimes = list(),
  administrationsConstraints = list(),
  samplingTimesConstraints = list(),
  evaluationModel = list(),
  evaluationGradients = list(),
  evaluationVariance = list(),
  evaluationFim = Fim()
)
}
\arguments{
\item{name}{A string giving the name of the arm.}

\item{size}{A integer giving the size of the arm.}

\item{administrations}{A list giving the objects of class \code{Administration} that define the administrations of the arm.}

\item{initialConditions}{A list giving the initial conditions for the ode model where the names are string that define the variable and their value are giving by double}

\item{samplingTimes}{A list giving the objects of class \code{SamplingTime} that define the sampling time of the arm.}

\item{administrationsConstraints}{A list giving the objects of class \code{AdministrationsConstraints} that define the administration constraints of the arm.}

\item{samplingTimesConstraints}{A list giving the objects of class \code{SamplingTimeConstraints} that define the sampling time constraints of the arm.}

\item{evaluationModel}{A list giving the evaluation of the responses of the arm.}

\item{evaluationGradients}{A list giving the evaluation of the responses gradient of the arm.}

\item{evaluationVariance}{A list giving the evaluation of the variance.}

\item{evaluationFim}{A object of class \code{Fim} giving the Fisher Information Matrix.}
}
\description{
The class \code{Arm} represents an arm and stores information concerning an arm.
}
