% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{armAdministration}
\alias{armAdministration}
\title{getArmAdministration: get the administration parameters of an arm.}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}
}
\value{
A list giving the administration parameters of an arm.
}
\description{
getArmAdministration: get the administration parameters of an arm.
}
