% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R
\docType{class}
\name{BayesianFim-class}
\alias{BayesianFim-class}
\alias{BayesianFim}
\title{Class "BayesianFim"}
\description{
The class \code{BayesianFim} represents the population Fisher information matrix.
The class \code{BayesianFim} inherits from the class \code{Fim}.
}
