% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R
\name{BayesianFim}
\alias{BayesianFim}
\title{BayesianFim Class}
\usage{
BayesianFim(
  fisherMatrix = numeric(0),
  fixedEffects = numeric(0),
  varianceEffects = numeric(0),
  SEAndRSE = list(),
  condNumberFixedEffects = 0,
  condNumberVarianceEffects = 0,
  shrinkage = numeric(0)
)
}
\arguments{
\item{fisherMatrix}{A numerical matrix representing the FIM.}

\item{fixedEffects}{A matrix representing fixed effects information.}

\item{varianceEffects}{A matrix representing variance components information.}

\item{SEAndRSE}{A data frame containing calculated SE and RSE values.}

\item{condNumberFixedEffects}{A numeric value for the fixed effects condition number.}

\item{condNumberVarianceEffects}{A numeric value for the variance effects condition number.}

\item{shrinkage}{A numeric vector representing parameter shrinkage.}
}
\description{
The \code{BayesianFim} class stores the Bayesian Fisher Information Matrix (FIM).
It extends the standard \code{Fim} class by incorporating Bayesian-specific
metrics, such as the shrinkage of individual parameters.
}
\section{Slots}{

\describe{
\item{\code{fisherMatrix}}{\code{matrix}. The numerical values of the Bayesian FIM.}

\item{\code{shrinkage}}{\code{numeric vector}. The shrinkage values for each random effect.}

\item{\code{fixedEffects}}{\code{matrix}. The FIM components related to fixed effects.}

\item{\code{varianceEffects}}{\code{matrix}. The FIM components related to variance components (random effects).}

\item{\code{SEAndRSE}}{\code{data.frame}. Standard Errors (SE) and Relative Standard Errors (RSE).}

\item{\code{condNumberFixedEffects}}{\code{numeric}. The condition number for the fixed effects sub-matrix.}

\item{\code{condNumberVarianceEffects}}{\code{numeric}. The condition number for the variance effects sub-matrix.}
}}

\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
