% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{checkSamplingTimeConstraintsForMetaheuristic}
\alias{checkSamplingTimeConstraintsForMetaheuristic}
\title{checkSamplingTimeConstraintsForMetaheuristic}
\arguments{
\item{samplingTimesConstraints}{An object \code{SamplingTimeConstraints}.}

\item{arm}{An object \code{Arm}.}

\item{newSamplings}{A vector of numeric for the new samplings.}

\item{outcome}{A string giving the outcome.}
}
\value{
A boolean TRUE/FALSE, with a message error if FALSE.
}
\description{
checkSamplingTimeConstraintsForMetaheuristic
}
