% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{checkValiditySamplingConstraint}
\alias{checkValiditySamplingConstraint}
\title{checkValiditySamplingConstraint: check if the constraints used for the design optimization are valid.}
\arguments{
\item{design}{An object \code{Design} giving the design.}
}
\value{
A boolean TRUE / FALSE, if FALSE it also gives an error message.
}
\description{
checkValiditySamplingConstraint: check if the constraints used for the design optimization are valid.
}
