% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{checkValiditySamplingConstraint}
\alias{checkValiditySamplingConstraint}
\title{Check the validity of he sampling times constraints}
\usage{
checkValiditySamplingConstraint(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Design}.}
}
\value{
An error message if a constraint is not valid.
}
\description{
Check the validity of he sampling times constraints
}
