% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined1.R
\docType{class}
\name{Combined1-class}
\alias{Combined1-class}
\alias{Combined1}
\title{Class "Combined1"}
\description{
The class \code{Combined1} defines the the residual error variance according
to the formula g(sigmaInter, sigmaSlope, cError, f(x, theta)) = sigmaInter + sigmaSlope*f(x,theta)).
The class \code{Combined1} inherits from the class \code{ModelError}.
}
\section{Objects from the class}{

Combined1 objects are typically created by calls to \code{Combined1} and contain the following slots that are inherited from
the class \linkS4class{ModelError}:

\describe{
\item{\code{outcome}:}{A string giving the name of the outcome.}
\item{\code{equation}:}{An symbolic expression of the model error.}
\item{\code{derivatives}:}{A list containing the derivatives of the model error expression.}
\item{\code{sigmaInter}:}{A numeric value giving the sigma inter of the error model.}
\item{\code{sigmaSlope}:}{A numeric value giving the sigma slope of the error model.}
\item{\code{cError}:}{A numeric value giving the exponant c of the error model.}
}
}

