% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationFim.R
\name{computeVMat}
\alias{computeVMat}
\title{computeVMat}
\usage{
computeVMat(varParam1, varParam2, invCholV)
}
\arguments{
\item{varParam1}{varParam1}

\item{varParam2}{varParam2}

\item{invCholV}{invCholV}
}
\value{
VMat
}
\description{
computeVMat
}
