% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constant.R
\docType{class}
\name{Constant-class}
\alias{Constant-class}
\alias{Constant}
\title{Class "Constant"}
\description{
The class \code{Constant} defines the the residual error variance according
to the formula g(sigma_inter, sigma_slope, c_error, f(x, theta)) = sigma_inter.
The class \code{Constant} inherits from the class \code{ModelError}.
}
\section{Objects from the class}{
 \code{Constant} objects are typically created by calls to \code{Constant}
and contain the following slots that are inherited from the class \linkS4class{ModelError}:

\describe{
\item{\code{outcome}:}{A string giving the name of the outcome.}
\item{\code{equation}:}{An symbolic expression of the model error.}
\item{\code{derivatives}:}{A list containing the derivatives of the model error expression.}
\item{\code{sigmaInter}:}{A numeric value giving the sigma inter of the error model.}
\item{\code{sigmaSlope}:}{A numeric value giving the sigma slope of the error model.}
\item{\code{cError}:}{A numeric value giving the exponant c of the error model.}
}
}

