% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constant.R
\name{Constant}
\alias{Constant}
\title{Constant}
\usage{
Constant(
  output = character(0),
  equation = expression(sigmaInter),
  derivatives = list(),
  sigmaInter = 0,
  sigmaSlope = 0,
  sigmaInterFixed = FALSE,
  sigmaSlopeFixed = FALSE,
  cError = 1
)
}
\arguments{
\item{output}{A string giving the model error output.}

\item{equation}{A expression giving the model error equation.}

\item{derivatives}{A list giving the derivatives of the model error equation.}

\item{sigmaInter}{A double giving the sigma inter.}

\item{sigmaSlope}{A double giving the sigma slope}

\item{sigmaInterFixed}{A boolean giving if the  sigma inter is fixed or not.}

\item{sigmaSlopeFixed}{A boolean giving if the  sigma slope is fixed or not.}

\item{cError}{A integer giving the power parameter.}
}
\description{
The class \code{Constant} represents and stores information for the error model Constant.
}
