% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGBOAlgorithm.R, R/PSOAlgorithm.R,
%   R/SimplexAlgorithm.R, R/FedorovWynnAlgorithm.R, R/MultiplicativeAlgorithm.R
\name{constraintsTableForReport}
\alias{constraintsTableForReport}
\title{constraintsTableForReport: table of the PGBOAlgorithm constraints for the report.}
\arguments{
\item{optimizationAlgorithm}{A object \code{MultiplicativeAlgorithm}.}

\item{arms}{List of the arms.}
}
\value{
The table for the constraints in the arms.

The table for the constraints in the arms.

The table for the constraints in the arms.

armsConstraintsTable

The table for the constraints in the arms.
}
\description{
constraintsTableForReport: table of the PGBOAlgorithm constraints for the report.

constraintsTableForReport: table of the PSOAlgorithm constraints for the report.

constraintsTableForReport: table of the SimplexAlgorithm constraints for the report.

constraintsTableForReport

constraintsTableForReport: table of the MultiplicativeAlgorithm constraints for the report.
}
