% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{convertPKModelAnalyticToPKModelODE}
\alias{convertPKModelAnalyticToPKModelODE}
\title{Convert an analytic model to a ode model.}
\usage{
convertPKModelAnalyticToPKModelODE(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
A ode model.
}
\description{
Convert an analytic model to a ode model.
}
