% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R, R/ModelInfusion.R,
%   R/ModelAnalyticInfusion.R
\name{convertPKModelAnalyticToPKModelODE}
\alias{convertPKModelAnalyticToPKModelODE}
\title{convertPKModelAnalyticToPKModelODE: conversion from analytic to ode}
\arguments{
\item{pkModel}{An object of class \code{ModelAnalyticInfusion} that defines the model.}
}
\description{
convertPKModelAnalyticToPKModelODE: conversion from analytic to ode

convertPKModelAnalyticToPKModelODE: conversion from analytic to ode

convertPKModelAnalyticToPKModelODE: conversion from analytic infusion to ode
}
