% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R, R/ModelInfusion.R,
%   R/ModelAnalyticInfusion.R
\name{convertPKModelAnalyticToPKModelODE}
\alias{convertPKModelAnalyticToPKModelODE}
\title{conversion from analytic to ode}
\arguments{
\item{pkModel}{An object of class \code{ModelInfusion} that defines the model.}
}
\description{
conversion from analytic to ode

conversion from analytic to ode
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.

Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
