% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{Dcriterion}
\alias{Dcriterion}
\title{Dcriterion: get the D-criterion of the Fim.}
\arguments{
\item{Fim}{A object \code{Fim} giving the Fim.}
}
\value{
A double giving the D-criterion of the Fim.
}
\description{
Dcriterion: get the D-criterion of the Fim.
}
