% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{defineFim}
\alias{defineFim}
\title{define the type of Fisher information matrix: population, individual or Bayesian}
\arguments{
\item{pfimproject}{An object \code{PFIMProject}.}
}
\value{
An object \code{Fim}.
}
\description{
define the type of Fisher information matrix: population, individual or Bayesian
}
