% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R, R/ModelAnalyticInfusion.R,
%   R/ModelAnalyticInfusionSteadyState.R, R/ModelODEBolus.R,
%   R/ModelODEDoseInEquations.R, R/ModelODEDoseNotInEquations.R,
%   R/ModelODEInfusionDoseInEquation.R
\name{defineModelAdministration}
\alias{defineModelAdministration}
\title{defineModelAdministration: define the administration}
\arguments{
\item{model}{An object of class \code{ModelODEInfusionDoseInEquation} that defines the model.}

\item{arm}{An object of class \code{Arm} that defines the arm.}
}
\value{
The model with samplings, solverInputs

The model with samplings, solverInputs

The model with samplings, solverInputs

The model with updated slots.

The model with samplings, solverInputs

The model with samplings, solverInputs

The model with updated slots.
}
\description{
defineModelAdministration: define the administration

defineModelAdministration: define the administration

defineModelAdministration: define the administration

defineModelAdministration: define the administration

defineModelAdministration: define the administration

defineModelAdministration: define the administration

defineModelAdministration: define the administration
}
