% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R, R/ModelAnalyticInfusion.R,
%   R/ModelAnalyticInfusionSteadyState.R, R/ModelAnalyticSteadyState.R,
%   R/ModelODEBolus.R, R/ModelODEDoseInEquations.R,
%   R/ModelODEDoseNotInEquations.R, R/ModelODEInfusionDoseInEquation.R
\name{defineModelAdministration}
\alias{defineModelAdministration}
\title{Define the administration for an analytic model}
\arguments{
\item{model}{An object of class \code{ModelAnalytic} that defines the model.}

\item{arm}{An object of class \code{Arm} that defines the arm.}
}
\value{
The model with samplings, solverInputs
}
\description{
Define the administration for an analytic model
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
