% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{defineModelEquationsFromLibraryOfModel}
\alias{defineModelEquationsFromLibraryOfModel}
\title{defineModelEquationsFromLibraryOfModel: define the model equations giving the models in the library of models.}
\arguments{
\item{pfimproject}{An object \code{PFIMProject} giving the evaluation to be run.}
}
\value{
A list giving the model equations.
}
\description{
defineModelEquationsFromLibraryOfModel: define the model equations giving the models in the library of models.
}
