% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{defineModelType}
\alias{defineModelType}
\alias{defineModelType,Model-method}
\title{Define the type of a model.}
\usage{
defineModelType(object, designs)

\S4method{defineModelType}{Model}(object, designs)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{designs}{A list of objects from the class \linkS4class{Design}.}
}
\value{
Return a model defined as analytic, ode, etc.
}
\description{
Define the type of a model.
}
