% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{defineModelType}
\alias{defineModelType}
\title{defineModelType: define the class of the model to be evaluated.}
\arguments{
\item{pfimproject}{An object \code{PFIMProject} giving the evaluation to be run.}
}
\value{
An object \code{Model} giving the model to be evaluated with its modelParameters, odeSolverParameters, modelError, modelEquations.
}
\description{
defineModelType: define the class of the model to be evaluated.
}
