% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{defineModelType}
\alias{defineModelType}
\title{defineModelType}
\usage{
defineModelType(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{\link{PFIMProject}} containing
the project specifications.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{Model} (or a subclass thereof) initialized
with \code{modelParameters}, \code{odeSolverParameters}, \code{modelError},
and \code{modelEquations}.
}
\description{
The method acts as a constructor for the specific model class
required for analysis. It extracts configurations from a \code{\link{PFIMProject}}
and instantiates a \code{Model} object, integrating equations, parameter
structures, error models, and solver settings.
}
\details{
This method determines whether the model should be treated as a:
\itemize{
  \item \bold{Library Model:} Pre-defined structural models (e.g., 1-compartment PK).
  \item \bold{User-Defined Model:} Custom equations provided via \code{modelEquations}.
  \item \bold{ODE Model:} Models requiring numerical integration using specified
  \code{odeSolverParameters}.
}
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
