% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R, R/ModelAnalyticInfusion.R,
%   R/ModelAnalyticInfusionSteadyState.R, R/ModelAnalyticSteadyState.R,
%   R/ModelODEBolus.R, R/ModelODEDoseInEquations.R,
%   R/ModelODEDoseNotInEquations.R, R/ModelODEInfusionDoseInEquation.R
\name{defineModelWrapper}
\alias{defineModelWrapper}
\title{defineModelWrapper: define the model wrapper for the ode solver}
\arguments{
\item{model}{An object of class \code{ModelODEInfusionDoseInEquation} that defines the model.}

\item{evaluation}{An object of class Evaluation that defines the evaluation}
}
\value{
The model with wrapperModelAnalytic, functionArgumentsModelAnalytic, functionArgumentsSymbolModelAnalytic, outputNames, outcomesWithAdministration

The model with wrapperModelAnalyticInfusion, functionArgumentsModelAnalyticInfusion, functionArgumentsSymbolModelAnalyticInfusion, outputNames, outcomesWithAdministration

The model with wrapperModelAnalyticInfusion, functionArgumentsModelAnalyticInfusion, functionArgumentsSymbolModelAnalyticInfusion, outputNames, outcomesWithAdministration

The model with wrapperModelAnalytic, functionArgumentsModelAnalytic, functionArgumentsSymbolModelAnalytic, outputNames, outcomesWithAdministration

The model with updated slots.

The model with the updated slots.

The model with the updated slots.

The model with updated slots.
}
\description{
defineModelWrapper: define the model wrapper for the ode solver

defineModelWrapper: define the model wrapper for the ode solver

defineModelWrapper: define the model wrapper for the ode solver

defineModelWrapper: define the model wrapper for the ode solver

defineModelWrapper: define the model wrapper for the ode solver

defineModelWrapper: define the model wrapper for the ode solver

defineModelWrapper: define the model wrapper for the ode solver

defineModelWrapper: define the model wrapper for the ode solver
}
