% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{defineOptimizationAlgorithm}
\alias{defineOptimizationAlgorithm}
\title{Define optimization algorithm}
\arguments{
\item{optimization}{An \code{Optimization} object.}
}
\value{
An optimization algorithm.
}
\description{
Define optimization algorithm
}
