% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{definePKModel}
\alias{definePKModel}
\title{Define a PK model.}
\usage{
definePKModel(object, outcomes)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{outcomes}{A list giving the outcomes of the  PK model.}
}
\value{
A model giving a PK model.
}
\description{
Define a PK model.
}
