% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R, R/ModelAnalytic.R,
%   R/ModelAnalyticSteadyState.R, R/ModelAnalyticInfusion.R,
%   R/ModelODEDoseInEquations.R, R/ModelODEDoseNotInEquations.R,
%   R/ModelODEInfusionDoseInEquations.R
\name{definePKModel}
\alias{definePKModel}
\alias{definePKModel,ModelAnalytic-method}
\alias{definePKModel,ModelAnalyticSteadyState-method}
\alias{definePKModel,ModelAnalyticInfusion-method}
\alias{definePKModel,ModelODEDoseInEquations-method}
\alias{definePKModel,ModelODE-method}
\alias{definePKModel,ModelODEInfusionDoseInEquations-method}
\title{Define a PK model.}
\usage{
definePKModel(object, outcomes)

\S4method{definePKModel}{ModelAnalytic}(object, outcomes)

\S4method{definePKModel}{ModelAnalyticSteadyState}(object, outcomes)

\S4method{definePKModel}{ModelAnalyticInfusion}(object, outcomes)

\S4method{definePKModel}{ModelODEDoseInEquations}(object, outcomes)

\S4method{definePKModel}{ModelODE}(object, outcomes)

\S4method{definePKModel}{ModelODEInfusionDoseInEquations}(object, outcomes)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{outcomes}{A list giving the outcomes of the  PK model.}
}
\value{
A model giving a PK model.
}
\description{
Define a PK model.
}
