% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{definePKPDModel}
\alias{definePKPDModel}
\title{Define a PKPD model.}
\usage{
definePKPDModel(PKModel, PDModel, outcomes)
}
\arguments{
\item{PKModel}{An object from the class \linkS4class{Model}.}

\item{PDModel}{An object from the class \linkS4class{Model}.}

\item{outcomes}{A list giving the outcomes of the  PKPD model.}
}
\value{
A model giving a PKPD model.
}
\description{
Define a PKPD model.
}
