% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R, R/ModelAnalyticInfusion.R,
%   R/ModelAnalyticInfusionSteadyState.R, R/ModelAnalyticSteadyState.R,
%   R/ModelODEInfusionDoseInEquation.R
\name{definePKPDModel}
\alias{definePKPDModel}
\title{definePKPDModel:  define a PKPD model from library of model}
\arguments{
\item{pkModel}{An object of class \code{ModelODE} that defines the PD model.}

\item{pfimproject}{An object of class \code{PFIMProject} that defines the pfimproject.}
}
\description{
definePKPDModel:  define a PKPD model from library of model

definePKPDModel:  define a PKPD model from library of model

definePKPDModel ModelAnalyticInfusion, ModelAnalytic

definePKPDModel ModelAnalyticInfusion, ModelODE

definePKPDModel

definePKPDModel

definePKPDModel

definePKPDModel: define a PKPD model from library of model
}
