% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{Design}
\alias{Design}
\title{Design}
\usage{
Design(
  name = character(0),
  size = 0,
  arms = list(),
  evaluationArms = list(),
  numberOfArms = 0,
  fim = Fim()
)
}
\arguments{
\item{name}{A string giving the name of the design.}

\item{size}{A integer giving the size of the design.}

\item{arms}{A list giving the arms of the design.}

\item{evaluationArms}{A list giving the valuation of the arms of the design.}

\item{numberOfArms}{A integer giving the number of arms.}

\item{fim}{A object \code{Fim} giving the Fim of the design.}
}
\description{
The class \code{Design} represents and stores information for the Design.
}
\details{
Design
}
