% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{class}
\name{Distribution-class}
\alias{Distribution-class}
\alias{Distribution}
\title{Class "Distribution"}
\description{
The class defines all the required methods for a distribution object.
}
\section{Objects from the class}{

Objects form the class \code{Distribution} can be created by calls of the form \code{Distribution(...)} where
(...) are the parameters for the \code{Distribution} objects.
}

\section{Slots for \code{Distribution} objects}{

 \describe{
   \item{\code{parameters}:}{A list containing the distribution parameters.}
 }
}

