% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{Distribution}
\alias{Distribution}
\title{Distribution}
\usage{
Distribution(name = character(0), mu = 0, omega = 0)
}
\arguments{
\item{name}{A string giving the name of the distribution.}

\item{mu}{A double giving the mean mu.}

\item{omega}{A double giving omega.}
}
\description{
The class \code{Distribution} represents and stores information for the parameter distribution.
}
