% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{evaluateArm}
\alias{evaluateArm}
\title{evaluateArm: evaluation of the model with the arm parameters.}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{model}{A object of class \code{Model} giving the model.}

\item{fim}{A object of class \code{Fim} giving the fim.}
}
\value{
The object arm with the slots evaluationModel, evaluationGradients, evaluationVariance and evaluationFim.
}
\description{
evaluateArm: evaluation of the model with the arm parameters.
}
