% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{EvaluateArm}
\alias{EvaluateArm}
\title{Evaluate an arm.}
\usage{
EvaluateArm(object, model, fim)
}
\arguments{
\item{object}{An object \code{arm} from the class \linkS4class{Arm}.}

\item{model}{An object \code{model} from the class \linkS4class{Model}.}

\item{fim}{An object \code{fim} from the class \linkS4class{Fim}.}
}
\value{
The object \code{fim} containing the Fisher Information Matrix
the two lists \code{evaluationOutcomes}, \code{outcomesGradient} containing the results of
the evaluation of the outcome and the sensitivity indices.
}
\description{
Evaluate an arm.
}
