% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{evaluateArm}
\alias{evaluateArm}
\title{Evaluate an arm}
\usage{
evaluateArm(arm, model, fim, ...)
}
\arguments{
\item{arm}{An object of class \code{Arm}.}

\item{model}{An object of class \code{Model}.}

\item{fim}{An object of class \code{Fim}.}

\item{...}{Additional arguments}
}
\value{
The \code{Arm} object with updated \code{evaluationModel},
\code{evaluationGradients}, \code{evaluationVariance}, and \code{evaluationFim}.
}
\description{
Evaluates the model, gradients, variance, and FIM for an arm.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
