% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{EvaluateDesign}
\alias{EvaluateDesign}
\title{Evaluate an design}
\usage{
EvaluateDesign(object, model, fim)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}

\item{model}{An object \code{model} from the class \linkS4class{Model}.}

\item{fim}{An object \code{fim} from the class \linkS4class{Fim}.}
}
\value{
The object \code{Design} with its slot \code{fim}, \code{evaluationOutcomes}, \code{outcomesGradient} updated.
}
\description{
Evaluate an design
}
