% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{evaluateDesign}
\alias{evaluateDesign}
\title{evaluateDesign: evaluation of a design.}
\arguments{
\item{design}{An object \code{Design} giving the design.}

\item{model}{An object \code{Model} giving the model.}

\item{fim}{An object \code{Fim} giving the Fim.}
}
\value{
The object \code{Design} with its evaluation results.
}
\description{
evaluateDesign: evaluation of a design.
}
