% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{EvaluateDesign}
\alias{EvaluateDesign}
\title{Evaluate the design for each arm.}
\usage{
EvaluateDesign(object, fimType, TheDesign)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}

\item{fimType}{A character string giving the type of FIM: "Population", "Individual" or "Bayesian".}

\item{TheDesign}{A \code{Design} object to be evaluated.}
}
\value{
The \code{PFIMProject} object with the list \code{designs} that contains the evaluation of each design for each arm.
}
\description{
Evaluate the design for each arm.
}
