% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{evaluateDesign}
\alias{evaluateDesign}
\title{Evaluation of a clinical design}
\arguments{
\item{design}{An object \code{Design} to evaluate.}

\item{model}{An object \code{Model} used for evaluation.}

\item{fim}{An object \code{Fim} to store results.}
}
\value{
The \code{Design} object with evaluated arms and aggregated global FIM.
}
\description{
Evaluation of a clinical design
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
