% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response.R
\name{EvaluateErrorModelDerivatives}
\alias{EvaluateErrorModelDerivatives}
\title{Evaluate the Error Model Derivatives.}
\usage{
EvaluateErrorModelDerivatives(object, f_x_i_theta)
}
\arguments{
\item{object}{A \code{Response} object.}

\item{f_x_i_theta}{The nonlinear structural model \code{f_x_i_theta}}
}
\value{
A list giving the error variance \code{V_sig} and sigma derivatives \code{sigmaDerivatives} of the model error.
}
\description{
Evaluate the Error Model Derivatives.
}
