% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{evaluateErrorModelDerivatives}
\alias{evaluateErrorModelDerivatives}
\title{evaluateErrorModelDerivatives; evaluate the derivatives of the model error.}
\arguments{
\item{modelError}{An object \code{ModelError} that defines the model error.}

\item{evaluationModel}{A dataframe giving the outputs for the model evaluation.}
}
\value{
The matrices sigmaDerivatives and errorVariance.
}
\description{
evaluateErrorModelDerivatives; evaluate the derivatives of the model error.
}
