% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{EvaluateErrorModelDerivatives}
\alias{EvaluateErrorModelDerivatives}
\title{Evaluate the error model derivatives.}
\usage{
EvaluateErrorModelDerivatives(object, evaluationOutcome)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelError}.}

\item{evaluationOutcome}{A list giving the results of the model evaluation.}
}
\value{
A list giving the error variance and the Sigma derivatives.
}
\description{
Evaluate the error model derivatives.
}
