% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{evaluateFim}
\alias{evaluateFim}
\title{evaluateFim: evaluation of the Fim}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{model}{An object \code{Model} giving the model.}

\item{arm}{An object \code{Arm} giving the arm.}
}
\value{
The object  \code{Fim} with the fisherMatrix and the shrinkage.

The object \code{IndividualFim} with the fisherMatrix and the shrinkage.

The object \code{IndividualFim} with the fisherMatrix and the shrinkage.
}
\description{
evaluateFim: evaluation of the Fim

evaluateFim: evaluation of the Fim

evaluateFim: evaluation of the Fim
}
