% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{evaluateFim}
\alias{evaluateFim}
\title{evaluateFim: compute the Bayesian FIM for one arm}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{model}{An object \code{Model} giving the model.}

\item{arm}{An object \code{Arm} giving the arm.}
}
\value{
The \code{BayesianFim} with fisherMatrix and shrinkage updated.

The \code{IndividualFim} with fisherMatrix updated.

The object \code{PopulationFim} with the fisherMatrix.
}
\description{
evaluateFim: compute the Bayesian FIM for one arm

evaluateFim: compute the individual FIM for one arm

evaluateFim: evaluation of the Fim
}
