% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{EvaluateFisherMatrix}
\alias{EvaluateFisherMatrix}
\title{Evaluate the Fisher matrix ( population, individual and Bayesian )}
\usage{
EvaluateFisherMatrix(object, model, arm, modelEvaluation, modelVariance)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{model}{An object from the class \linkS4class{Model}.}

\item{arm}{An object from the class \linkS4class{Arm}.}

\item{modelEvaluation}{A list containing the evaluation results.}

\item{modelVariance}{A list containing the model variance.}
}
\value{
An object from the class \linkS4class{Fim} containing the Fisher matrix.
}
\description{
Evaluate the Fisher matrix ( population, individual and Bayesian )
}
