% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODE.R, R/ModelODEBolus.R,
%   R/ModelODEInfusion.R
\name{evaluateInitialConditions}
\alias{evaluateInitialConditions}
\title{evaluateInitialConditions: evaluate the initial conditions.}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{model}{A object of class \code{ModelODEInfusion} giving the model.}

\item{doseEvent}{A data frame giving the dose event for the ode solver.}
}
\value{
A list giving the evaluated initial conditions.
}
\description{
evaluateInitialConditions: evaluate the initial conditions.

evaluateInitialConditions: evaluate the initial conditions.

evaluateInitialConditions: evaluate the initial conditions.
}
