% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{EvaluateModel}
\alias{EvaluateModel}
\title{Evaluate a model.}
\usage{
EvaluateModel(object, arm)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{arm}{An object from the class \linkS4class{Arm}.}
}
\value{
A list giving the results of the model evaluation.
}
\description{
Evaluate a model.
}
