% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEquations.R
\name{EvaluateModel}
\alias{EvaluateModel}
\title{Evaluate an analytic model.}
\usage{
EvaluateModel(object, samplingTimesModel, inputsModel, computeFIM)
}
\arguments{
\item{object}{An object \code{EvaluateModel}.}

\item{samplingTimesModel}{A vector containing the sampling times.}

\item{inputsModel}{A list containing the models input.}

\item{computeFIM}{A boolean for computing the FIM or not (plot or evaluation).}
}
\value{
A list containing the evaluated responses and the gradients.
}
\description{
Evaluate an analytic model.
}
