% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R, R/ModelAnalytic.R,
%   R/ModelAnalyticSteadyState.R, R/ModelAnalyticInfusion.R,
%   R/ModelAnalyticInfusionSteadyState.R, R/ModelODEBolus.R,
%   R/ModelODEDoseInEquations.R, R/ModelODEDoseNotInEquations.R,
%   R/ModelODEInfusionDoseInEquations.R
\name{EvaluateModel}
\alias{EvaluateModel}
\alias{EvaluateModel,ModelAnalytic-method}
\alias{EvaluateModel,ModelAnalyticSteadyState-method}
\alias{EvaluateModel,ModelAnalyticInfusion-method}
\alias{EvaluateModel,ModelAnalyticInfusionSteadyState-method}
\alias{EvaluateModel,ModelODEBolus-method}
\alias{EvaluateModel,ModelODEDoseInEquations-method}
\alias{EvaluateModel,ModelODEDoseNotInEquations-method}
\alias{EvaluateModel,ModelODEInfusionDoseInEquations-method}
\title{Evaluate a model.}
\usage{
EvaluateModel(object, dataForModelEvaluation, arm)

\S4method{EvaluateModel}{ModelAnalytic}(object, dataForModelEvaluation, arm)

\S4method{EvaluateModel}{ModelAnalyticSteadyState}(object, dataForModelEvaluation, arm)

\S4method{EvaluateModel}{ModelAnalyticInfusion}(object, dataForModelEvaluation, arm)

\S4method{EvaluateModel}{ModelAnalyticInfusionSteadyState}(object, dataForModelEvaluation, arm)

\S4method{EvaluateModel}{ModelODEBolus}(object, dataForModelEvaluation, arm)

\S4method{EvaluateModel}{ModelODEDoseInEquations}(object, dataForModelEvaluation, arm)

\S4method{EvaluateModel}{ModelODEDoseNotInEquations}(object, dataForModelEvaluation, arm)

\S4method{EvaluateModel}{ModelODEInfusionDoseInEquations}(object, dataForModelEvaluation, arm)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{dataForModelEvaluation}{...}

\item{arm}{An object from the class \linkS4class{Arm}.}
}
\value{
A list giving the results of the model evaluation.
}
\description{
Evaluate a model.
}
