% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R, R/ModelAnalyticInfusion.R,
%   R/ModelAnalyticInfusionSteadyState.R, R/ModelAnalyticSteadyState.R,
%   R/ModelODEBolus.R, R/ModelODEDoseInEquations.R,
%   R/ModelODEDoseNotInEquations.R, R/ModelODEInfusionDoseInEquation.R
\name{evaluateModel}
\alias{evaluateModel}
\title{evaluateModel: evaluate the model}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{model}{A object of class \code{ModelODEInfusionDoseInEquation} giving the model.}
}
\value{
A list of dataframes that contains the results for the evaluation of the model.

A list of dataframes that contains the results for the evaluation of the model.

A list of dataframes that contains the results for the evaluation of the model.

A list of dataframes that contains the results for the evaluation of the model.

A list of dataframes that contains the evaluation of the model.

A data frame giving the output of the model evaluation.

A list of dataframes that contains the results for the evaluation of the model.

A list of dataframes that contains the results for the evaluation of the model.

A data frame giving the output of the model evaluation.
}
\description{
evaluateModel: evaluate the model

evaluateModel: evaluate the ModelAnalyticInfusion

evaluateModel: evaluate the ModelAnalyticInfusion

evaluateModel: evaluate the ModelAnalyticInfusion

evaluateModel

evaluateModel

evaluateModel: evaluate the model

evaluateModel: evaluate the model

evaluateModel
}
