% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{evaluateModelGradient}
\alias{evaluateModelGradient}
\title{evaluateModelGradient: evaluate the gradient of the model}
\arguments{
\item{model}{An object \code{Model} that defines the model.}

\item{arm}{A object \code{Arm} giving the arm}
}
\value{
A data frame that contains the gradient of the model.
}
\description{
evaluateModelGradient: evaluate the gradient of the model
}
