% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{evaluateModelVariance}
\alias{evaluateModelVariance}
\title{evaluateModelVariance: evaluate the variance of the model}
\arguments{
\item{model}{A object \code{Model} giving the model.}

\item{arm}{A object \code{Arm} giving the arm}
}
\value{
A list giving errorVariance  and sigmaDerivatives.
}
\description{
evaluateModelVariance: evaluate the variance of the model
}
