% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{EvaluateVarianceFIM}
\alias{EvaluateVarianceFIM}
\title{Evaluate the variance of the Fisher information matrix.}
\usage{
EvaluateVarianceFIM(object, model, arm, modelEvaluation, modelVariance)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{model}{An object from the class \linkS4class{Model}.}

\item{arm}{An object from the class \linkS4class{Arm}.}

\item{modelEvaluation}{A list containing the evaluation results.}

\item{modelVariance}{A list containing the model variance.}
}
\value{
A list containing the matrices of the variance of the FIM.
}
\description{
Evaluate the variance of the Fisher information matrix.
}
