% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{evaluateVarianceFIM}
\alias{evaluateVarianceFIM}
\title{evaluateVarianceFIM: evaluate the variance}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{model}{A object of class \code{Model} giving the model.}

\item{fim}{A object of class \code{PopulationFim} giving the Fim.}
}
\value{
The matrices MFbeta and V.

The matrices MFbeta and V.

The matrices MFVar and V.
}
\description{
evaluateVarianceFIM: evaluate the variance

evaluateVarianceFIM: evaluate the variance

evaluateVarianceFIM: evaluate the variance
}
