% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{EvaluateVarianceModel}
\alias{EvaluateVarianceModel}
\title{Evaluate the variance of a model.}
\usage{
EvaluateVarianceModel(object, arm, evaluationModel)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{arm}{An object from the class \linkS4class{Arm}.}

\item{evaluationModel}{A list giving the outputs of the model evaluation.}
}
\value{
Return a list giving the results of the evaluation of the model variance.
}
\description{
Evaluate the variance of a model.
}
