% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{Evaluation}
\alias{Evaluation}
\title{Evaluation}
\usage{
Evaluation(
  evaluationDesign = list(),
  name = character(0),
  modelParameters = list(),
  modelEquations = list(),
  modelFromLibrary = list(),
  modelError = list(),
  designs = list(),
  outputs = list(),
  fimType = character(0),
  odeSolverParameters = list()
)
}
\arguments{
\item{evaluationDesign}{A list giving the evaluation of the design.}

\item{name}{A string giving the name of the design evaluation.}

\item{modelParameters}{A list giving the model parameters.}

\item{modelEquations}{A list giving the model equations.}

\item{modelFromLibrary}{A list giving the model equations from the library of model.}

\item{modelError}{A list giving the model error.}

\item{designs}{A list giving the designs to be evaluated.}

\item{outputs}{A list giving the model outputs.}

\item{fimType}{A string giving the type of Fim being evaluated.}

\item{odeSolverParameters}{A list giving the atol and rtol parameters for the ode solver.}
}
\description{
The class \code{Evaluation} represents and stores information for the evaluation of a design
}
