% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{FedorovWynnAlgorithm_Rcpp}
\alias{FedorovWynnAlgorithm_Rcpp}
\title{Fedorov-Wynn algorithm in Rcpp.}
\usage{
FedorovWynnAlgorithm_Rcpp(
  protocols_input,
  ndimen_input,
  nbprot_input,
  numprot_input,
  freq_input,
  nbdata_input,
  vectps_input,
  fisher_input,
  nok_input,
  protdep_input,
  freqdep_input
)
}
\arguments{
\item{protocols_input}{parameter protocols_input}

\item{ndimen_input}{parameter ndimen_input}

\item{nbprot_input}{parameter nbprot_input}

\item{numprot_input}{parameter numprot_input}

\item{freq_input}{parameter freq_input}

\item{nbdata_input}{parameter nbdata_input}

\item{vectps_input}{parameter vectps_input}

\item{fisher_input}{parameter fisher_input}

\item{nok_input}{parameter nok_input}

\item{protdep_input}{parameter protdep_input}

\item{freqdep_input}{parameter freqdep_input}
}
\value{
A list giving the results of the outputs of the FedorovWynn algorithm.
}
\description{
Run the FedorovWynnAlgorithm in Rcpp
}
