% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\docType{class}
\name{Fim-class}
\alias{Fim-class}
\alias{Fim}
\title{Class "Fim" representing the Fisher information matrix,
a parent class used by three classes PopulationFim, IndividualFim and BayesianFim.}
\description{
A class storing information regarding the Fisher computation matrix.
Type of the Fisher information: population ("PopulationFIM"), individual ("IndividualFIM") or Bayesian ("BayesianFIM").
The computation method for population and Bayesian matrix is first order linearisation (FO).
}
\section{Objects from the class}{

Objects form the class \code{Fim} can be created by calls of the form \code{Fim(...)} where
(...) are the parameters for the \code{Fim} objects.
}

\section{Slots for \code{Fim} objects}{

\describe{
\item{\code{isOptimizationResult}:}{A Boolean giving TRUE for an optimization result and FALSE an evaluation result.}
\item{\code{mfisher}:}{A matrix of numeric giving the Fisher information.}
\item{\code{omega}:}{A matrix of numeric giving the variances.}
\item{\code{mu}:}{A matrix of numeric giving the means.}
\item{\code{fim_comput_method}:}{Name of the method used to approximate the population matrix : character strings, 'FO' }
}
}

