% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{finiteDifferenceHessian}
\alias{finiteDifferenceHessian}
\title{finiteDifferenceHessian: compute the Hessian}
\arguments{
\item{model}{A object \code{Model} giving the model.}
}
\value{
The model with the slots parametersForComputingGradient with XcolsInv, shifted, frac.
}
\description{
finiteDifferenceHessian: compute the Hessian
}
