% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexAlgorithm.R
\name{fisher.simplex}
\alias{fisher.simplex}
\title{Compute the fisher.simplex}
\usage{
fisher.simplex(samplingTimes, data)
}
\arguments{
\item{samplingTimes}{A \code{SamplnigTimes} object.}

\item{data}{A list containing the design, arm, response names, statistical model, constraint and FIM.}
}
\value{
The fisher.simplex giving the evalation of the optimization criterion ( i.e. D-criterion)
}
\description{
Compute the fisher.simplex
}
