% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexAlgorithm.R
\name{fun.amoeba}
\alias{fun.amoeba}
\title{Compute the fun.amoeba}
\usage{
fun.amoeba(p, y, ftol, itmax, funk, outcomes, data, showProcess)
}
\arguments{
\item{p}{parameter p}

\item{y}{parameter y}

\item{ftol}{parameter ftol}

\item{itmax}{parameter itmax}

\item{funk}{parameter funk}

\item{outcomes}{The model outcomes.}

\item{data}{parameter data}

\item{showProcess}{Boolean.}
}
\value{
fun.amoeba
}
\description{
Compute the fun.amoeba
}
