% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexAlgorithm.R
\name{fun.amoeba}
\alias{fun.amoeba}
\title{function fun.amoeba}
\usage{
fun.amoeba(p, y, ftol, itmax, funk, outcomes, data, showProcess)
}
\arguments{
\item{p}{input is a matrix p whose ndim+1 rows are ndim-dimensional vectors which are the vertices of the starting simplex.}

\item{y}{vector whose components must be pre-initialized to the values of funk evaluated at the ndim+1 vertices (rows) of p.}

\item{ftol}{the fractional convergence tolerance to be achieved in the function value.}

\item{itmax}{maximal number of iterations.}

\item{funk}{multidimensional function to be optimized.}

\item{outcomes}{A vector giving the outcomes.}

\item{data}{a fixed set of data.}

\item{showProcess}{A boolean for showing the process or not.}
}
\value{
A list containing the components of the optimized simplex.
}
\description{
function fun.amoeba
}
