% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{generateDosesCombination}
\alias{generateDosesCombination}
\title{generateDosesCombination: generate the combination for the doses.}
\arguments{
\item{design}{An object \code{Design} giving the design.}
}
\value{
dosesForFIMs, numberOfDoses used in the design optimization.
}
\description{
generateDosesCombination: generate the combination for the doses.
}
