% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{generateFimsFromConstraints}
\alias{generateFimsFromConstraints}
\title{Generate FIMs from constraints}
\arguments{
\item{optimization}{An \code{Optimization} object.}
}
\value{
A list containing FIMs from constraints.
}
\description{
Generate FIMs from constraints
}
