% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{generateFimsFromConstraints}
\alias{generateFimsFromConstraints}
\title{Generate the fim from the constraints}
\usage{
generateFimsFromConstraints(object, fims)
}
\arguments{
\item{object}{An object from the class \linkS4class{Optimization}.}

\item{fims}{A list of object from the class \linkS4class{Fim}.}
}
\value{
A list giving the arms with their fims.
}
\description{
Generate the fim from the constraints
}
