% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{generateFimsFromConstraints}
\alias{generateFimsFromConstraints}
\title{generateFimsFromConstraints: enumerate and evaluate all FIMs from constraints}
\arguments{
\item{optimization}{An \code{Optimization} object.}
}
\value{
A list with listArms, dimFim, listFimsAlgoFW, listFimsAlgoMult,
  samplingsForFedorovWynnAlgo.
}
\description{
generateFimsFromConstraints: enumerate and evaluate all FIMs from constraints
}
