% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{generateFimsFromConstraints}
\alias{generateFimsFromConstraints}
\title{Generate Fisher Information Matrices (FIM) from Design Constraints}
\usage{
generateFimsFromConstraints(optimization, ...)
}
\arguments{
\item{optimization}{An object of class \code{Optimization}
containing the model definitions, parameter values, and design constraints.}

\item{...}{Additional arguments.}
}
\value{
The updated \code{optimization} object, where the slot
\code{fisherMatrices} contains the list of matrices calculated for each
candidate arm.
}
\description{
This method explores the design space defined by the project constraints and
computes the Fisher Information Matrix (FIM) for every candidate elementary
design arm. These matrices are then stored as a library to be used by the
optimization algorithms.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
