% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{generateReportEvaluation}
\alias{generateReportEvaluation}
\title{Generate the report for the evaluation}
\usage{
generateReportEvaluation(
  object,
  evaluationObject,
  outputPath,
  outputFile,
  plotOptions
)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{evaluationObject}{A list giving the results of the evaluation of the model.}

\item{outputPath}{A string giving the output path.}

\item{outputFile}{A string giving the name of the output file.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
Return the report for the evaluation in html.
}
\description{
Generate the report for the evaluation
}
