% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{generateReportEvaluation}
\alias{generateReportEvaluation}
\title{generateReportEvaluation: generate the report for the model evaluation.}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{tablesForReport}{The output list giving by the method tablesForReport.}
}
\value{
The html report for the design evaluation.

The html report for the model evaluation.

The html report for the model evaluation.
}
\description{
generateReportEvaluation: generate the report for the model evaluation.

generateReportEvaluation: generate the report for the model evaluation.

generateReportEvaluation: generate the report for the model evaluation.
}
