% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{generateReportOptimization}
\alias{generateReportOptimization}
\title{generateReportOptimization: generate the report for the design optimization.}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{optimizationAlgorithm}{An object \code{PGBOAlgorithm} giving the PGBOAlgorithm}

\item{tablesForReport}{The output list giving by the method tablesForReport.}
}
\value{
The html report for the design optimization.

The html report for the design optimization.

The html report.

The html report.

The html report.

The html report.

The html report.

The html report.

The html report.

The html report.

The html report.

The html report.
}
\description{
generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.

generateReportOptimization: generate the report for the design optimization.
}
