% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationAlgorithm.R
\name{generateReportOptimization}
\alias{generateReportOptimization}
\title{Generate report for the optimization.}
\usage{
generateReportOptimization(
  object,
  optimizationObject,
  outputPath,
  outputFile,
  plotOptions
)
}
\arguments{
\item{object}{An object from the class \linkS4class{OptimizationAlgorithm}.}

\item{optimizationObject}{An object from the class \linkS4class{Optimization}.}

\item{outputPath}{A string giving the output path.}

\item{outputFile}{A string giving the name of the output file.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
The report for the optimization in html.
}
\description{
Generate report for the optimization.
}
