% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{generateSamplingsFromSamplingConstraints}
\alias{generateSamplingsFromSamplingConstraints}
\title{Generate samplings from sampling constraints}
\usage{
generateSamplingsFromSamplingConstraints(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{SamplingTimeConstraints}.}
}
\value{
A list of sampling times generated from the sampling constraints.
}
\description{
Generate samplings from sampling constraints
}
