% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{generateSamplingsFromSamplingConstraints}
\alias{generateSamplingsFromSamplingConstraints}
\title{generateSamplingsFromSamplingConstraints}
\arguments{
\item{samplingTimeConstraints}{An object \code{SamplingTimeConstraints}}
}
\value{
A list intervalsConstraints.
}
\description{
generateSamplingsFromSamplingConstraints
}
