% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{generateSamplingTimesCombination}
\alias{generateSamplingTimesCombination}
\title{generateSamplingTimesCombination: generate the combination for the samplings.}
\arguments{
\item{design}{An object \code{Design} giving the design.}
}
\value{
samplingTimesCombinations used in the design optimization.
}
\description{
generateSamplingTimesCombination: generate the combination for the samplings.
}
