% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{getAdjustedGradient}
\alias{getAdjustedGradient}
\title{Get the adjusted gradient.}
\usage{
getAdjustedGradient(object, outcomesGradient)
}
\arguments{
\item{object}{An object \code{distribution} from the class \linkS4class{Distribution}.}

\item{outcomesGradient}{A list containing the evaluation of the outcome gradients.}
}
\value{
A list giving the adjusted gradient.
}
\description{
Get the adjusted gradient.
}
