% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getAdministration}
\alias{getAdministration}
\title{Get the administrations by outcome.}
\usage{
getAdministration(object, outcome)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{outcome}{A string giving the name of the outcome.}
}
\value{
The element of the list \code{administrations} containing the administration of the outcome \code{outcome}
}
\description{
Get the administrations by outcome.
}
