% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getAdministrationConstraint}
\alias{getAdministrationConstraint}
\title{Get the administration constraints by outcome.}
\usage{
getAdministrationConstraint(object, outcome)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{outcome}{A string giving the name of the outcome.}
}
\value{
The element of the list \code{getAdministrationConstraint} containing the administration constraints of the outcome \code{outcome}
}
\description{
Get the administration constraints by outcome.
}
