% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getArmConstraints}
\alias{getArmConstraints}
\title{getArmConstraints: get the administration and sampling time constraints for the MultiplicativeAlgorithm.}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{optimizationAlgorithm}{A object of class \code{Optimization} giving the optimization algorithm.}
}
\value{
A list giving the administration and sampling time constraints for the MultiplicativeAlgorithm.

A list giving the administration and sampling time constraints for the FedorovWynnAlgorithm.

A list giving the administration and sampling time constraints for the SimplexAlgorithm.

A list giving the administration and sampling time constraints for the PSOAlgorithm.

A list giving the administration and sampling time constraints for the PGBOAlgorithm.
}
\description{
getArmConstraints: get the administration and sampling time constraints for the MultiplicativeAlgorithm.

getArmConstraints: get the administration and sampling time constraints for the FedorovWynnAlgorithm.

getArmConstraints: get the administration and sampling time constraints for the SimplexAlgorithm.

getArmConstraints: get the administration and sampling time constraints for the PSOAlgorithm.

getArmConstraints: get the administration and sampling time constraints for the PGBOAlgorithm.
}
