% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getArmData}
\alias{getArmData}
\title{getArmData: extract arm data for The Report}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}
}
\value{
A list giving the name, Number of subjects, Outcome, Dose and Sampling times of the arm.
}
\description{
getArmData: extract arm data for The Report
}
