% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getArms}
\alias{getArms}
\title{Get the arms of an object.}
\usage{
getArms(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
A list containing the arms of the object.
}
\description{
Get the arms of an object.
}
