% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getArms}
\alias{getArms}
\title{Get the arms of a design.}
\usage{
getArms(object)
}
\arguments{
\item{object}{A \code{Design} object.}
}
\value{
A list \code{arms} of the arms of a design.
}
\description{
Get the arms of a design.
}
