% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getCError}
\alias{getCError}
\title{Get the \code{CError} of a \code{ModelError} object.}
\usage{
getCError(object)
}
\arguments{
\item{object}{\code{ModelError} object.}
}
\value{
The numeric \code{c_error} giving the CError.
}
\description{
Get the \code{CError} of a \code{ModelError} object.
}
