% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{getColumnAndParametersNamesFIMInLatex}
\alias{getColumnAndParametersNamesFIMInLatex}
\alias{getColumnAndParametersNamesFIMInLatex,BayesianFim-method}
\alias{getColumnAndParametersNamesFIMInLatex,IndividualFim-method}
\alias{getColumnAndParametersNamesFIMInLatex,PopulationFim-method}
\title{Get the names of the names of the parameters associated to each column of the fim in Latex format.}
\usage{
getColumnAndParametersNamesFIMInLatex(object, model)

\S4method{getColumnAndParametersNamesFIMInLatex}{BayesianFim}(object, model)

\S4method{getColumnAndParametersNamesFIMInLatex}{IndividualFim}(object, model)

\S4method{getColumnAndParametersNamesFIMInLatex}{PopulationFim}(object, model)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{model}{An object from the class \linkS4class{Model}.}
}
\value{
A list giving the names of the parameters associated to each column of the fim in Latex format.
}
\description{
Get the names of the names of the parameters associated to each column of the fim in Latex format.
}
