% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getConditionNumberFixedEffects}
\alias{getConditionNumberFixedEffects}
\title{Get the condition number of the matrix of the fixed effects.}
\usage{
getConditionNumberFixedEffects(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
A numeric giving the condition number of the matrix of the fixed effects.
}
\description{
Get the condition number of the matrix of the fixed effects.
}
