% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R
\name{getConditionNumberVarianceEffects}
\alias{getConditionNumberVarianceEffects}
\alias{getConditionNumberVarianceEffects,Fim-method}
\alias{getConditionNumberVarianceEffects,BayesianFim-method}
\title{Get the condition number of the matrix of the variance effects.}
\usage{
getConditionNumberVarianceEffects(object)

\S4method{getConditionNumberVarianceEffects}{Fim}(object)

\S4method{getConditionNumberVarianceEffects}{BayesianFim}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}..}
}
\value{
A numeric giving the condition number of the matrix of the variance effects.
}
\description{
Get the condition number of the matrix of the variance effects.
}
