% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Fim.R, R/Evaluation.R,
%   R/Optimization.R
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\alias{getCorrelationMatrix,Fim-method}
\alias{getCorrelationMatrix,Evaluation-method}
\alias{getCorrelationMatrix,Optimization-method}
\title{Get the correlation matrix.}
\usage{
getCorrelationMatrix(object)

\S4method{getCorrelationMatrix}{Fim}(object)

\S4method{getCorrelationMatrix}{Evaluation}(object)

\S4method{getCorrelationMatrix}{Optimization}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
The correlation matrix of the fim.
}
\description{
Get the correlation matrix.
}
