% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\title{Get the correlation matrix.}
\usage{
getCorrelationMatrix(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
The correlation matrix of the fim.
}
\description{
Get the correlation matrix.
}
