% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\title{getCorrelationMatrix : get the correlation matrix}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The correlation matrix

The Dcriterion
}
\description{
getCorrelationMatrix : get the correlation matrix

getCorrelationMatrix : get the correlation matrix
}
