% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationAlgorithm.R,
%   R/FedorovWynnAlgorithm.R, R/MultiplicativeAlgorithm.R, R/Optimization.R
\name{getDataFrameResults}
\alias{getDataFrameResults}
\alias{getDataFrameResults,FedorovWynnAlgorithm-method}
\alias{getDataFrameResults,MultiplicativeAlgorithm-method}
\alias{getDataFrameResults,Optimization-method}
\title{Get the dataframe of the results.}
\usage{
getDataFrameResults(object)

\S4method{getDataFrameResults}{FedorovWynnAlgorithm}(object)

\S4method{getDataFrameResults}{MultiplicativeAlgorithm}(object)

\S4method{getDataFrameResults}{Optimization}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{OptimizationAlgorithm}.}
}
\value{
Return the dataframe of the results.
}
\description{
Get the dataframe of the results.
}
