% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{getDataFrameResults}
\alias{getDataFrameResults}
\alias{getDataFrameResults,MultiplicativeAlgorithm-method}
\title{Get the dataframe of the results.}
\usage{
getDataFrameResults(object, threshold)

\S4method{getDataFrameResults}{MultiplicativeAlgorithm}(object, threshold)
}
\arguments{
\item{object}{An object from the class \linkS4class{MultiplicativeAlgorithm}.}

\item{threshold}{The threshold for the optimal weights.}
}
\value{
Return the dataframe of the results.
}
\description{
Get the dataframe of the results.
}
