% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getDcriterion}
\alias{getDcriterion}
\title{Get the D-criterion for a \code{Fim} object.}
\usage{
getDcriterion(object)
}
\arguments{
\item{object}{A \code{Fim} object.}
}
\value{
A numeric \code{Dcriterion} giving the D-criterion of a Fisher Information Matrix.
}
\description{
Get the D-criterion for a \code{Fim} object.
}
