% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getDcriterion}
\alias{getDcriterion}
\title{getDcriterion : get the Dcriterion}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The Dcriterion of the FIM.

The Dcriterion
}
\description{
getDcriterion : get the Dcriterion

getDcriterion : get the Dcriterion
}
