% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getDcriterion}
\alias{getDcriterion}
\title{getDcriterion: Extract the D-optimality criterion}
\usage{
getDcriterion(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{...}{Additional arguments.}
}
\value{
A numeric value representing the D-criterion.
}
\description{
Returns the D-criterion derived from the determinant of the FIM, normalized
by the number of parameters for cross-design comparisons.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{

# Examples from Vignette 1 and 2

# Extract the D-criterion from the FIM evaluation results
dCriterion = getDcriterion(evaluationPopulationFIMResults)

# Display the D-criterion value
print(dCriterion)

}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
