% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getDerivatives}
\alias{getDerivatives}
\title{Get the derivatives of the model error equation.}
\usage{
getDerivatives(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelError}.}
}
\value{
The derivatives of the model error equation.
}
\description{
Get the derivatives of the model error equation.
}
