% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getDeterminant}
\alias{getDeterminant}
\title{Get the Determinant of a Fisher Information Matrix.}
\usage{
getDeterminant(object)
}
\arguments{
\item{object}{\code{Fim} object.}
}
\value{
A numeric \code{Det} giving the determinant of a Fisher Information Matrix.
}
\description{
Get the Determinant of a Fisher Information Matrix.
}
