% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Fim.R, R/Evaluation.R,
%   R/Optimization.R
\name{getDeterminant}
\alias{getDeterminant}
\alias{getDeterminant,Fim-method}
\alias{getDeterminant,Evaluation-method}
\alias{getDeterminant,Optimization-method}
\title{Get the determinant of the fim.}
\usage{
getDeterminant(object)

\S4method{getDeterminant}{Fim}(object)

\S4method{getDeterminant}{Evaluation}(object)

\S4method{getDeterminant}{Optimization}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
A numeric giving the determinant of the fim.
}
\description{
Get the determinant of the fim.
}
