% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getDeterminant}
\alias{getDeterminant}
\title{Get the determinant of the fim.}
\usage{
getDeterminant(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
A numeric giving the determinant of the fim.
}
\description{
Get the determinant of the fim.
}
