% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getDeterminant}
\alias{getDeterminant}
\title{getDeterminant: get the determinant}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The determinant of the FIM.

The determinant
}
\description{
getDeterminant: get the determinant

getDeterminant: get the determinant
}
