% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getDeterminant}
\alias{getDeterminant}
\title{getDeterminant}
\usage{
getDeterminant(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{...}{Additional arguments.}
}
\value{
A numeric value representing the determinant of the FIM.
}
\description{
Returns the determinant of the Fisher Information Matrix (FIM),
a global measure of design information used for D-optimality.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{

# Extract the determinant of the Fisher Information Matrix (FIM)
determinant = getDeterminant(evaluationPopulationFIMResults)

# Display the determinant value
print(determinant)

}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
