% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{getDistribution}
\alias{getDistribution}
\title{Get the distribution.}
\usage{
getDistribution(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelParameter}.}
}
\value{
The parameter distribution.
}
\description{
Get the distribution.
}
