% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{getDose}
\alias{getDose}
\title{Get the amount of doses.}
\usage{
getDose(object)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}
}
\value{
The numeric \code{amount_dose} giving the amount of doses.
}
\description{
Get the amount of doses.
}
