% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{getElementaryProtocols}
\alias{getElementaryProtocols}
\title{Get the elementary protocols.}
\usage{
getElementaryProtocols(object, fims)
}
\arguments{
\item{object}{An object from the class \linkS4class{Optimization}.}

\item{fims}{A list of object from the class \linkS4class{Fim}.}
}
\value{
A list containing the results of the evaluation of the elementary protocols giving
the numberOfTimes, nbOfDimensions, totalCost, samplingTimes and the fisherMatrices
}
\description{
Get the elementary protocols.
}
