% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getEquation}
\alias{getEquation}
\title{Get the equation of a model error.}
\usage{
getEquation(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelError}.}
}
\value{
An expression giving the equation of a model error.
}
\description{
Get the equation of a model error.
}
