% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getEquations}
\alias{getEquations}
\alias{getEquations,Model-method}
\title{Get the equations of a model.}
\usage{
getEquations(object)

\S4method{getEquations}{Model}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
The list giving the equations of the model.
}
\description{
Get the equations of a model.
}
