% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{getEvaluationFIMResults}
\alias{getEvaluationFIMResults}
\alias{getEvaluationFIMResults,Optimization-method}
\title{Get the results of the evaluation.}
\usage{
getEvaluationFIMResults(object)

\S4method{getEvaluationFIMResults}{Optimization}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Optimization}.}
}
\value{
An object from the class \linkS4class{Evaluation} giving the evaluation results for the optimal design.
}
\description{
Get the results of the evaluation.
}
