% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{getFim}
\alias{getFim}
\title{getFim: get the Fisher matrix.}
\arguments{
\item{evaluation}{An object \code{Evaluation} giving the evaluation to be run.}
}
\value{
The matrices fisherMatrix, fixedEffects, varianceEffects.
}
\description{
getFim: get the Fisher matrix.
}
