% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getFim}
\alias{getFim}
\title{Get the fim of an of an object.}
\usage{
getFim(object)

getFim(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
The \code{FIM} of the object.

The fim of the object.
}
\description{
Get the fim of an of an object.

Get the fim of an object.
}
