% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Design.R, R/PFIMProject.R,
%   R/OptimizationAlgorithm.R
\name{getFim}
\alias{getFim}
\alias{getFim,Design-method}
\alias{getFim,PFIMProject-method}
\alias{getFim,OptimizationAlgorithm-method}
\title{getFim}
\usage{
getFim(object)

getFim(object)

\S4method{getFim}{Design}(object)

\S4method{getFim}{PFIMProject}(object)

\S4method{getFim}{OptimizationAlgorithm}(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
The \code{FIM} of the object.

The fim of the object.
}
\description{
Get the fim of an of an object.
}
