% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getFim}
\alias{getFim}
\title{Get the Fisher Information Matrix.}
\usage{
getFim(object, ...)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}

\item{...}{A list giving the index of the Fim.}
}
\value{
A \code{Fim} object giving the Fisher Information Matrix of a design.
}
\description{
Get the Fisher Information Matrix.
}
