% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Fim.R, R/Evaluation.R,
%   R/Optimization.R
\name{getFisherMatrix}
\alias{getFisherMatrix}
\alias{getFisherMatrix,Fim-method}
\alias{getFisherMatrix,Evaluation-method}
\alias{getFisherMatrix,Optimization-method}
\title{Get the FIM.}
\usage{
getFisherMatrix(object)

\S4method{getFisherMatrix}{Fim}(object)

\S4method{getFisherMatrix}{Evaluation}(object)

\S4method{getFisherMatrix}{Optimization}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
A matrix giving the FIM.
}
\description{
Get the FIM.
}
