% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getFisherMatrix}
\alias{getFisherMatrix}
\title{getFisherMatrix: display the Fisher matrix components}
\arguments{
\item{evaluation}{An object \code{Evaluation} giving the evaluation to be run.}
}
\value{
The matrices fisherMatrix, fixedEffects, varianceEffects.

The matrices fisherMatrix, fixedEffects, varianceEffects.
}
\description{
getFisherMatrix: display the Fisher matrix components

getFisherMatrix: display the Fisher matrix components
}
