% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getFisherMatrix}
\alias{getFisherMatrix}
\title{getFisherMatrix}
\usage{
getFisherMatrix(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{Evaluation}.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with \code{fisherMatrix}, \code{fixedEffects},
  and \code{varianceEffects}.
}
\description{
Extracts partitioned components of the FIM for an \code{Evaluation} object.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{
fisherMatrixComponents = getFisherMatrix(evaluationPopulationFIMResults)
# Access specific matrices
FIM             = fisherMatrixComponents$fisherMatrix
fixedEffects    = fisherMatrixComponents$fixedEffects
varianceEffects = fisherMatrixComponents$varianceEffects
}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
