% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{getFixedParameters}
\alias{getFixedParameters}
\title{Get the fixed and non fixed model parameters.}
\usage{
getFixedParameters(object)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}
}
\value{
A list that contains the name and indices of the fixed parameters.
}
\description{
Get the fixed and non fixed model parameters.
}
