% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{getFixedTimes}
\alias{getFixedTimes}
\alias{getFixedTimes,SamplingTimeConstraints-method}
\title{Get the fixed sampling times.}
\usage{
getFixedTimes(object)

\S4method{getFixedTimes}{SamplingTimeConstraints}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{SamplingTimeConstraints}.}
}
\value{
A vector giving the foxed sampling times.
}
\description{
Get the fixed sampling times.
}
