% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingConstraint.R
\name{getfixedTimes}
\alias{getfixedTimes}
\title{Get the fixed times.}
\usage{
getfixedTimes(object)
}
\arguments{
\item{object}{A \code{SamplingConstraint} object.}
}
\value{
The fixed times of the \code{SamplingConstraint} object.
}
\description{
Get the fixed times.
}
