% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getInitialConditions}
\alias{getInitialConditions}
\title{Get the initial condition for the evaluation of an ode model.}
\usage{
getInitialConditions(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
The list \code{initialConditions} for the object \code{Arm}.
}
\description{
Get the initial condition for the evaluation of an ode model.
}
