% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{getLambda}
\alias{getLambda}
\alias{getLambda,MultiplicativeAlgorithm-method}
\title{Get the parameter lambda.}
\usage{
getLambda(object)

\S4method{getLambda}{MultiplicativeAlgorithm}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{MultiplicativeAlgorithm}.}
}
\value{
A numeric giving the parameter lambda.
}
\description{
Get the parameter lambda.
}
