% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{getListLastName}
\alias{getListLastName}
\title{getListLastName: routine to get the names of last element of a nested list.}
\usage{
getListLastName(list)
}
\arguments{
\item{list}{The list to be used.}
}
\value{
The names of last element.
}
\description{
getListLastName: routine to get the names of last element of a nested list.
}
