% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{getMinSampling}
\alias{getMinSampling}
\alias{getMinSampling,SamplingTimeConstraints-method}
\title{Get the minimal sampling times.}
\usage{
getMinSampling(object)

\S4method{getMinSampling}{SamplingTimeConstraints}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{SamplingTimeConstraints}.}
}
\value{
A numeric giving the minimal sampling times.
}
\description{
Get the minimal sampling times.
}
