% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{getModel}
\alias{getModel}
\title{Get a model of the \code{LibraryOfModels} object.}
\usage{
getModel(object, ...)
}
\arguments{
\item{object}{A \code{LibraryOfModels} object.}

\item{...}{The three-dots for passing one name or two names as arguments. One name to get a PK or PD model and two names for the PK and PD models of a PKPD model.}
}
\value{
Return a \code{Model} object giving a PK or PD model.
}
\description{
Get a model of the \code{LibraryOfModels} object.
}
