% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getModel}
\alias{getModel}
\alias{getModel,PFIMProject-method}
\title{Get the model.}
\usage{
getModel(object)

\S4method{getModel}{PFIMProject}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{PFIMProject}.}
}
\value{
The model of the object.
}
\description{
Get the model.
}
