% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getModelError}
\alias{getModelError}
\title{Get the model error.}
\usage{
getModelError(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
The model error of the object.
}
\description{
Get the model error.
}
