% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Model.R, R/PFIMProject.R
\name{getModelError}
\alias{getModelError}
\alias{getModelError,Model-method}
\alias{getModelError,PFIMProject-method}
\title{Get the model error.}
\usage{
getModelError(object)

\S4method{getModelError}{Model}(object)

\S4method{getModelError}{PFIMProject}(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
The model error of the object.
}
\description{
Get the model error.
}
