% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getModelErrorParametersValues}
\alias{getModelErrorParametersValues}
\title{Get the values of the model error parameters.}
\usage{
getModelErrorParametersValues(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
A list giving the values of the model error parameters.
}
\description{
Get the values of the model error parameters.
}
