% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getModelParameters}
\alias{getModelParameters}
\title{Get the model parameters.}
\usage{
getModelParameters(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{PFIMProject}.}
}
\value{
A list giving the model parameters.
}
\description{
Get the model parameters.
}
