% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{getModelParametersData}
\alias{getModelParametersData}
\title{getModelParametersData: get model parameters data for report.}
\arguments{
\item{modelParameter}{An object if class \code{Model} giving the model.}
}
\value{
A data frame with the data of all the parameters.
}
\description{
getModelParametersData: get model parameters data for report.
}
