% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Distribution.R,
%   R/ModelParameter.R
\name{getMu}
\alias{getMu}
\alias{getMu,Distribution-method}
\alias{getMu,ModelParameter-method}
\title{getMu}
\usage{
getMu(object)

\S4method{getMu}{Distribution}(object)

\S4method{getMu}{ModelParameter}(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
The object with the updated fixed effect.
}
\description{
Get the fixed effect of an object.
}
