% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{getMu}
\alias{getMu}
\title{Get \code{mu} for a \code{ModelParameter} object.}
\usage{
getMu(object)
}
\arguments{
\item{object}{\code{ModelParameter} object.}
}
\value{
A numeric \code{mu} giving the value of the mean \code{mu} for a \code{ModelParameter} object.
}
\description{
Get \code{mu} for a \code{ModelParameter} object.
}
