% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Arm.R, R/Design.R,
%   R/ModelParameter.R, R/LibraryOfModels.R, R/Model.R, R/PFIMProject.R
\name{getName}
\alias{getName}
\alias{getName,Arm-method}
\alias{getName,Design-method}
\alias{getName,ModelParameter-method}
\alias{getName,LibraryOfModels-method}
\alias{getName,Model-method}
\alias{getName,PFIMProject-method}
\title{getName}
\usage{
getName(object)

\S4method{getName}{Arm}(object)

\S4method{getName}{Design}(object)

\S4method{getName}{ModelParameter}(object)

\S4method{getName}{LibraryOfModels}(object)

\S4method{getName}{Model}(object)

\S4method{getName}{PFIMProject}(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
A character string \code{name} giving the name of the object.
}
\description{
Get the name of an object
}
