% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Model.R, R/PFIMProject.R
\name{getOdeSolverParameters}
\alias{getOdeSolverParameters}
\alias{getOdeSolverParameters,Model-method}
\alias{getOdeSolverParameters,PFIMProject-method}
\title{getOdeSolverParameters}
\usage{
getOdeSolverParameters(object)

\S4method{getOdeSolverParameters}{Model}(object)

\S4method{getOdeSolverParameters}{PFIMProject}(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
The list giving the parameters for the ode solvers.
}
\description{
Get the parameters for the ode solvers of an object.
}
