% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getOmega}
\alias{getOmega}
\title{Get the matrix omega of an object.}
\usage{
getOmega(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
The matrix omega of an object.
}
\description{
Get the matrix omega of an object.
}
