% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Distribution.R,
%   R/ModelParameter.R
\name{getOmega}
\alias{getOmega}
\alias{getOmega,Distribution-method}
\alias{getOmega,ModelParameter-method}
\title{Get the matrix omega of an object.}
\usage{
getOmega(object)

\S4method{getOmega}{Distribution}(object)

\S4method{getOmega}{ModelParameter}(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
The matrix omega of an object.
}
\description{
Get the matrix omega of an object.
}
