% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getOptimizer}
\alias{getOptimizer}
\alias{getOptimizer,PFIMProject-method}
\title{Get the optimization algorithm.}
\usage{
getOptimizer(object)

\S4method{getOptimizer}{PFIMProject}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{PFIMProject}.}
}
\value{
A string giving the name of the optimization algorithm.
}
\description{
Get the optimization algorithm.
}
