% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getOutcome}
\alias{getOutcome}
\title{Get the outcome of an object.}
\usage{
getOutcome(object)
}
\arguments{
\item{object}{An object defined from a class of PFIM.}
}
\value{
A string giving the outcome of the object.
}
\description{
Get the outcome of an object.
}
