% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{getOutcomesGradient}
\alias{getOutcomesGradient}
\alias{getOutcomesGradient,Design-method}
\title{getOutcomesGradient}
\usage{
getOutcomesGradient(object)

\S4method{getOutcomesGradient}{Design}(object)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}
}
\value{
The list \code{outcomesGradient} containing the results of the design evaluation for the outcome gradients.
}
\description{
Get the results of the evaluation of the outcome gradients.
}
