% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getParameters}
\alias{getParameters}
\title{Get the parameters of an object.}
\usage{
getParameters(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
Return the list of the parameters of the object.
}
\description{
Get the parameters of an object.
}
