% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/ModelError.R,
%   R/Distribution.R, R/Model.R
\name{getParameters}
\alias{getParameters}
\alias{getParameters,ModelError-method}
\alias{getParameters,Distribution-method}
\alias{getParameters,Model-method}
\title{Get the parameters of an object.}
\usage{
getParameters(object)

\S4method{getParameters}{ModelError}(object)

\S4method{getParameters}{Distribution}(object)

\S4method{getParameters}{Model}(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
Return the list of the parameters of the object.
}
\description{
Get the parameters of an object.
}
