% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKPDModels.R
\name{getPDModel}
\alias{getPDModel}
\alias{getPDModel,LibraryOfPKPDModels-method}
\title{Get a PD model.}
\usage{
getPDModel(object, PDModelName)

\S4method{getPDModel}{LibraryOfPKPDModels}(object, PDModelName)
}
\arguments{
\item{object}{An object from the class \linkS4class{LibraryOfPKPDModels}.}

\item{PDModelName}{A string giving the name of the PD model.}
}
\value{
Return a PD model.
}
\description{
Get a PD model.
}
