% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKPDModels.R
\name{getPKModel}
\alias{getPKModel}
\alias{getPKModel,LibraryOfPKPDModels-method}
\title{Get a PK model.}
\usage{
getPKModel(object, PKModelName)

\S4method{getPKModel}{LibraryOfPKPDModels}(object, PKModelName)
}
\arguments{
\item{object}{An object from the class \linkS4class{LibraryOfPKPDModels}.}

\item{PKModelName}{A string giving the name of the PK model.}
}
\value{
Return a PK model.
}
\description{
Get a PK model.
}
