% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKPDModels.R
\name{getPKPDModel}
\alias{getPKPDModel}
\title{Get a PKPD model.}
\usage{
getPKPDModel(object, namesModel)
}
\arguments{
\item{object}{An object from the class \linkS4class{LibraryOfPKPDModels}.}

\item{namesModel}{A vector of strings giving the names of the PK and PD models.}
}
\value{
Return a PKPD model.
}
\description{
Get a PKPD model.
}
