% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{getProportionsOfSubjects}
\alias{getProportionsOfSubjects}
\alias{getProportionsOfSubjects,Optimization-method}
\title{Get the proportion of subjects.}
\usage{
getProportionsOfSubjects(object)

\S4method{getProportionsOfSubjects}{Optimization}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Optimization}.}
}
\value{
A vector giving the proportion of subjects.
}
\description{
Get the proportion of subjects.
}
