% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getRSE}
\alias{getRSE}
\title{Get the RSE}
\usage{
getRSE(object, model)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{model}{An object from the class \linkS4class{Model}.}
}
\value{
A vector giving the RSE.
}
\description{
Get the RSE
}
