% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getRSE}
\alias{getRSE}
\title{getRSE: get the RSE}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The RSE of the parameters.

The RSE
}
\description{
getRSE: get the RSE

getRSE: get the RSE
}
