% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{getRSE}
\alias{getRSE}
\alias{getRSE,BayesianFim-method}
\alias{getRSE,IndividualFim-method}
\alias{getRSE,PopulationFim-method}
\title{Get the RSE}
\usage{
getRSE(object, model)

\S4method{getRSE}{BayesianFim}(object, model)

\S4method{getRSE}{IndividualFim}(object, model)

\S4method{getRSE}{PopulationFim}(object, model)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{model}{An object from the class \linkS4class{Model}.}
}
\value{
A vector giving the RSE.
}
\description{
Get the RSE
}
