% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getRSE}
\alias{getRSE}
\title{getRSE}
\usage{
getRSE(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of RSE (\%) values for each model parameter.
}
\description{
Retrieves the Relative Standard Errors (RSE, \%) from the FIM.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{
# Extract RSE (\%) for all model parameters for evaluationPopulationFIMResults
rse = getRSE(evaluationPopulationFIMResults)
# Display the RSE values
print(rse)
}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
