% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getSamplingData}
\alias{getSamplingData}
\title{getSamplingData: extract sampling times and max sampling time used for plot.}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}
}
\value{
A list giving the \code{samplingTimes} object, the vector samplings and the double samplingMax.
}
\description{
getSamplingData: extract sampling times and max sampling time used for plot.
}
