% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/SamplingTimeConstraints.R,
%   R/SamplingTimes.R
\name{getSamplings}
\alias{getSamplings}
\alias{getSamplings,SamplingTimeConstraints-method}
\alias{getSamplings,SamplingTimes-method}
\title{Get the sampling of an object.}
\usage{
getSamplings(object)

\S4method{getSamplings}{SamplingTimeConstraints}(object)

\S4method{getSamplings}{SamplingTimes}(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
A list of the samplings of the object.
}
\description{
Get the sampling of an object.
}
