% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getSamplings}
\alias{getSamplings}
\title{Get the vectors of sampling times for an arm.}
\usage{
getSamplings(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
A list of objects \code{samplings} from the class \code{SamplingTimes} giving the vector.
of sampling times for the object \code{Arm}.
}
\description{
Get the vectors of sampling times for an arm.
}
