% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getSamplingTimesConstraints}
\alias{getSamplingTimesConstraints}
\alias{getSamplingTimesConstraints,Arm-method}
\title{getSamplingTimesConstraints}
\usage{
getSamplingTimesConstraints(object)

\S4method{getSamplingTimesConstraints}{Arm}(object)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}
}
\value{
The list \code{getSamplingTimesConstraints}.
}
\description{
Get the sampling times constraints.
}
