% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getSE}
\alias{getSE}
\title{Get the Standard Errors for a \code{Fim} object..}
\usage{
getSE(object)
}
\arguments{
\item{object}{A \code{Fim} object.}
}
\value{
A vector of numerical values giving the Standard Errors from the Fisher Information Matrix.
}
\description{
Get the Standard Errors for a \code{Fim} object..
}
