% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getSE}
\alias{getSE}
\alias{getSE,Fim-method}
\title{Get the SE.}
\usage{
getSE(object)

\S4method{getSE}{Fim}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
A vector giving the SE.
}
\description{
Get the SE.
}
