% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getSE}
\alias{getSE}
\title{getSE: get the SE}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The SE of the parameters.

The SE.
}
\description{
getSE: get the SE

getSE: get the SE
}
