% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Fim.R, R/Evaluation.R,
%   R/Optimization.R
\name{getSE}
\alias{getSE}
\alias{getSE,Fim-method}
\alias{getSE,Evaluation-method}
\alias{getSE,Optimization-method}
\title{Get the SE.}
\usage{
getSE(object)

\S4method{getSE}{Fim}(object)

\S4method{getSE}{Evaluation}(object)

\S4method{getSE}{Optimization}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
A vector giving the SE.
}
\description{
Get the SE.
}
