% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getSE}
\alias{getSE}
\title{getSE}
\usage{
getSE(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of SE values for each model parameter.
}
\description{
Retrieves the Standard Errors (SE) from the Fisher Information Matrix.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{
# Extract Standard Errors from evaluationPopulationFIMResults
se = getSE(evaluationPopulationFIMResults)
print(se)
}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
