% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{getShrinkage}
\alias{getShrinkage}
\title{Get the shrinkage.}
\usage{
getShrinkage(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
A vector giving the shrinkage of the Bayesian fim.
}
\description{
Get the shrinkage.
}
