% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{getShrinkage}
\alias{getShrinkage}
\title{getShrinkage: get the shrinkage}
\arguments{
\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The shrinkage of the FIM.

The shrinkage
}
\description{
getShrinkage: get the shrinkage

getShrinkage: get the shrinkage
}
