% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R
\name{getShrinkage}
\alias{getShrinkage}
\title{Calculates the shrinkage of individual parameters from a \code{BayesianFim} object.}
\usage{
getShrinkage(object)
}
\arguments{
\item{object}{An object \code{BayesianFim} from the class \linkS4class{BayesianFim}.}
}
\value{
A numeric vector giving the shrinkage of individual parameters from a Bayesian matrix.
}
\description{
Calculates the shrinkage of individual parameters from a \code{BayesianFim} object.
}
