% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{getShrinkage}
\alias{getShrinkage}
\alias{getShrinkage,BayesianFim-method}
\alias{getShrinkage,IndividualFim-method}
\alias{getShrinkage,PopulationFim-method}
\title{Get the shrinkage.}
\usage{
getShrinkage(object)

\S4method{getShrinkage}{BayesianFim}(object)

\S4method{getShrinkage}{IndividualFim}(object)

\S4method{getShrinkage}{PopulationFim}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
A vector giving the shrinkage of the Bayesian fim.
}
\description{
Get the shrinkage.
}
