% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{getShrinkage}
\alias{getShrinkage}
\title{getShrinkage}
\usage{
getShrinkage(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of shrinkage values for each random effect.
}
\description{
Retrieves shrinkage values for the random effects (omega), measuring how
individual estimates are shrunk toward the population mean.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{

# Extract the shrinkage values from the Bayesian FIM evaluation results
shrinkage = getShrinkage(evaluationBayesianFIMResults)
print(shrinkage)

}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
