% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{getSigmaInter}
\alias{getSigmaInter}
\title{Get the parameter sigma inter.}
\usage{
getSigmaInter(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelError}.}
}
\value{
A numeric giving the parameter sigma inter.
}
\description{
Get the parameter sigma inter.
}
