% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Arm.R, R/Design.R
\name{getSize}
\alias{getSize}
\alias{getSize,Arm-method}
\alias{getSize,Design-method}
\title{getSize}
\usage{
getSize(object)

\S4method{getSize}{Arm}(object)

\S4method{getSize}{Design}(object)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}
}
\value{
A numeric giving the size of the object.
}
\description{
Get the size of an object.
}
