% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Administration.R
\name{getTimeDose}
\alias{getTimeDose}
\title{Get the times vector when doses are given.}
\usage{
getTimeDose(object)
}
\arguments{
\item{object}{An object \code{Administration} from the class \linkS4class{Administration}.}
}
\value{
The vector \code{timeDose} giving the times when the doses are given.
}
\description{
Get the times vector when doses are given.
}
