% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualFim.R
\docType{class}
\name{IndividualFim-class}
\alias{IndividualFim-class}
\alias{IndividualFim}
\title{Class "individualFim" representing the individual Fisher information matrix}
\description{
A class storing information regarding the individual Fisher computation matrix.
}
\section{Objects from the class}{
 IndividualFim objects are typically created by calls to \code{{fim}} or \code{{pfim}} and contain the following slots:

\describe{
\item{IndividualFim}{Create a new object \code{{Fim}}}
}
}

