% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\docType{class}
\name{LibraryOfModels-class}
\alias{LibraryOfModels-class}
\alias{LibraryOfModels}
\title{Class "LibraryOfModels"}
\description{
The class \code{LibraryOfModels} represents the library of models.
}
\section{Objects from the class}{

Objects form the class \code{LibraryOfModels} can be created by calls of the form \code{LibraryOfModels(...)} where
(...) are the parameters for the \code{LibraryOfModels} objects.
}

\section{Slots for \code{LibraryOfModels} objects}{

\describe{
\item{\code{name}:}{A string giving the name of the library of models.}
\item{\code{content}:}{A list giving the content of the library of model.}
}
}

