% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{LibraryOfModels}
\alias{LibraryOfModels}
\title{LibraryOfModels}
\usage{
LibraryOfModels(models = list())
}
\arguments{
\item{models}{A list giving all the PK and PD models.}
}
\description{
The class \code{LibraryOfModels} represents and stores information for the LibraryOfModels.
}
