% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPDModels.R
\docType{data}
\name{LibraryOfPDModels}
\alias{LibraryOfPDModels}
\title{LibraryOfPDModels}
\format{
An object of class \code{PFIM::LibraryOfPDModels} (inherits from \code{PFIM::LibraryOfModels}, \code{S7_object}) of length 1.
}
\usage{
LibraryOfPDModels
}
\description{
The class \code{LibraryOfPDModels} represents and stores information for the LibraryOfPDModels.
}
\keyword{datasets}
