% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R
\docType{data}
\name{LibraryOfPKModels}
\alias{LibraryOfPKModels}
\title{LibraryOfPKModels}
\format{
An object of class \code{PFIM::LibraryOfPKModels} (inherits from \code{PFIM::LibraryOfModels}, \code{S7_object}) of length 1.
}
\usage{
LibraryOfPKModels
}
\description{
The class \code{LibraryOfPKModels} represents and stores information for the LibraryOfPKModels.
}
\keyword{datasets}
