% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\alias{Model}
\title{Class "Model"}
\description{
The class \code{Model} defines information concerning the construction of a model.
}
\section{Objects from the class}{

Objects form the class \code{Model} can be created by calls of the form \code{Model(...)} where
(...) are the parameters for the \code{Model} objects.
}

\section{Slots for \code{Administration} objects}{

 \describe{
   \item{\code{name}:}{A string giving the name of the model.}
   \item{\code{description}:}{A list of string giving the description of the model.}
   \item{\code{equations}:}{A list giving the equations of the model.}
   \item{\code{outcomes}:}{A list giving the outcomes of the model.}
   \item{\code{outcomesForEvaluation}:}{A list giving the outcomes used for the evaluation of the model.}
   \item{\code{parameters}:}{A list giving the parameters of the model.}
   \item{\code{modelError}:}{A list giving the model error of the model.}
   \item{\code{initialConditions}:}{A list giving the initial conditions of the model.}
   \item{\code{odeSolverParameters}:}{A list giving the parameters for the solver of the model.}
   \item{\code{modelFromLibrary}:}{A list giving the model equations when the model is constructed from the library of model.}
 }
}

