% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{ModelError}
\alias{ModelError}
\title{ModelError}
\usage{
ModelError(
  output = "output",
  equation = expression(),
  derivatives = list(),
  sigmaInter = 0.1,
  sigmaSlope = 0,
  sigmaInterFixed = FALSE,
  sigmaSlopeFixed = FALSE,
  cError = 1
)
}
\arguments{
\item{output}{A string giving the model error output.}

\item{equation}{A expression giving the model error equation.}

\item{derivatives}{A list giving the derivatives of the model error equation.}

\item{sigmaInter}{A double giving the sigma inter.}

\item{sigmaSlope}{A double giving the sigma slope}

\item{sigmaInterFixed}{A boolean giving if the  sigma inter is fixed or not. - not in the v7.0}

\item{sigmaSlopeFixed}{A boolean giving if the  sigma slope is fixed or not. - not in the v7.0}

\item{cError}{A integer giving the power parameter.}
}
\description{
The class \code{ModelError} is used to defined a model error.
}
\details{
ModelError
}
