% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelODEDoseInEquations.R
\name{ModelODEDoseInEquations}
\alias{ModelODEDoseInEquations}
\title{ModelODEDoseNotInEquations}
\usage{
ModelODEDoseInEquations(
  name = character(0),
  modelParameters = list(),
  samplings = numeric(0),
  modelEquations = list(),
  wrapper = function() NULL,
  outputFormula = list(),
  outputNames = character(0),
  variableNames = character(0),
  outcomesWithAdministration = character(0),
  outcomesWithNoAdministration = character(0),
  modelError = list(),
  odeSolverParameters = list(),
  parametersForComputingGradient = list(),
  initialConditions = numeric(0),
  functionArguments = character(0),
  functionArgumentsSymbol = list(),
  modelODEDoseInEquations = function() NULL,
  solverInputs = list()
)
}
\arguments{
\item{name}{Character vector specifying the model name}

\item{modelParameters}{List of model parameters}

\item{samplings}{Numeric vector of sampling times}

\item{modelEquations}{List containing the model equations}

\item{wrapper}{Function wrapper for the model (default: function () NULL)}

\item{outputFormula}{List of output formulas}

\item{outputNames}{Character vector of output names}

\item{variableNames}{Character vector of variable names}

\item{outcomesWithAdministration}{Character vector of outcomes with administration}

\item{outcomesWithNoAdministration}{Character vector of outcomes without administration}

\item{modelError}{List defining the error model}

\item{odeSolverParameters}{List of ODE solver parameters}

\item{parametersForComputingGradient}{List of parameters for gradient computation}

\item{initialConditions}{Numeric vector of initial conditions}

\item{functionArguments}{Character vector of function arguments}

\item{functionArgumentsSymbol}{List of function argument symbols}

\item{modelODEDoseInEquations}{An object \code{modelODEDoseInEquations}.}

\item{solverInputs}{A list giving the solver inputs.}
}
\description{
The class \code{ModelODEDoseNotInEquations} is used to defined a ModelODEDoseNotInEquations
}
