% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\docType{class}
\name{ModelParameter-class}
\alias{ModelParameter-class}
\alias{ModelParameter}
\title{Class "ModelParameter"}
\description{
The class \code{ModelParameter} defines information concerning the model parameters.
}
\section{Objects from the class}{

Objects form the class \code{ModelParameter} can be created by calls of the form \code{ModelParameter(...)} where
(...) are the parameters for the \code{ModelParameter} objects.
}

\section{Slots for \code{ModelParameter} objects}{

 \describe{
   \item{\code{name}:}{A string giving the name of the parameter.}
   \item{\code{distribution}:}{An object from the class \code{Distribution} giving the distribution of the parameter.}
   \item{\code{fixedMu}:}{A boolean giving if mu is fixed or not.}
   \item{\code{fixedOmega}:}{A boolean giving if omega is fixed or not.}
 }
}

