% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R
\name{ModelParameter}
\alias{ModelParameter}
\title{ModelParameter}
\usage{
ModelParameter(
  name = character(0),
  distribution = Distribution(),
  fixedMu = FALSE,
  fixedOmega = FALSE
)
}
\arguments{
\item{name}{A string giving the name of the parameter.}

\item{distribution}{A string giving the distribution of the parameter.}

\item{fixedMu}{A Boolean setting TRUE/FALSE if the mu is estimated or not.}

\item{fixedOmega}{A Boolean setting TRUE/FALSE if the omega is estimated or not.}
}
\description{
The class \code{ModelParameter} is used to defined the model parameters.
}
\details{
ModelParameter
}
