% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\docType{class}
\name{MultiplicativeAlgorithm-class}
\alias{MultiplicativeAlgorithm-class}
\alias{MultiplicativeAlgorithm}
\alias{MultiplicativeAlgorithmClass}
\title{Class "MultiplicativeAlgorithm"}
\description{
Class "MultiplicativeAlgorithm" implements the Multiplicative algorithm.
}
\section{Objects from the class}{

Objects form the class \code{MultiplicativeAlgorithm} can be created by calls of the form \code{MultiplicativeAlgorithm(...)} where
(...) are the parameters for the \code{MultiplicativeAlgorithm} objects.
}

\section{Slots for \code{MultiplicativeAlgorithm} objects}{

\describe{
\item{\code{lambda}:}{A numeric giving the lambda parameter of the multiplicative algorithm.}
\item{\code{delta}:}{A numeric giving the delta parameter of the multiplicative algorithm.}
\item{\code{iteration_init}:}{A numeric giving the first iteration of the optimization process.}
\item{\code{iteration_fin}:}{A numeric giving the last iteration of the optimization process.}
\item{\code{FinalWeights}:}{A vector givint hte optimal weights.}
\item{\code{showProcess}:}{A boolean for showing or not the process of optimization.}
\item{\code{OptimalDesign}:}{A object from the class \code{Design}}
\item{\code{allArms}:}{A list of all arms.}
}
}

