% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{MultiplicativeAlgorithm_Rcpp}
\alias{MultiplicativeAlgorithm_Rcpp}
\title{Function MultiplicativeAlgorithm_Rcpp}
\usage{
MultiplicativeAlgorithm_Rcpp(
  fisherMatrices_input,
  numberOfFisherMatrices_input,
  weights_input,
  numberOfParameters_input,
  dim_input,
  lambda_input,
  delta_input,
  iterationInit_input
)
}
\arguments{
\item{fisherMatrices_input}{The parameter fotfisherMatrices_input.}

\item{numberOfFisherMatrices_input}{The parameter numberOfFisherMatrices_input.}

\item{weights_input}{The parameter weights_input.}

\item{numberOfParameters_input}{The parameter numberOfParameters_input.}

\item{dim_input}{The parameter dim_input.}

\item{lambda_input}{The parameter lambda_input.}

\item{delta_input}{The parameter delta_input.}

\item{iterationInit_input}{The parameter iterationInit_input.}
}
\value{
The list output with the outputs of the MultiplicativeAlgorithm_Rcpp.
}
\description{
Run the MultiplicativeAlgorithm_Rcpp in Rcpp.
}
