% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{MultiplicativeAlgorithm_Rcpp}
\alias{MultiplicativeAlgorithm_Rcpp}
\title{Function MultiplicativeAlgorithm_Rcpp}
\usage{
MultiplicativeAlgorithm_Rcpp(
  fisherMatrices_input,
  numberOfFisherMatrices_input,
  weights_input,
  numberOfParameters_input,
  dim_input,
  lambda_input,
  delta_input,
  iterationInit_input
)
}
\arguments{
\item{fisherMatrices_input}{fisherMatrices_input}

\item{numberOfFisherMatrices_input}{numberOfFisherMatrices_input}

\item{weights_input}{weights_input}

\item{numberOfParameters_input}{numberOfParameters_input}

\item{dim_input}{dim_input}

\item{lambda_input}{lambda_input}

\item{delta_input}{delta_input}

\item{iterationInit_input}{iterationInit_input}
}
\description{
Run the MultiplicativeAlgorithm_Rcpp in Rcpp
}
