% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationAlgorithm.R,
%   R/FedorovWynnAlgorithm.R, R/MultiplicativeAlgorithm.R, R/PGBOAlgorithm.R,
%   R/PSOAlgorithm.R, R/SimplexAlgorithm.R
\name{optimize}
\alias{optimize}
\alias{optimize,FedorovWynnAlgorithm-method}
\alias{optimize,MultiplicativeAlgorithm-method}
\alias{optimize,PGBOAlgorithm-method}
\alias{optimize,PSOAlgorithm-method}
\alias{optimize,SimplexAlgorithm-method}
\title{Optimize a design.}
\usage{
optimize(object, optimizerParameters, optimizationObject)

\S4method{optimize}{FedorovWynnAlgorithm}(object, optimizerParameters, optimizationObject)

\S4method{optimize}{MultiplicativeAlgorithm}(object, optimizerParameters, optimizationObject)

\S4method{optimize}{PGBOAlgorithm}(object, optimizationObject)

\S4method{optimize}{PSOAlgorithm}(object, optimizationObject)

\S4method{optimize}{SimplexAlgorithm}(object, optimizerParameters, optimizationObject)
}
\arguments{
\item{object}{An object from the class \linkS4class{OptimizationAlgorithm}.}

\item{optimizerParameters}{A list giving the optimization parameters.}

\item{optimizationObject}{An object giving the optimization algorithm.}
}
\value{
A list giving the results if the optimization.
}
\description{
Optimize a design.
}
