% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGBOAlgorithm.R, R/PSOAlgorithm.R,
%   R/SimplexAlgorithm.R, R/FedorovWynnAlgorithm.R, R/MultiplicativeAlgorithm.R
\name{optimizeDesign}
\alias{optimizeDesign}
\title{Optimization PGBOAlgorithm}
\arguments{
\item{optimizationObject}{A object \code{Optimization}.}

\item{optimizationAlgorithm}{A object \code{MultiplicativeAlgorithm}.}
}
\value{
The object \code{optimizationObject} with the slots updated.

The object \code{optimizationObject} with the slots updated.

The object \code{optimizationObject} with the slots updated.

The object \code{optimizationObject} with the slots updated.

The object \code{optimizationObject} with the slots updated.
}
\description{
Optimization PGBOAlgorithm

Optimization PSOAlgorithm

Optimization SimplexAlgorithm

Optimization FedorovWynnAlgorithm

Optimization MultiplicativeAlgorithm
}
