% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{OptimizeDesign}
\alias{OptimizeDesign}
\title{Optimize the designs for each arms.}
\usage{
OptimizeDesign(object, optimizer, typeOfFim)
}
\arguments{
\item{object}{A \code{PFIMProject} object.}

\item{optimizer}{A \code{Optimization} object.}

\item{typeOfFim}{A character string giving the type of Fisher Information Matrix (Population, Indidivual or Bayesian).}
}
\value{
The \code{PFIMProject} object with the optimized designs for each arms.
}
\description{
Optimize the designs for each arms.
}
