% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{parametersForComputingGradient}
\alias{parametersForComputingGradient}
\title{Parameters used for computing the model gradient by finite-differences.}
\usage{
parametersForComputingGradient(object)
}
\arguments{
\item{object}{A \code{StatisticalModel} object.}
}
\value{
A list containing the parameters used for computing the gradient of the model.
}
\description{
Parameters used for computing the model gradient by finite-differences.
}
