% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\docType{class}
\name{PFIMProject-class}
\alias{PFIMProject-class}
\alias{PFIMProject}
\title{Class "PFIMProject"}
\description{
A class storing information concerning a PFIM project.
}
\section{Objects from the class}{

Objects form the class \code{PFIMProject} can be created by calls of the form \code{PFIMProject(...)} where
(...) are the parameters for the \code{PFIMProject} objects.
}

\section{Slots for \code{PFIMProject} objects}{

 \describe{
   \item{\code{name}:}{A character string giving the name of the PFIM project.}
   \item{\code{description}:}{A list giving the description of the PFIM project.}
 }
}

