% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEvaluation.R
\name{plot}
\alias{plot}
\alias{plot,Evaluation-method}
\title{Graphs of the results of the evaluation.}
\usage{
plot(object, plotOptions)

\S4method{plot}{Evaluation}(object, plotOptions)
}
\arguments{
\item{object}{An object from the class \linkS4class{Evaluation}.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
A list giving the graphs for the evaluation of the responses and sensitivity indices.
}
\description{
Graphs of the results of the evaluation.
}
