% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{plotEvaluation}
\alias{plotEvaluation}
\title{plotEvaluation: plots for the evaluation of the model responses.}
\arguments{
\item{pfimproject}{A object \code{PFIMProject}.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
All the plots for the evaluation of the model responses.
}
\description{
plotEvaluation: plots for the evaluation of the model responses.
}
