% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotEvaluation}
\alias{plotEvaluation}
\title{plotEvaluation}
\usage{
plotEvaluation(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{...}{Additional arguments.}
}
\value{
A named list of plots per design.
}
\description{
Generates graphical representations of model responses based on the design
and parameters defined in the PFIM project.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{
# Plot the model responses from evaluationPopulationFIMResults
plotEvaluation(evaluationPopulationFIMResults, plotOptions = plotOptions)
}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
