% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{plotEvaluationResults}
\alias{plotEvaluationResults}
\title{plotEvaluationResults: process for the evaluation of the responses.}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{evaluationModel}{A list giving the evaluation of the model.}

\item{outputNames}{A list of string giving the output of the evaluation of the model.}

\item{samplingData}{A list giving the sampling data from the method getSamplingData.}

\item{unitXAxis}{A list giving the unit of the x-axis.}

\item{unitYAxis}{A list giving the unit of the y-axis.}

\item{designName}{A string giving the design name.}
}
\value{
A list giving the plot of the evaluation of the model responses.
}
\description{
plotEvaluationResults: process for the evaluation of the responses.
}
