% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{plotEvaluationResults}
\alias{plotEvaluationResults}
\title{Plot model responses for an arm}
\usage{
plotEvaluationResults(
  arm,
  evaluationModel,
  outputNames,
  samplingData,
  designName,
  plotOptions
)
}
\arguments{
\item{arm}{An object of class \code{Arm}.}

\item{evaluationModel}{A list of data frames from \code{evaluateModel}.}

\item{outputNames}{A list of strings giving the output names.}

\item{samplingData}{A list from \code{getSamplingData}.}

\item{designName}{A string giving the design name.}

\item{plotOptions}{A list with \code{unitTime} and \code{unitOutcomes}.}
}
\value{
A named list of \code{ggplot} objects.
}
\description{
Generates \code{ggplot2} line plots of model responses, overlaying design
sampling points in red on the secondary x-axis.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
