% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{plotEvaluationSI}
\alias{plotEvaluationSI}
\title{plotEvaluationSI: process for the evaluation of the gradient of the responses.}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{evaluationModelGradient}{A list giving the evaluation of the gradient of the model responses.}

\item{parametersNames}{A vector of string giving the parameter names?}

\item{outputNames}{A list of string giving the name of the outputs.}

\item{samplingData}{A list giving the sampling data from the method getSamplingData.}

\item{unitXAxis}{A list giving the unit of the x-axis.}

\item{unitYAxis}{A list giving the unit of the y-axis.}

\item{designName}{A string giving the design name.}
}
\value{
A list giving the plot of the evaluation of gradient of the model responses.
}
\description{
plotEvaluationSI: process for the evaluation of the gradient of the responses.
}
