% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationAlgorithm.R,
%   R/FedorovWynnAlgorithm.R, R/Optimization.R
\name{plotFrequencies}
\alias{plotFrequencies}
\alias{plotFrequencies,FedorovWynnAlgorithm-method}
\alias{plotFrequencies,Optimization-method}
\title{Graph of the frequencies for the FW algorithm.}
\usage{
plotFrequencies(object)

\S4method{plotFrequencies}{FedorovWynnAlgorithm}(object)

\S4method{plotFrequencies}{Optimization}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{OptimizationAlgorithm}.}
}
\value{
The graphs of the frequencies for the FW algorithm.
}
\description{
Graph of the frequencies for the FW algorithm.
}
