% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{plotFrequencies}
\alias{plotFrequencies}
\title{Plot frequencies for the FedorovWynn algorithm}
\arguments{
\item{optimization}{An \code{Optimization} object.}
}
\value{
Graph of the optimal frequencies.
}
\description{
Plot frequencies for the FedorovWynn algorithm
}
