% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{plotFrequenciesFedorovWynnAlgorithm}
\alias{plotFrequenciesFedorovWynnAlgorithm}
\title{plotFrequenciesFedorovWynnAlgorithm}
\arguments{
\item{optimization}{optimization}

\item{optimizationAlgorithm}{optimizationAlgorithm}
}
\value{
plotFrequenciesFedorovWynnAlgorithm
}
\description{
plotFrequenciesFedorovWynnAlgorithm
}
