% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{plotOutcomesEvaluation}
\alias{plotOutcomesEvaluation}
\alias{plotOutcomesEvaluation,Design-method}
\title{plotOutcomesEvaluation}
\usage{
plotOutcomesEvaluation(
  object,
  outcomesEvaluationInitialDesign,
  model,
  plotOptions
)

\S4method{plotOutcomesEvaluation}{Design}(
  object,
  outcomesEvaluationInitialDesign,
  model,
  plotOptions
)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}

\item{outcomesEvaluationInitialDesign}{A list containing the evaluation of the initial design.}

\item{model}{An object \code{model} from the class \linkS4class{Model}.}

\item{plotOptions}{A list containing the plot options.}
}
\value{
A list containing the plots the evaluation of the outcomes.
}
\description{
Plot the evaluation of the outcomes.
}
