% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{plotRSE}
\alias{plotRSE}
\title{Plot relative standard errors}
\arguments{
\item{optimization}{An \code{Optimization} object.}

\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
Graph of relative standard errors

The bar plot of the RSE.
}
\description{
Plot relative standard errors

plotRSE: bar plot of the RSE.
}
