% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEvaluation.R
\name{plotRSE}
\alias{plotRSE}
\alias{plotRSE,PFIMProject-method}
\title{Graph of the RSE.}
\usage{
plotRSE(object, plotOptions)

\S4method{plotRSE}{PFIMProject}(object, plotOptions)
}
\arguments{
\item{object}{An object from the class \linkS4class{Evaluation}.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
A graph of the RSE.
}
\description{
Graph of the RSE.
}
