% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotRSE}
\alias{plotRSE}
\title{Plot the relative standard errors RSE of the model parameters.}
\usage{
plotRSE(object)
}
\arguments{
\item{object}{\code{PFIMProject} object.}
}
\value{
A list containing the plots of the RSE of the model parameters.
}
\description{
Plot the relative standard errors RSE of the model parameters.
}
