% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotRSE}
\alias{plotRSE}
\title{plotRSE}
\usage{
plotRSE(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{...}{Additional arguments.}
}
\value{
A bar plot of RSE (\%) values.
}
\description{
Bar plot of Relative Standard Errors (RSE, \%) for the model parameters.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{
# Extract Relative Standard Errors from evaluationPopulationFIMResults
se = getSE(evaluationPopulationFIMResults)
print(se)
}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
