% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{plotRSEFIM}
\alias{plotRSEFIM}
\title{plotRSEFIM: barplot for the RSE}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{evaluation}{An object \code{Evaluation} giving the evaluation of the model.}
}
\value{
The bar plot of the RSE.

The bar plot of the RSE.

The bar plot of the RSE.
}
\description{
plotRSEFIM: barplot for the RSE

plotRSEFIM: barplot for the RSE

plotRSEFIM: barplot for the RSE
}
