% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationFim.R
\name{plotRSEFIM}
\alias{plotRSEFIM}
\title{plotRSEFIM: barplot for the RSE}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{evaluation}{An object \code{Evaluation} giving the evaluation of the model.}
}
\value{
A ggplot bar chart of RSE by parameter type.
}
\description{
plotRSEFIM: barplot for the RSE
}
