% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{plotSE}
\alias{plotSE}
\title{Plot standard errors}
\arguments{
\item{optimization}{An \code{Optimization} object.}

\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
Graph of standard errors

The bar plot of the SE.
}
\description{
Plot standard errors

plotSE: bar plot of the SE.
}
