% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotSE}
\alias{plotSE}
\title{Plot the standard errors SE of the model parameters.}
\usage{
plotSE(object)
}
\arguments{
\item{object}{\code{PFIMProject} object.}
}
\value{
A list containing the plots of the standard errors SE of the model parameters.
}
\description{
Plot the standard errors SE of the model parameters.
}
