% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEvaluation.R
\name{plotSE}
\alias{plotSE}
\alias{plotSE,PFIMProject-method}
\title{Graph the SE.}
\usage{
plotSE(object, plotOptions)

\S4method{plotSE}{PFIMProject}(object, plotOptions)
}
\arguments{
\item{object}{An object from the class \linkS4class{Evaluation}.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
A graph of the SE.
}
\description{
Graph the SE.
}
