% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotSE}
\alias{plotSE}
\title{plotSE}
\usage{
plotSE(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{...}{Additional arguments.}
}
\value{
A bar plot of SE values.
}
\description{
Bar plot of Standard Errors (SE) for fixed effects and variance components.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{
# plotSE from evaluationPopulationFIMResults
plotSE( evaluationPopulationFIMResults )
}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
