% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{plotSEFIM}
\alias{plotSEFIM}
\title{plotSEFIM: barplot for the SE}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{evaluation}{An object \code{Evaluation} giving the evaluation of the model.}
}
\value{
The bar plot of the SE.

The bar plot of the SE.

The bar plot of the SE.
}
\description{
plotSEFIM: barplot for the SE

plotSEFIM: barplot for the SE

plotSEFIM: barplot for the SE
}
