% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/PopulationFim.R
\name{plotSEFIM}
\alias{plotSEFIM}
\title{plotSEFIM / plotRSEFIM / plotShrinkage: bar charts for SE, RSE, shrinkage}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{evaluation}{An object \code{Evaluation} giving the evaluation of the model.}
}
\value{
A ggplot bar chart of SE by parameter type.
}
\description{
plotSEFIM / plotRSEFIM / plotShrinkage: bar charts for SE, RSE, shrinkage

plotSEFIM: barplot for the SE
}
