% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{plotSensitivityIndices}
\alias{plotSensitivityIndices}
\title{Plot sensitivity indices.}
\arguments{
\item{optimization}{An \code{Optimization} object.}

\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
Graph of sensitivity indices.

All the plots for the evaluation of the gradient of the model responses.
}
\description{
Plot sensitivity indices.

plotSensitivityIndices:  plots for the evaluation of the gradient of the model responses.
}
