% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotSensitivityIndices}
\alias{plotSensitivityIndices}
\title{plotSensitivityIndices}
\usage{
plotSensitivityIndices(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{...}{Additional arguments.}
}
\value{
A named list of sensitivity index plots per design.
}
\description{
Generates sensitivity index plots (partial derivatives of model responses
with respect to population parameters).
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{
plot the sensitivity indices from evaluationPopulationFIMResults
plotSensitivityIndices( evaluationPopulationFIMResults, plotOptions = plotOptions )
}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
