% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R
\name{plotShrinkage}
\alias{plotShrinkage}
\title{plotShrinkage: plot the shrinkage values.}
\arguments{
\item{fim}{An object \code{BayesianFim} giving the Fim.}

\item{evaluation}{An object \code{Evaluation} giving the evaluation of the model.}
}
\value{
The bar plot of the shrinkage.
}
\description{
plotShrinkage: plot the shrinkage values.
}
