% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{plotShrinkage}
\alias{plotShrinkage}
\title{Plot the shrinkage data.}
\usage{
plotShrinkage(object)
}
\arguments{
\item{object}{\code{PFIMProject} object.}
}
\value{
A list containing the plots of the shrinkage of the model parameters.
}
\description{
Plot the shrinkage data.
}
