% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEvaluation.R
\name{plotShrinkage}
\alias{plotShrinkage}
\title{Graph of the shrinkage.}
\usage{
plotShrinkage(object, plotOptions)
}
\arguments{
\item{object}{An object from the class \linkS4class{Evaluation}.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
A graph of the shrinkage.
}
\description{
Graph of the shrinkage.
}
