% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R
\name{plotShrinkage}
\alias{plotShrinkage}
\title{Bar plot of the Bayesian shrinkage}
\arguments{
\item{fim}{An object of class \code{BayesianFim} giving the Fim.}

\item{evaluation}{An object of class \code{PFIMProject} giving the evaluation.}
}
\value{
A \code{ggplot2} object representing the shrinkage bar plot.
}
\description{
Bar plot of the Bayesian shrinkage
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
