% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R
\name{plotShrinkage}
\alias{plotShrinkage}
\title{plotShrinkage: Bar plot for Bayesian Shrinkage}
\arguments{
\item{fim}{An object \code{BayesianFim}.}

\item{evaluation}{An object \code{PFIMProject}.}
}
\value{
A ggplot object representing the shrinkage values.
}
\description{
plotShrinkage: Bar plot for Bayesian Shrinkage
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
