% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{plotWeights}
\alias{plotWeights}
\title{Plot weights for the multiplicative algorithm}
\arguments{
\item{optimization}{An \code{Optimization} object.}
}
\value{
Plot of weights
}
\description{
Plot weights for the multiplicative algorithm
}
