% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{plotWeightsMultiplicativeAlgorithm}
\alias{plotWeightsMultiplicativeAlgorithm}
\title{plotWeightsMultiplicativeAlgorithm: plot the optimal weight.}
\arguments{
\item{optimization}{A object \code{Optimization}.}

\item{optimizationAlgorithm}{A object \code{MultiplicativeAlgorithm}.}
}
\value{
The graph plotWeight.
}
\description{
plotWeightsMultiplicativeAlgorithm: plot the optimal weight.
}
