% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationFim.R
\docType{class}
\name{PopulationFim-class}
\alias{PopulationFim-class}
\alias{PopulationFim}
\title{Class "PopulationFim"}
\description{
Class "PopulationFim" representing the population Fisher information matrix

A class storing information regarding the population Fisher computation matrix
(computation method: first order linearisation (FO))
}
\section{Objects from the class \code{PopulationFim}}{

Objects form the class \code{PopulationFim} can be created by calls of the form \code{PopulationFim(...)} where
(...) are the parameters for the \code{PopulationFim} objects.
}

\section{Slots for \code{PopulationFim} objects}{

 \describe{
   \item{\code{mfisher}:}{A matrix giving the Fisher Information matrix.}
 }
}

