% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationFim.R
\name{PopulationFim}
\alias{PopulationFim}
\title{PopulationFim}
\usage{
PopulationFim(
  fisherMatrix = numeric(0),
  fixedEffects = numeric(0),
  varianceEffects = numeric(0),
  SEAndRSE = list(),
  condNumberFixedEffects = 0,
  condNumberVarianceEffects = 0,
  shrinkage = numeric(0)
)
}
\arguments{
\item{fisherMatrix}{A matrix giving the numerical values of the Fim.}

\item{fixedEffects}{A matrix giving the numerical values of the fixedEffects of the Fim.}

\item{varianceEffects}{A matrix giving the numerical values of varianceEffects of the Fim.}

\item{SEAndRSE}{A data frame giving the value of the SE and RSE.}

\item{condNumberFixedEffects}{The conditional number of the fixedEffects of the Fim.}

\item{condNumberVarianceEffects}{The conditional number of the varianceEffects of the Fim.}

\item{shrinkage}{A vector giving the shrinkage values.}
}
\description{
The class \code{PopulationFim} represents and stores information for the PopulationFim.
}
