% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{processArmEvaluationResults}
\alias{processArmEvaluationResults}
\title{processArmEvaluationResults: process for the evaluation of an arm.}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{model}{A object of class \code{Model} giving the model.}

\item{fim}{A object of class \code{Fim} giving the fim.}

\item{designName}{A string giving the name of the design.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
A list of ggplot object giving the plot of the responses ans the gradient responses of the the model.
}
\description{
processArmEvaluationResults: process for the evaluation of an arm.
}
