% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{processArmEvaluationSI}
\alias{processArmEvaluationSI}
\title{processArmEvaluationSI: process for the evaluation of the gradient of the responses.}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{model}{A object of class \code{Model} giving the model.}

\item{fim}{A object of class \code{Fim} giving the fim.}

\item{designName}{A string giving the name of the design.}
}
\value{
A list giving the ggplot object of the plots of the gradient.
}
\description{
processArmEvaluationSI: process for the evaluation of the gradient of the responses.
}
