% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOAlgorithm.R
\docType{class}
\name{PSOAlgorithm-class}
\alias{PSOAlgorithm-class}
\alias{PSOAlgorithm}
\title{Class "PSOAlgorithm"}
\description{
The class "PSOAlgorithm" implements the PSO algorithm.
}
\section{Objects from the class \code{PSOAlgorithm}}{

Objects form the class \code{PSOAlgorithm} can be created by calls of the form \code{PSOAlgorithm(...)} where
(...) are the parameters for the \code{PSOAlgorithm} objects.
}

\section{Slots for \code{PSOAlgorithm} objects}{

 \describe{
   \item{\code{maxIteration}:}{A numeric giving the maximum of iterations.}
   \item{\code{populationSize}:}{A numeric giving the population size.}
   \item{\code{seed}:}{A numeric giving the seed.}
   \item{\code{personalLearningCoefficient}:}{A numeric giving the personal learning coefficient.}
   \item{\code{globalLearningCoefficient}:}{A numeric giving the global learning coefficient.}
   \item{\code{showProcess}:}{A boolean to show or not the process.}
   \item{\code{optimalDesign}:}{A \code{Design} object giving the optimal design.}
   \item{\code{iterationAndCriteria}:}{A list giving the optimal criteria at each iteration.}
 }
}

