% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{replaceVariablesLibraryOfModels}
\alias{replaceVariablesLibraryOfModels}
\title{replaceVariablesLibraryOfModels: replace variable in the LibraryOfModels}
\usage{
replaceVariablesLibraryOfModels(text, old, new)
}
\arguments{
\item{text}{the text}

\item{old}{old string}

\item{new}{new string}
}
\value{
text with new string
}
\description{
replaceVariablesLibraryOfModels: replace variable in the LibraryOfModels
}
