% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{Report}
\alias{Report}
\title{Generate optimization report}
\arguments{
\item{optimization}{An \code{Optimization} object.}

\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation or Optimization.}

\item{outputPath}{A string giving the path where the output are saved.}

\item{outputFile}{A string giving the name of the output file.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
Generated report.

The html report of the design evaluation or optimization.
}
\description{
Generate optimization report

Report: generate the report.
}
