% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R, R/Evaluation.R, R/Optimization.R
\name{Report}
\alias{Report}
\alias{Report,Evaluation-method}
\alias{Report,Optimization-method}
\title{Report}
\usage{
Report(object, outputPath, outputFile, plotOptions)

\S4method{Report}{Evaluation}(object, outputPath, outputFile, plotOptions)

\S4method{Report}{Optimization}(object, outputPath, outputFile, plotOptions)
}
\arguments{
\item{object}{An object from the class \linkS4class{PFIMProject}.}

\item{outputPath}{A string giving the output path.}

\item{outputFile}{A string giving the name of the output file.}

\item{plotOptions}{A list giving the plot options.}
}
\value{
The report in html.
}
\description{
Report
}
