% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{Report}
\alias{Report}
\title{Report}
\usage{
Report(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject}.}

\item{...}{Additional arguments: outputPath, outputFile, plotOptions}
}
\value{
Generates and saves an HTML report to \code{outputPath/outputFile}.
}
\description{
Creates a detailed HTML report from the design evaluation results,
including tables, matrices, and plots.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{

# Examples from Vignette 1 and 2

# Generate a comprehensive HTML report for the design evaluation
Report(
  pfimproject = evaluationPopulationFIMResults,
  outputPath  = "C:/MyResults",
  outputFile  = "Design_Evaluation_Report.html",
  plotOptions = plotOptions
)

}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
