% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{reportTablesAdministration}
\alias{reportTablesAdministration}
\alias{reportTablesAdministration,Design-method}
\title{reportTablesAdministration}
\usage{
reportTablesAdministration(object)

\S4method{reportTablesAdministration}{Design}(object)
}
\arguments{
\item{object}{An object \code{design} from the class \linkS4class{Design}.}
}
\value{
A table of the administration parameters for the report.
}
\description{
Generate table for the report.
}
