% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{reportTablesModelParameters}
\alias{reportTablesModelParameters}
\title{Generate the tables for model parameters for the evaluation report.}
\usage{
reportTablesModelParameters(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}
}
\value{
A kable table for the evaluation report.
}
\description{
Generate the tables for model parameters for the evaluation report.
}
