% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{reportTablesPlot}
\alias{reportTablesPlot}
\alias{reportTablesPlot,Evaluation-method}
\title{reportTablesPlot}
\usage{
reportTablesPlot(object, plotOptions)

\S4method{reportTablesPlot}{Evaluation}(object, plotOptions)
}
\arguments{
\item{object}{An object \code{evaluation} from the class \linkS4class{Evaluation}.}

\item{plotOptions}{A list containing the options for the plots.}
}
\value{
The list \code{tables} containing the tables for the evaluation report.
}
\description{
Generate all the table for the evaluation report
}
