% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{resizeFisherMatrix}
\alias{resizeFisherMatrix}
\title{Resize the fisher Matrix from a vector to a matrix.}
\usage{
resizeFisherMatrix(nb_dimensions, fisherMatrix)
}
\arguments{
\item{nb_dimensions}{: a numeric for the dimensions of the fisher matrix.}

\item{fisherMatrix}{: a vector that contain the low triangular Fisher matrix + its main diagonal.}
}
\value{
The Fisher matrix of size nb_dimensions*nb_dimensions.
}
\description{
Resize the fisher Matrix from a vector to a matrix.
}
