% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{run}
\alias{run}
\title{Run optimization}
\arguments{
\item{optimization}{An \code{Optimization} object.}

\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
The optimization design results.

The object \code{Evaluation} giving the design evaluation.
}
\description{
Run optimization

run: run the evaluation of a design.
}
