% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R, R/Evaluation.R, R/Optimization.R
\name{run}
\alias{run}
\alias{run,Evaluation-method}
\alias{run,Optimization-method}
\title{run}
\usage{
run(object)

\S4method{run}{Evaluation}(object)

\S4method{run}{Optimization}(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{PFIMProject}.}
}
\value{
A list giving the results of evaluation or optimization.
}
\description{
run
}
