% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{run}
\alias{run}
\title{run}
\usage{
run(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{PFIMProject}.}
}
\value{
A list giving the results of evaluation or optimization.
}
\description{
run
}
