% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{run}
\alias{run}
\title{run}
\usage{
run(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An object of class \code{PFIMProject} containing the
project settings, model definitions, and design parameters.}

\item{...}{Additional arguments.}
}
\description{
This function performs the evaluation or the optimization of a experimental design
based on the settings provided in a \code{PFIMProject} object.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{
# Execute the evaluation process
results = run(evaluationPopObject)
}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
