% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\docType{class}
\name{SamplingTimes-class}
\alias{SamplingTimes-class}
\alias{SamplingTimes}
\title{Class "SamplingTimes"}
\description{
The class "SamplingTimes" implements the sampling times.
}
\section{Objects from the class \code{SamplingTimes}}{

Objects form the class \code{SamplingTimes} can be created by calls of the form \code{SamplingTimes(...)} where
(...) are the parameters for the \code{SamplingTimes} objects.
}

\section{Slots for \code{SamplingTimes} objects}{

 \describe{
   \item{\code{outcome}:}{A string giving the outcome.}
   \item{\code{samplings}:}{A vector giving the sampling times.}
 }
}

