% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\name{SamplingTimes}
\alias{SamplingTimes}
\title{SamplingTimes}
\usage{
SamplingTimes(outcome = character(0), samplings = numeric(0))
}
\arguments{
\item{outcome}{A string giving the outcome.}

\item{samplings}{A vector of numeric giving the samplings.}
}
\description{
The class \code{SamplingTimes} is used to defined SamplingTimes.
}
