% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateTest.R
\name{saveCovariateTest}
\alias{saveCovariateTest}
\title{saveCovariateTest: sauvegarde les trois slots dans un fichier texte}
\usage{
saveCovariateTest(relevanceResults, file_name, folder = NULL)
}
\arguments{
\item{relevanceResults}{Un objet \code{CovariateTest}.}

\item{file_name}{Nom du fichier de sortie.}

\item{folder}{Dossier de destination (optionnel).}
}
\description{
saveCovariateTest: sauvegarde les trois slots dans un fichier texte
}
