% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setArm}
\alias{setArm}
\alias{setArm,Design-method}
\title{setArm}
\usage{
setArm(object, arm)

\S4method{setArm}{Design}(object, arm)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}

\item{arm}{A list of object \code{Arm} giving the arms of the design.}
}
\value{
An object \code{Design} with the list \code{Arm} updated.
}
\description{
Set the arms in a design.
}
