% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{setArms}
\alias{setArms}
\title{Set the arms of an object.}
\usage{
setArms(object, arms)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{arms}{A list of arms.}
}
\value{
The object with the updated arms.
}
\description{
Set the arms of an object.
}
