% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Design.R,
%   R/OptimizationAlgorithm.R
\name{setArms}
\alias{setArms}
\alias{setArms,Design-method}
\alias{setArms,OptimizationAlgorithm-method}
\title{Set the arms of an object.}
\usage{
setArms(object, arms)

\S4method{setArms}{Design}(object, arms)

\S4method{setArms}{OptimizationAlgorithm}(object, arms)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{arms}{A list of arms.}
}
\value{
The object with the updated arms.
}
\description{
Set the arms of an object.
}
