% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{setcError}
\alias{setcError}
\title{Set the parameter c.}
\usage{
setcError(object, cError)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelError}.}

\item{cError}{A numeric giving the parameter c.}
}
\value{
The model error with the parameter c.
}
\description{
Set the parameter c.
}
