% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setDescription}
\alias{setDescription}
\alias{setDescription,Model-method}
\title{Set the description of a model.}
\usage{
setDescription(object, description)

\S4method{setDescription}{Model}(object, description)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{description}{A list giving the description of a model.}
}
\value{
The model with the updated description.
}
\description{
Set the description of a model.
}
