% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{setEquation}
\alias{setEquation}
\alias{setEquation,ModelError-method}
\title{Set the equation of a model error.}
\usage{
setEquation(object, equation)

\S4method{setEquation}{ModelError}(object, equation)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelError}.}

\item{equation}{An expression giving the equation of a model error.}
}
\value{
The model error with the updated equation.
}
\description{
Set the equation of a model error.
}
