% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setEquations}
\alias{setEquations}
\title{Set the equations of a model.}
\usage{
setEquations(object, equations)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{equations}{A list giving the equations of the model.}
}
\value{
The model with the updated equations.
}
\description{
Set the equations of a model.
}
