% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{setEvaluationFim}
\alias{setEvaluationFim}
\title{setEvaluationFim: set the Fim results.}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{evaluation}{An object \code{Evaluation} giving the evaluation of the model.}
}
\value{
The object \code{Fim} with its fisherMatrix, fixedEffects, shrinkage, condNumberFixedEffects, SEAndRSE.

The object \code{IndividualFim} with its fisherMatrix, fixedEffects, shrinkage, condNumberFixedEffects, SEAndRSE.

The object \code{PopulationFim} with its fisherMatrix, fixedEffects, shrinkage, condNumberFixedEffects, SEAndRSE.
}
\description{
setEvaluationFim: set the Fim results.

setEvaluationFim: set the Fim results.

setEvaluationFim: set the Fim results.
}
