% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{setEvaluationFim}
\alias{setEvaluationFim}
\title{setEvaluationFim: populate FIM result slots after optimisation/evaluation}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{evaluation}{An object \code{Evaluation} giving the evaluation of the model.}
}
\value{
The updated \code{BayesianFim}.

The updated \code{IndividualFim}.

The object \code{PopulationFim} with fisherMatrix, fixedEffects,
  condNumberFixedEffects, and SEAndRSE populated.
}
\description{
setEvaluationFim: populate FIM result slots after optimisation/evaluation

setEvaluationFim: populate FIM result slots

setEvaluationFim: set the Fim results.
}
