% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setFim}
\alias{setFim}
\title{Set the fim of the design.}
\usage{
setFim(object, fim)
}
\arguments{
\item{object}{An object \code{Design} from the class \linkS4class{Design}.}

\item{fim}{An object \code{fim} from the class \linkS4class{Fim}.}
}
\value{
An object \code{Design} with the \code{fim} updated.
}
\description{
Set the fim of the design.
}
