% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{setFixedEffects}
\alias{setFixedEffects}
\title{Set the fixed effects.}
\usage{
setFixedEffects(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}
}
\value{
Update the matrix of the fixed effects.
}
\description{
Set the fixed effects.
}
