% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{setInitialConditions}
\alias{setInitialConditions}
\title{Set the initial conditions of an Arm for an ODE model.}
\usage{
setInitialConditions(object, values)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{values}{A list of numeric giving the values of the initial conditions.}
}
\value{
The object \code{Arm} with the new initial conditions for an ODE model.
}
\description{
Set the initial conditions of an Arm for an ODE model.
}
