% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Arm.R, R/Model.R
\name{setInitialConditions}
\alias{setInitialConditions}
\alias{setInitialConditions,Arm-method}
\alias{setInitialConditions,Model-method}
\title{setInitialConditions}
\usage{
setInitialConditions(object, initialConditions)

\S4method{setInitialConditions}{Arm}(object, initialConditions)

\S4method{setInitialConditions}{Model}(object, initialConditions)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{initialConditions}{A list giving the initial conditions.}
}
\value{
The model with the updated initial conditions.
}
\description{
Set the initial conditions of a ode model.
}
