% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{setModel}
\alias{setModel}
\title{Set the model.}
\usage{
setModel(object, model)
}
\arguments{
\item{object}{An object from the class \linkS4class{PFIMProject}.}

\item{model}{An object from the class \linkS4class{Model}.}
}
\value{
The object with the updated model.
}
\description{
Set the model.
}
