% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setModelError}
\alias{setModelError}
\alias{setModelError,Model-method}
\title{Set the model error.}
\usage{
setModelError(object, modelError)

\S4method{setModelError}{Model}(object, modelError)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{modelError}{An object from the class \linkS4class{ModelError}.}
}
\value{
The model with the updated model error.
}
\description{
Set the model error.
}
