% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{setMu}
\alias{setMu}
\title{Set the mu vector.}
\usage{
setMu(object, mu)
}
\arguments{
\item{object}{A \code{Fim} object.}

\item{mu}{A vector mu of the new values of mu.}
}
\value{
The \code{Fim} object with the mu vector with the new values.
}
\description{
Set the mu vector.
}
