% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Distribution.R,
%   R/ModelParameter.R
\name{setMu}
\alias{setMu}
\alias{setMu,Distribution-method}
\alias{setMu,ModelParameter-method}
\title{Set the value of the fixed effect mu of an object.}
\usage{
setMu(object, value)

\S4method{setMu}{Distribution}(object, value)

\S4method{setMu}{ModelParameter}(object, value)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{value}{The value of the fixed effect mu.}
}
\value{
The object with the updated fixed effect mu.
}
\description{
Set the value of the fixed effect mu of an object.
}
