% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Arm.R, R/Design.R,
%   R/Model.R
\name{setName}
\alias{setName}
\alias{setName,Arm-method}
\alias{setName,Design-method}
\alias{setName,Model-method}
\title{Set the name of an object.}
\usage{
setName(object, name)

\S4method{setName}{Arm}(object, name)

\S4method{setName}{Design}(object, name)

\S4method{setName}{Model}(object, name)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{name}{A string giving the name of the object.}
}
\value{
The object with the updated name.
}
\description{
Set the name of an object.
}
