% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setOdeSolverParameters}
\alias{setOdeSolverParameters}
\title{Set the parameters of the ode solver.}
\usage{
setOdeSolverParameters(object, odeSolverParameters)
}
\arguments{
\item{object}{An object from the class \linkS4class{Model}.}

\item{odeSolverParameters}{A list giving the parameters of the ode solver.}
}
\value{
The model with the updated parameters of the ode solver.
}
\description{
Set the parameters of the ode solver.
}
