% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{setOmega}
\alias{setOmega}
\title{Set the matrix omega of an object.}
\usage{
setOmega(object, value)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{value}{The matrix omega.}
}
\value{
The object with the updated matrix omega.
}
\description{
Set the matrix omega of an object.
}
