% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{setOptimalArms}
\alias{setOptimalArms}
\title{setOptimalArms: set the optimal arms of an optimization algorithm.}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{optimizationAlgorithm}{An object \code{FedorovWynnAlgorithm} giving the optimization algorithm.}
}
\value{
The optimal arms.

The optimal arms.

The list optimalArms.

The list optimalArms.

The list optimalArms.

The list optimalArms.
}
\description{
setOptimalArms: set the optimal arms of an optimization algorithm.

setOptimalArms: set the optimal arms of an optimization algorithm.

setOptimalArms: set the optimal arms of an optimization algorithm.

setOptimalArms: set the optimal arms of an optimization algorithm.

setOptimalArms: set the optimal arms of an optimization algorithm.

setOptimalArms: set the optimal arms of an optimization algorithm.
}
