% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{setOptimalArms}
\alias{setOptimalArms}
\title{setOptimalArms: MultiplicativeAlgorithm}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{optimizationAlgorithm}{An object \code{FedorovWynnAlgorithm}.}
}
\value{
List of optimal arms.

List of optimal arms.

List of optimal arms.

List of optimal arms.

The list optimalArms.

The list optimalArms.
}
\description{
setOptimalArms: MultiplicativeAlgorithm

setOptimalArms: FedorovWynnAlgorithm

setOptimalArms: MultiplicativeAlgorithm

setOptimalArms: FedorovWynnAlgorithm

setOptimalArms: set the optimal arms of a MultiplicativeAlgorithm.

setOptimalArms: set the optimal arms of a FedorovWynnAlgorithm.
}
