% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{setOptimalArms}
\alias{setOptimalArms}
\title{Set the optimal arms for Multiplicative Algorithm}
\arguments{
\item{fim}{An object of class \code{\link{PopulationFim}} containing the
Fisher Information Matrix evaluated at the optimal design points.}

\item{optimizationAlgorithm}{An object of class \code{\link{MultiplicativeAlgorithm}\link{FedorovWynnAlgorithm}}
representing the solver that has completed its execution.
sampling schedule and dosing protocols that maximize the optimization criterion.}
}
\value{
A list of the optimal arms.

A list of the optimal arms.

The list optimalArms.
}
\description{
The \code{setOptimalArms} method identifies and extracts the best performing
experimental arms from an optimization routine. It converts the output of
the \code{\link{MultiplicativeAlgorithm}} into a structured list of
optimized experimental designs.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.

Copyright (c) 2026-present Romain Leroux. All rights reserved.

Copyright (c) 2026-present Romain Leroux. All rights reserved.

Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
