% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/Administration.R,
%   R/SamplingTimes.R
\name{setOutcome}
\alias{setOutcome}
\alias{setOutcome,Administration-method}
\alias{setOutcome,SamplingTimes-method}
\title{setOutcome}
\usage{
setOutcome(object, outcome)

\S4method{setOutcome}{Administration}(object, outcome)

\S4method{setOutcome}{SamplingTimes}(object, outcome)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{outcome}{A string defined the outcome.}
}
\value{
A string giving the updated outcome of the object.
}
\description{
Set the outcome of an object.
}
