% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{setParameters}
\alias{setParameters}
\title{Set the parameters of an object.}
\usage{
setParameters(object, parameters)
}
\arguments{
\item{object}{An object defined form a class of PFIM.}

\item{parameters}{A list of parameters.}
}
\value{
The object with the updated list of parameters.
}
\description{
Set the parameters of an object.
}
