% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setSamplingConstraintForOptimization}
\alias{setSamplingConstraintForOptimization}
\title{Set the sampling times constraint for optimization with PSO, PGBO and Simplex}
\usage{
setSamplingConstraintForOptimization(object)
}
\arguments{
\item{object}{An object from the class \linkS4class{Design}.}
}
\value{
The arms with the sampling times constraints.
}
\description{
Set the sampling times constraint for optimization with PSO, PGBO and Simplex
}
