% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{setSamplingConstraintForOptimization}
\alias{setSamplingConstraintForOptimization}
\title{setSamplingConstraintForOptimization: set the sampling time constraints for an arm for the design optimization.}
\arguments{
\item{design}{An object \code{Design} giving the design.}
}
\value{
The arm with the sampling time constraint for the design optimization.
}
\description{
setSamplingConstraintForOptimization: set the sampling time constraints for an arm for the design optimization.
}
