% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{setSamplingTimes}
\alias{setSamplingTimes}
\title{Set the vectors of sampling times for an arm.}
\usage{
setSamplingTimes(object, samplingTimes)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{samplingTimes}{The list containing the new sampling times.}
}
\value{
An object \code{Arm} from the class \linkS4class{Arm} with the new sampling times \code{samplingTimes}.
}
\description{
Set the vectors of sampling times for an arm.
}
