% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{setSamplingTimesConstraints}
\alias{setSamplingTimesConstraints}
\alias{setSamplingTimesConstraints,Arm-method}
\title{setSamplingTimesConstraints}
\usage{
setSamplingTimesConstraints(object, samplingTimesConstraints)

\S4method{setSamplingTimesConstraints}{Arm}(object, samplingTimesConstraints)
}
\arguments{
\item{object}{An object \code{Arm} from the class \linkS4class{Arm}.}

\item{samplingTimesConstraints}{An object \code{SamplingTimeConstraints} from the class \linkS4class{SamplingTimeConstraints}.}
}
\value{
The arm with the new sampling time constraints.
}
\description{
Set the sampling times constraints.
}
