% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{setShrinkage}
\alias{setShrinkage}
\title{Set the shrinkage.}
\usage{
setShrinkage(object, value)
}
\arguments{
\item{object}{An object from the class \linkS4class{Fim}.}

\item{value}{A vector giving the shrinkage of the Bayesian fim.}
}
\value{
The object with the updated shrinkage.
}
\description{
Set the shrinkage.
}
