% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{setSigmaInter}
\alias{setSigmaInter}
\title{Set the parameter sigma inter.}
\usage{
setSigmaInter(object, sigmaInter)
}
\arguments{
\item{object}{An object from the class \linkS4class{ModelError}.}

\item{sigmaInter}{A numeric giving the parameter sigma inter.}
}
\value{
The model error with the updated sigma inter.
}
\description{
Set the parameter sigma inter.
}
