% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/Evaluation.R
\name{show}
\alias{show}
\title{Show optimization results}
\arguments{
\item{optimization}{An \code{Optimization} object.}

\item{pfimproject}{A object \code{PFIMProject} giving the Evaluation.}
}
\value{
Prints results to console.

The show of the evaluation of the design.
}
\description{
Show optimization results

show: show the evaluation in the R console.
}
