% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/IndividualFim.R, R/ModelError.R,
%   R/Combined1.R, R/Combined1c.R, R/Combined2c.R, R/Combined2.R,
%   R/Constant.R, R/Optimization.R, R/Design.R, R/DesignConstraint.R,
%   R/MultiplicativeAlgorithm.R, R/StatisticalModel.R, R/PFIMProject.R
\name{show,Fim-method}
\alias{show,Fim-method}
\alias{show,IndividualFim-method}
\alias{show,ModelError-method}
\alias{show,Combined1-method}
\alias{show,Combined1c-method}
\alias{show,Combined2c-method}
\alias{show,Combined2-method}
\alias{show,Constant-method}
\alias{show,Optimization-method}
\alias{show,Design-method}
\alias{show,DesignConstraint-method}
\alias{show,MultiplicativeAlgorithm-method}
\alias{show,StatisticalModel-method}
\alias{show,PFIMProject-method}
\title{Show the Fisher Information Matrix for a \code{Fim} object and its information: Determinant, D-criterion, SE, Eigenvalues, Correlation.}
\usage{
\S4method{show}{Fim}(object)

\S4method{show}{IndividualFim}(object)

\S4method{show}{ModelError}(object)

\S4method{show}{Combined1}(object)

\S4method{show}{Combined1c}(object)

\S4method{show}{Combined2c}(object)

\S4method{show}{Combined2}(object)

\S4method{show}{Constant}(object)

\S4method{show}{Optimization}(object)

\S4method{show}{Design}(object)

\S4method{show}{DesignConstraint}(object)

\S4method{show}{MultiplicativeAlgorithm}(object)

\S4method{show}{StatisticalModel}(object)

\S4method{show}{PFIMProject}(object)
}
\arguments{
\item{object}{\code{PFIMProject} object.}
}
\value{
Print the Fisher Information Matrix and its informations: Determinant, D-criterion, SE, Eigenvalues, Correlation.

Show the Individual Fim.

Show the values of \code{sigma_inter}, \code{sigma_slope}, and  \code{c_error}.

Display the model errors

Display the model errors.

Display the model errors.

Display the model errors.

The model errors.

The content of an \code{Optimization} object.

Return the FIM of the design and the data summary of the arm in the design.

Show the content of a design.

Print the end of the process for the multiplicative algorithm.

Display the responses name of the model equations,
the ordinary derivatives of the model equations (for an ODE model), the parameters of the model.

Summary of the PFIMProject \code{PFIMProject} object given by the:
PopulationFim, IndividualFim, Fim, Determinant, Dcriterion,
Eigenvalues, SEfixedEffects,
SEstandardDeviationOfRandomEffect, SEerrorModel, Designs.
}
\description{
Show the Fisher Information Matrix for a \code{Fim} object and its information: Determinant, D-criterion, SE, Eigenvalues, Correlation.

Show the Individual Fim.

Show the values of \code{sigma_inter}, \code{sigma_slope}, and  \code{c_error}.

Show the model errors

Show the model errors.

Show the model errors.

Show the model errors.

Show the model errors.

Show for an \code{Optimization} object.

Show a design.

Show the content of a design.

Show the end of the process for the multiplicative algorithm.

Show the content of a \code{StatisticalModel} object.

show the content of the \code{PFIMProject} object.
}
