% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R, R/Evaluation.R,
%   R/FedorovWynnAlgorithm.R, R/MultiplicativeAlgorithm.R, R/Optimization.R,
%   R/PGBOAlgorithm.R, R/PSOAlgorithm.R, R/SimplexAlgorithm.R
\name{show,Design-method}
\alias{show,Design-method}
\alias{show,Evaluation-method}
\alias{show,FedorovWynnAlgorithm-method}
\alias{show,MultiplicativeAlgorithm-method}
\alias{show,Optimization-method}
\alias{show,PGBOAlgorithm-method}
\alias{show,PSOAlgorithm-method}
\alias{show,SimplexAlgorithm-method}
\title{show}
\usage{
\S4method{show}{Design}(object)

\S4method{show}{Evaluation}(object)

\S4method{show}{FedorovWynnAlgorithm}(object)

\S4method{show}{MultiplicativeAlgorithm}(object)

\S4method{show}{Optimization}(object)

\S4method{show}{PGBOAlgorithm}(object)

\S4method{show}{PSOAlgorithm}(object)

\S4method{show}{SimplexAlgorithm}(object)
}
\arguments{
\item{object}{object}
}
\description{
show

show

show

show

show

show

show

show
}
