% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{show}
\alias{show}
\title{show}
\usage{
show(pfimproject, ...)
}
\arguments{
\item{pfimproject}{The object to be displayed.}

\item{...}{Additional arguments.}
}
\value{
Invisibly prints the FIM summary to the console.
}
\description{
Displays a formatted summary of a \code{PFIMProject} object including the FIM,
standard errors, and design metrics.
}
\note{
Copyright (c) 2026-present Romain Leroux. All rights reserved.
}
\examples{
\dontrun{
# Show the results of the evaluationPopulationFIMResults
show(evaluationPopulationFIMResults)
}
}
\author{
Romain Leroux \email{romainlerouxPFIM@gmail.com}
}
