% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{showFIM}
\alias{showFIM}
\title{showFIM: show the Fim in the R console.}
\arguments{
\item{fim}{An object \code{IndividualFim} giving the Fim.}
}
\value{
The fisherMatrix, fixedEffects, Determinant, condition numbers and D-criterion, Shrinkage and Parameters estimation

The fisherMatrix, fixedEffects, Determinant, condition numbers and D-criterion, Shrinkage and Parameters estimation

The fisherMatrix, fixedEffects, Determinant, condition numbers and D-criterion, Shrinkage and Parameters estimation
}
\description{
showFIM: show the Fim in the R console.

showFIM: show the Fim in the R console.

showFIM: show the Fim in the R console.
}
