% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{showFIM}
\alias{showFIM}
\title{showFIM: print the Bayesian FIM to the console}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}
}
\value{
Printed console output (invisibly returns fim).
}
\description{
showFIM: print the Bayesian FIM to the console

showFIM: print the individual FIM to the console

showFIM: show the Fim in the R console.
}
