% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexAlgorithm.R
\docType{class}
\name{SimplexAlgorithm-class}
\alias{SimplexAlgorithm-class}
\alias{SimplexAlgorithm}
\title{Class "SimplexAlgorithm"}
\description{
Class "SimplexAlgorithm" implements the Multiplicative algorithm.
}
\section{Objects from the class \code{SimplexAlgorithm}}{

Objects form the class \code{SimplexAlgorithm} can be created by calls of the form \code{SimplexAlgorithm(...)} where
(...) are the parameters for the \code{SimplexAlgorithm} objects.
}

\section{Slots for \code{SamplingTimes} objects}{

 \describe{
   \item{\code{pctInitialSimplexBuilding}:}{A numeric giving the percentage of the initial simplex.}
   \item{\code{maxIteration}:}{A numeric giving the number of maximum iteration.}
   \item{\code{tolerance}:}{A numeric giving the tolerance threshold.}
   \item{\code{showProcess}:}{A boolean to show or not the process.}
   \item{\code{optimalDesign}:}{A \code{Design} object giving the optimal design.}
   \item{\code{iterationAndCriteria}:}{A list giving the optimal criteria at each iteration.}
 }
}

