% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{tablesForReport}
\alias{tablesForReport}
\title{tablesForReport: generate the table for the report.}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{evaluation}{An object \code{Evaluation} giving the evaluation of the model.}
}
\value{
fixedEffectsTable, FIMCriteriaTable, SEAndRSETable.

fixedEffectsTable, FIMCriteriaTable, SEAndRSETable.

fixedEffectsTable, FIMCriteriaTable, SEAndRSETable.
}
\description{
tablesForReport: generate the table for the report.

tablesForReport: generate the table for the report.

tablesForReport: generate the table for the report.
}
