% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{tablesForReport}
\alias{tablesForReport}
\title{tablesForReport: kableExtra tables for the Bayesian FIM report}
\arguments{
\item{fim}{An object \code{PopulationFim} giving the Fim.}

\item{evaluation}{An object \code{Evaluation} giving the evaluation of the model.}
}
\value{
List: fixedEffectsTable, FIMCriteriaTable, SEAndRSETable.

List: fixedEffectsTable, varianceEffectsTable, FIMCriteriaTable, SEAndRSETable.

List: fixedEffectsTable, varianceEffectsTable, FIMCriteriaTable, SEAndRSETable.
}
\description{
tablesForReport: kableExtra tables for the Bayesian FIM report

tablesForReport: kableExtra tables for the individual FIM report

tablesForReport: generate the tables for the HTML report.
}
