% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{updateSamplingTimes}
\alias{updateSamplingTimes}
\title{updateSamplingTimes: update sampling times for plotting used for plot}
\arguments{
\item{arm}{A object of class \code{Arm} giving the arm.}

\item{samplingData}{The list giving as output in the method getSamplingData.}
}
\value{
The updated sampling times.
}
\description{
updateSamplingTimes: update sampling times for plotting used for plot
}
